Difference variables

This is the list of difference variables of the form Delta_var_stage, where var is the variable name (from ReboundTools::key_analysis_vars) and stage is the rebound stage being considered (from ReboundTools::rebound_stages).

Usage

Delta_vars

Format

A string list with 84 entries.

Delta_eta_star

Calculated as eta_star - eta_orig.

Delta_p_s_star

Calculated as p_s_star - p_s_orig.

.

Etc.

.

Etc.

.

Etc.

Delta_M_dot_tilde

Calculated as M_dot_tilde - M_dot_bar.

Delta_N_dot_tilde

Calculated as N_dot_tilde - N_dot_bar.

Examples

Delta_vars
#> $Delta_M_dot_star
#> [1] "Delta_M_dot_star"
#> 
#> $Delta_p_E_star
#> [1] "Delta_p_E_star"
#> 
#> $Delta_t_life_star
#> [1] "Delta_t_life_star"
#> 
#> $Delta_tau_alpha_star
#> [1] "Delta_tau_alpha_star"
#> 
#> $Delta_tau_omega_star
#> [1] "Delta_tau_omega_star"
#> 
#> $Delta_eta_engr_units_star
#> [1] "Delta_eta_engr_units_star"
#> 
#> $Delta_eta_star
#> [1] "Delta_eta_star"
#> 
#> $Delta_p_s_star
#> [1] "Delta_p_s_star"
#> 
#> $Delta_E_dot_emb_star
#> [1] "Delta_E_dot_emb_star"
#> 
#> $Delta_C_dot_cap_star
#> [1] "Delta_C_dot_cap_star"
#> 
#> $Delta_tau_alpha_C_dot_cap_star
#> [1] "Delta_tau_alpha_C_dot_cap_star"
#> 
#> $Delta_C_dot_om_star
#> [1] "Delta_C_dot_om_star"
#> 
#> $Delta_C_d_star
#> [1] "Delta_C_d_star"
#> 
#> $Delta_C_dot_d_star
#> [1] "Delta_C_dot_d_star"
#> 
#> $Delta_tau_omega_C_dot_d_star
#> [1] "Delta_tau_omega_C_dot_d_star"
#> 
#> $Delta_C_dot_omd_star
#> [1] "Delta_C_dot_omd_star"
#> 
#> $Delta_E_dot_s_star
#> [1] "Delta_E_dot_s_star"
#> 
#> $Delta_C_dot_s_star
#> [1] "Delta_C_dot_s_star"
#> 
#> $Delta_N_dot_star
#> [1] "Delta_N_dot_star"
#> 
#> $Delta_q_dot_s_star
#> [1] "Delta_q_dot_s_star"
#> 
#> $Delta_C_dot_g_star
#> [1] "Delta_C_dot_g_star"
#> 
#> $Delta_M_dot_hat
#> [1] "Delta_M_dot_hat"
#> 
#> $Delta_p_E_hat
#> [1] "Delta_p_E_hat"
#> 
#> $Delta_t_life_hat
#> [1] "Delta_t_life_hat"
#> 
#> $Delta_tau_alpha_hat
#> [1] "Delta_tau_alpha_hat"
#> 
#> $Delta_tau_omega_hat
#> [1] "Delta_tau_omega_hat"
#> 
#> $Delta_eta_engr_units_hat
#> [1] "Delta_eta_engr_units_hat"
#> 
#> $Delta_eta_hat
#> [1] "Delta_eta_hat"
#> 
#> $Delta_p_s_hat
#> [1] "Delta_p_s_hat"
#> 
#> $Delta_E_dot_emb_hat
#> [1] "Delta_E_dot_emb_hat"
#> 
#> $Delta_C_dot_cap_hat
#> [1] "Delta_C_dot_cap_hat"
#> 
#> $Delta_tau_alpha_C_dot_cap_hat
#> [1] "Delta_tau_alpha_C_dot_cap_hat"
#> 
#> $Delta_C_dot_om_hat
#> [1] "Delta_C_dot_om_hat"
#> 
#> $Delta_C_d_hat
#> [1] "Delta_C_d_hat"
#> 
#> $Delta_C_dot_d_hat
#> [1] "Delta_C_dot_d_hat"
#> 
#> $Delta_tau_omega_C_dot_d_hat
#> [1] "Delta_tau_omega_C_dot_d_hat"
#> 
#> $Delta_C_dot_omd_hat
#> [1] "Delta_C_dot_omd_hat"
#> 
#> $Delta_E_dot_s_hat
#> [1] "Delta_E_dot_s_hat"
#> 
#> $Delta_C_dot_s_hat
#> [1] "Delta_C_dot_s_hat"
#> 
#> $Delta_N_dot_hat
#> [1] "Delta_N_dot_hat"
#> 
#> $Delta_q_dot_s_hat
#> [1] "Delta_q_dot_s_hat"
#> 
#> $Delta_C_dot_g_hat
#> [1] "Delta_C_dot_g_hat"
#> 
#> $Delta_M_dot_bar
#> [1] "Delta_M_dot_bar"
#> 
#> $Delta_p_E_bar
#> [1] "Delta_p_E_bar"
#> 
#> $Delta_t_life_bar
#> [1] "Delta_t_life_bar"
#> 
#> $Delta_tau_alpha_bar
#> [1] "Delta_tau_alpha_bar"
#> 
#> $Delta_tau_omega_bar
#> [1] "Delta_tau_omega_bar"
#> 
#> $Delta_eta_engr_units_bar
#> [1] "Delta_eta_engr_units_bar"
#> 
#> $Delta_eta_bar
#> [1] "Delta_eta_bar"
#> 
#> $Delta_p_s_bar
#> [1] "Delta_p_s_bar"
#> 
#> $Delta_E_dot_emb_bar
#> [1] "Delta_E_dot_emb_bar"
#> 
#> $Delta_C_dot_cap_bar
#> [1] "Delta_C_dot_cap_bar"
#> 
#> $Delta_tau_alpha_C_dot_cap_bar
#> [1] "Delta_tau_alpha_C_dot_cap_bar"
#> 
#> $Delta_C_dot_om_bar
#> [1] "Delta_C_dot_om_bar"
#> 
#> $Delta_C_d_bar
#> [1] "Delta_C_d_bar"
#> 
#> $Delta_C_dot_d_bar
#> [1] "Delta_C_dot_d_bar"
#> 
#> $Delta_tau_omega_C_dot_d_bar
#> [1] "Delta_tau_omega_C_dot_d_bar"
#> 
#> $Delta_C_dot_omd_bar
#> [1] "Delta_C_dot_omd_bar"
#> 
#> $Delta_E_dot_s_bar
#> [1] "Delta_E_dot_s_bar"
#> 
#> $Delta_C_dot_s_bar
#> [1] "Delta_C_dot_s_bar"
#> 
#> $Delta_N_dot_bar
#> [1] "Delta_N_dot_bar"
#> 
#> $Delta_q_dot_s_bar
#> [1] "Delta_q_dot_s_bar"
#> 
#> $Delta_C_dot_g_bar
#> [1] "Delta_C_dot_g_bar"
#> 
#> $Delta_M_dot_tilde
#> [1] "Delta_M_dot_tilde"
#> 
#> $Delta_p_E_tilde
#> [1] "Delta_p_E_tilde"
#> 
#> $Delta_t_life_tilde
#> [1] "Delta_t_life_tilde"
#> 
#> $Delta_tau_alpha_tilde
#> [1] "Delta_tau_alpha_tilde"
#> 
#> $Delta_tau_omega_tilde
#> [1] "Delta_tau_omega_tilde"
#> 
#> $Delta_eta_engr_units_tilde
#> [1] "Delta_eta_engr_units_tilde"
#> 
#> $Delta_eta_tilde
#> [1] "Delta_eta_tilde"
#> 
#> $Delta_p_s_tilde
#> [1] "Delta_p_s_tilde"
#> 
#> $Delta_E_dot_emb_tilde
#> [1] "Delta_E_dot_emb_tilde"
#> 
#> $Delta_C_dot_cap_tilde
#> [1] "Delta_C_dot_cap_tilde"
#> 
#> $Delta_tau_alpha_C_dot_cap_tilde
#> [1] "Delta_tau_alpha_C_dot_cap_tilde"
#> 
#> $Delta_C_dot_om_tilde
#> [1] "Delta_C_dot_om_tilde"
#> 
#> $Delta_C_d_tilde
#> [1] "Delta_C_d_tilde"
#> 
#> $Delta_C_dot_d_tilde
#> [1] "Delta_C_dot_d_tilde"
#> 
#> $Delta_tau_omega_C_dot_d_tilde
#> [1] "Delta_tau_omega_C_dot_d_tilde"
#> 
#> $Delta_C_dot_omd_tilde
#> [1] "Delta_C_dot_omd_tilde"
#> 
#> $Delta_E_dot_s_tilde
#> [1] "Delta_E_dot_s_tilde"
#> 
#> $Delta_C_dot_s_tilde
#> [1] "Delta_C_dot_s_tilde"
#> 
#> $Delta_N_dot_tilde
#> [1] "Delta_N_dot_tilde"
#> 
#> $Delta_q_dot_s_tilde
#> [1] "Delta_q_dot_s_tilde"
#> 
#> $Delta_C_dot_g_tilde
#> [1] "Delta_C_dot_g_tilde"
#>