EEU bar data

This is the list of the derived variables at the bar stage (after income effect) of a rebound analysis.

Usage

bar_vars

Format

A string list with 26 entries.

R_alpha_bar

The post-income effect discount multiplier for beginning-of-life expenses, such as capital cost. Same as R_alpha_hat.

R_omega_bar

The post-income effect discount multiplier for end-of-life expenses, such as disposal cost. Same as R_omega_hat.

eta_bar

Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], exactly eta_hat.

p_s_bar

The energy service price after the income effect [$/service], exactly p_s_hat.

C_dot_cap_bar

The capital expenditure rate of the device without discounting [$/yr], calculated by C_cap/t_life.

R_alpha_C_dot_cap_bar

The product of R_alpha and the undiscounted capital cost rate, itself annualized and discounted in [$/yr].

C_dot_om_bar

The operations and maintenance expenditure rate of the device [$/yr].

C_d_bar

The disposal cost for the device [$/yr].

C_dot_d_bar

The disposal cost rate of the device without discounting [$/yr].

R_omega_C_dot_d_bar

The product of R_omega and the undiscounted disposal cost rate, itself annualized and discounted in [$/yr].

C_dot_omd_bar

The operations, maintenance, and disposal expenditure rate of the device [$/yr].

E_dot_emb_bar

The embodied energy rate after the income effect [MJ/yr], exactly E_dot_emb_hat.

M_dot_bar

Real income after the income effect [MJ/yr], exactly M_dot_hat.

q_dot_s_bar

The rate of energy service consumption after the income effect\ [service/yr], calculated by (1 + N_dot_hat/M_dot_hat_prime)^(e_qs_M).

C_dot_o_bar

The rate of other goods expenditures after the income effect [$/yr], calculated by (1 + N_dot_hat/M_dot_hat_prime)^(e_qo_M).

f_Cs_bar

The post-income effect fraction of the energy and other budget spent on the energy service [--], calculated by C_dot_s_star / (C_dot_s_star + C_dot_o_star).

e_qs_ps_C_bar

The post-income effect compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [--], calculated by e_qs_ps_UC_orig + f_Cs_orig*e_qs_M.

e_qo_ps_C_bar

The post-income effect compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [--], calculated by f_Cs_orig*(f_Cs_orig + e_qs_ps_UC_orig) / (f_Cs_orig - 1).

e_qs_ps_UC_bar

The post-income effect uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [--].

e_qo_ps_UC_bar

The post-income effect uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of other goods ("qo") consumption (cross-price elasticity) [--].

p_E_bar

The price of energy [$/MJ],

N_dot_bar

The freed cash rate after the income effect [$/yr], exactly 0.

[service/yr]: R:service/yr%5C

Examples

bar_vars
#> $t_life_bar
#> [1] "t_life_bar"
#> 
#> $R_alpha_bar
#> [1] "R_alpha_bar"
#> 
#> $R_omega_bar
#> [1] "R_omega_bar"
#> 
#> $eta_engr_units_bar
#> [1] "eta_engr_units_bar"
#> 
#> $eta_bar
#> [1] "eta_bar"
#> 
#> $p_s_bar
#> [1] "p_s_bar"
#> 
#> $C_dot_cap_bar
#> [1] "C_dot_cap_bar"
#> 
#> $R_alpha_C_dot_cap_bar
#> [1] "R_alpha_C_dot_cap_bar"
#> 
#> $E_dot_emb_bar
#> [1] "E_dot_emb_bar"
#> 
#> $M_dot_bar
#> [1] "M_dot_bar"
#> 
#> $q_dot_s_bar
#> [1] "q_dot_s_bar"
#> 
#> $p_E_bar
#> [1] "p_E_bar"
#> 
#> $E_dot_s_bar
#> [1] "E_dot_s_bar"
#> 
#> $C_dot_s_bar
#> [1] "C_dot_s_bar"
#> 
#> $C_dot_om_bar
#> [1] "C_dot_om_bar"
#> 
#> $C_d_bar
#> [1] "C_d_bar"
#> 
#> $C_dot_d_bar
#> [1] "C_dot_d_bar"
#> 
#> $R_omega_C_dot_d_bar
#> [1] "R_omega_C_dot_d_bar"
#> 
#> $C_dot_omd_bar
#> [1] "C_dot_omd_bar"
#> 
#> $C_dot_o_bar
#> [1] "C_dot_o_bar"
#> 
#> $f_Cs_bar
#> [1] "f_Cs_bar"
#> 
#> $e_qs_ps_C_bar
#> [1] "e_qs_ps_C_bar"
#> 
#> $e_qs_ps_UC_bar
#> [1] "e_qs_ps_UC_bar"
#> 
#> $e_qo_ps_C_bar
#> [1] "e_qo_ps_C_bar"
#> 
#> $e_qo_ps_UC_bar
#> [1] "e_qo_ps_UC_bar"
#> 
#> $N_dot_bar
#> [1] "N_dot_bar"
#>