Calculate energy rebound data at the bar stage

This function calculates energy rebound information for the bar stage (immediately after the income effect).

Usage

calc_bar(
  .hat_data = NULL,
  tol = 1e-06,
  t_life_hat = ReboundTools::hat_vars$t_life_hat,
  tau_alpha_hat = ReboundTools::hat_vars$tau_alpha_hat,
  tau_omega_hat = ReboundTools::hat_vars$tau_omega_hat,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qg_M = ReboundTools::eeu_base_params$e_qg_M,
  p_E = ReboundTools::orig_vars$p_E,
  eta_engr_units_hat = ReboundTools::hat_vars$eta_engr_units_hat,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime,
  C_d_hat = ReboundTools::hat_vars$C_d_hat,
  C_dot_d_hat = ReboundTools::hat_vars$C_dot_d_hat,
  C_dot_om_hat = ReboundTools::hat_vars$C_dot_om_hat,
  C_dot_omd_hat = ReboundTools::hat_vars$C_dot_omd_hat,
  C_dot_g_hat = ReboundTools::hat_vars$C_dot_g_hat,
  e_qs_ps_UC_hat = ReboundTools::hat_vars$e_qs_ps_UC_hat,
  e_qg_ps_UC_hat = ReboundTools::hat_vars$e_qg_ps_UC_hat,
  e_qs_ps_C_hat = ReboundTools::hat_vars$e_qs_ps_C_hat,
  e_qg_ps_C_hat = ReboundTools::hat_vars$e_qg_ps_C_hat,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  t_life_bar = ReboundTools::bar_vars$t_life_bar,
  tau_alpha_bar = ReboundTools::bar_vars$tau_alpha_bar,
  tau_omega_bar = ReboundTools::bar_vars$tau_omega_bar,
  eta_engr_units_bar = ReboundTools::bar_vars$eta_engr_units_bar,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_E_bar = ReboundTools::bar_vars$p_E_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  tau_alpha_C_dot_cap_bar = ReboundTools::bar_vars$tau_alpha_C_dot_cap_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_d_bar = ReboundTools::bar_vars$C_d_bar,
  C_dot_d_bar = ReboundTools::bar_vars$C_dot_d_bar,
  tau_omega_C_dot_d_bar = ReboundTools::bar_vars$tau_omega_C_dot_d_bar,
  C_dot_om_bar = ReboundTools::bar_vars$C_dot_om_bar,
  C_dot_omd_bar = ReboundTools::bar_vars$C_dot_omd_bar,
  C_dot_g_bar = ReboundTools::bar_vars$C_dot_g_bar,
  f_Cs_bar = ReboundTools::bar_vars$f_Cs_bar,
  e_qs_ps_UC_bar = ReboundTools::bar_vars$e_qs_ps_UC_bar,
  e_qg_ps_UC_bar = ReboundTools::bar_vars$e_qg_ps_UC_bar,
  e_qs_ps_C_bar = ReboundTools::bar_vars$e_qs_ps_C_bar,
  e_qg_ps_C_bar = ReboundTools::bar_vars$e_qg_ps_C_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar
)

Arguments

.hat_data

An optional data frame containing rebound calculations, original data, star data, and hat data, likely calculated by calc_hat().

tol

The tolerance with which the budget constraint should be satisfied. Default is 1e-6.

t_life_hat, C_d_hat, C_dot_d_hat, C_dot_om_hat, C_dot_omd_hat, tau_alpha_hat, tau_omega_hat, eta_engr_units_hat, eta_hat, p_s_hat, C_dot_cap_hat, E_dot_emb_hat, M_dot_hat, q_dot_s_hat, N_dot_hat, M_dot_hat_prime, C_dot_g_hat, e_qs_ps_UC_hat, e_qg_ps_UC_hat, e_qs_ps_C_hat, e_qg_ps_C_hat, E_dot_s_hat

See ReboundTools::hat_vars.

e_qs_M, e_qg_M, p_E

See ReboundTools::eeu_base_params.

t_life_bar, C_d_bar, C_dot_d_bar, tau_omega_C_dot_d_bar, C_dot_om_bar, C_dot_omd_bar, tau_alpha_bar, tau_omega_bar, eta_engr_units_bar, eta_bar, p_s_bar, C_dot_cap_bar, tau_alpha_C_dot_cap_bar, E_dot_emb_bar, M_dot_bar, q_dot_s_bar, E_dot_s_bar, C_dot_s_bar, C_dot_g_bar, f_Cs_bar, e_qs_ps_UC_bar, e_qg_ps_UC_bar, e_qs_ps_C_bar, e_qg_ps_C_bar, N_dot_bar, p_E_bar

See ReboundTools::bar_vars.

Value

A list or data frame of derived rebound values for the bar stage (after the income effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar()
#> # A tibble: 3 × 128
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 121 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qg_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   tau_alpha_orig <dbl>, tau_omega_orig <dbl>, p_E <dbl>, p_E_orig <dbl>, …