Calculate energy rebound data at the original stage

This function calculates energy rebound information for the original stage (pre-EEU).

Usage

calc_orig(
  .eeu_data = NULL,
  tol = 1e-06,
  r = ReboundTools::eeu_base_params$r,
  MJ_engr_unit = ReboundTools::eeu_base_params$MJ_engr_unit,
  p_E_engr_units = ReboundTools::eeu_base_params$p_E_engr_units,
  e_qs_ps_UC_orig = ReboundTools::eeu_base_params$e_qs_ps_UC_orig,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qg_M = ReboundTools::eeu_base_params$e_qg_M,
  eta_engr_units_orig = ReboundTools::orig_vars$eta_engr_units_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_cap_orig = ReboundTools::orig_vars$C_cap_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  C_dot_om_orig = ReboundTools::orig_vars$C_dot_om_orig,
  C_d_orig = ReboundTools::orig_vars$C_d_orig,
  E_emb_orig = ReboundTools::orig_vars$E_emb_orig,
  t_life_orig = ReboundTools::orig_vars$t_life_orig,
  tau_alpha_orig = ReboundTools::orig_vars$tau_alpha_orig,
  tau_omega_orig = ReboundTools::orig_vars$tau_omega_orig,
  p_E = ReboundTools::orig_vars$p_E,
  p_E_orig = ReboundTools::orig_vars$p_E_orig,
  eta_orig = ReboundTools::orig_vars$eta_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  tau_alpha_C_dot_cap_orig = ReboundTools::orig_vars$tau_alpha_C_dot_cap_orig,
  p_s_orig = ReboundTools::orig_vars$p_s_orig,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_d_orig = ReboundTools::orig_vars$C_dot_d_orig,
  tau_omega_C_dot_d_orig = ReboundTools::orig_vars$tau_omega_C_dot_d_orig,
  C_dot_omd_orig = ReboundTools::orig_vars$C_dot_omd_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  e_qg_ps_UC_orig = ReboundTools::orig_vars$e_qg_ps_UC_orig,
  e_qs_ps_C_orig = ReboundTools::orig_vars$e_qs_ps_C_orig,
  e_qg_ps_C_orig = ReboundTools::orig_vars$e_qg_ps_C_orig,
  sigma = ReboundTools::orig_vars$sigma,
  rho = ReboundTools::orig_vars$rho,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  N_dot_orig = ReboundTools::orig_vars$N_dot_orig
)

Arguments

.eeu_data

An optional data frame containing EEU base data. See ReboundTools::eeu_base_params.

tol

The tolerance for checking internal consistency of rebound calculations. Default is 1e-6.

r, MJ_engr_unit, p_E_engr_units, e_qs_ps_UC_orig, e_qs_M, e_qg_M

See ReboundTools::eeu_base_params.

tau_alpha_orig, tau_omega_orig, eta_engr_units_orig, q_dot_s_orig, M_dot_orig, C_cap_orig, tau_alpha_C_dot_cap_orig, C_d_orig, C_dot_omd_orig, C_dot_om_orig, C_dot_d_orig, tau_omega_C_dot_d_orig, E_emb_orig, t_life_orig, p_E, eta_orig, E_dot_s_orig, C_dot_cap_orig, p_s_orig, C_dot_s_orig, C_dot_g_orig, f_Cs_orig, e_qg_ps_UC_orig, e_qs_ps_C_orig, e_qg_ps_C_orig, sigma, rho, E_dot_emb_orig, N_dot_orig, p_E_orig

See ReboundTools::orig_vars.

Value

A list or data frame of derived rebound values.

Examples

load_eeu_data() %>% 
  calc_orig()
#> # A tibble: 3 × 50
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 43 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qg_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   tau_alpha_orig <dbl>, tau_omega_orig <dbl>, p_E <dbl>, p_E_orig <dbl>, …