Calculate rebound terms.

This function calculates rebound terms from a data frame that already contains Delta terms. Note that each rebound term is calculated twice as a way of validating the derived expression for rebound.

Usage

calc_rebound(
  .Deltas_data = NULL,
  tol = 1e-10,
  I_E = ReboundTools::eeu_base_params$I_E,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  k = ReboundTools::eeu_base_params$k,
  e_qs_ps_C = ReboundTools::orig_vars$e_qs_ps_C,
  e_qo_ps_C = ReboundTools::orig_vars$e_qo_ps_C,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime,
  R_alpha_orig = ReboundTools::orig_vars$R_alpha_orig,
  R_alpha_star = ReboundTools::star_vars$R_alpha_star,
  C_dot_cap_orig = ReboundTools::orig_vars$C_cap_orig,
  C_dot_cap_star = ReboundTools::star_vars$C_cap_star,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars$Delta_E_dot_emb_star,
  R_omega_orig = ReboundTools::orig_vars$R_omega_orig,
  R_omega_star = ReboundTools::star_vars$R_omega_star,
  C_dot_d_orig = ReboundTools::orig_vars$C_dot_d_orig,
  C_dot_d_star = ReboundTools::star_vars$C_dot_d_star,
  Delta_C_dot_om_star = ReboundTools::Delta_vars$Delta_C_dot_om_star,
  Delta_C_dot_omd_star = ReboundTools::Delta_vars$Delta_C_dot_omd_star,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars$Delta_E_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_E_dot_s_bar = ReboundTools::Delta_vars$Delta_E_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  Re_dempl = ReboundTools::rebound_terms$Re_dempl,
  Re_emb = ReboundTools::rebound_terms$Re_emb,
  Re_cap = ReboundTools::rebound_terms$Re_cap,
  Re_om = ReboundTools::rebound_terms$Re_om,
  Re_d = ReboundTools::rebound_terms$Re_d,
  Re_omd = ReboundTools::rebound_terms$Re_omd,
  Re_empl = ReboundTools::rebound_terms$Re_empl,
  Re_dsub = ReboundTools::rebound_terms$Re_dsub,
  Re_isub = ReboundTools::rebound_terms$Re_isub,
  Re_sub = ReboundTools::rebound_terms$Re_sub,
  Re_dinc = ReboundTools::rebound_terms$Re_dinc,
  Re_iinc = ReboundTools::rebound_terms$Re_iinc,
  Re_inc = ReboundTools::rebound_terms$Re_inc,
  Re_micro = ReboundTools::rebound_terms$Re_micro,
  Re_macro = ReboundTools::rebound_terms$Re_macro,
  Re_dir = ReboundTools::rebound_terms$Re_dir,
  Re_indir = ReboundTools::rebound_terms$Re_indir,
  Re_tot = ReboundTools::rebound_terms$Re_tot
)

Arguments

.Deltas_data

A data frame containing Delta values, likely created by ReboundTools::calc_Deltas()

tol

The tolerance for checking internal consistency of rebound calculations. Default is 1e-10.

I_E, e_qs_M, e_qo_M, k

See ReboundTools::eeu_base_params.

e_qs_ps_C, e_qo_ps_C, C_dot_o_orig, E_dot_s_orig

See ReboundTools::orig_vars.

S_dot_dev, eta_ratio

See ReboundTools::star_vars.

N_dot_star

See ReboundTools::star_vars.

M_dot_hat_prime

See ReboundTools::hat_vars.

R_alpha_orig, C_dot_cap_orig, R_omega_orig, C_dot_d_orig

See ReboundTools::orig_vars.

R_alpha_star, R_omega_star, C_dot_d_star, C_dot_cap_star, Delta_C_dot_om_star, Delta_C_dot_omd_star, Delta_E_dot_emb_star, Delta_E_dot_s_hat, Delta_C_dot_o_hat, Delta_E_dot_s_bar, Delta_C_dot_o_bar

See ReboundTools::Delta_vars.

Re_dempl, Re_emb, Re_cap, Re_om, Re_d, Re_omd, Re_empl, Re_dsub, Re_isub, Re_sub, Re_dinc, Re_iinc, Re_inc, Re_micro, Re_macro, Re_dir, Re_indir, Re_tot

See ReboundTools::rebound_terms.

Value

A data frame with rebound terms added as columns.

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde() %>% 
  calc_Deltas() %>% 
  calc_rebound()
#> # A tibble: 3 × 229
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 222 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>, R_alpha_orig <dbl>,
#> #   R_omega_orig <dbl>, p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, …