Calculate energy rebound data at the star stage

This function calculates energy rebound information for the star stage (immediately after the emplacement effect, with no behavior change).

Usage

calc_star(
  .orig_data = NULL,
  r = ReboundTools::eeu_base_params$r,
  MJ_engr_unit = ReboundTools::eeu_base_params$MJ_engr_unit,
  R_alpha_orig = ReboundTools::orig_vars$R_alpha_orig,
  p_E = ReboundTools::orig_vars$p_E,
  eta_orig = ReboundTools::orig_vars$eta_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_omd_orig = ReboundTools::orig_vars$C_dot_omd_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  e_qs_ps_UC_orig = ReboundTools::orig_vars$e_qs_ps_UC_orig,
  e_qo_ps_UC_orig = ReboundTools::orig_vars$e_qo_ps_UC_orig,
  e_qs_ps_C_orig = ReboundTools::orig_vars$e_qs_ps_C_orig,
  e_qo_ps_C_orig = ReboundTools::orig_vars$e_qo_ps_C_orig,
  eta_engr_units_star = ReboundTools::star_vars$eta_engr_units_star,
  E_emb_star = ReboundTools::star_vars$E_emb_star,
  t_life_star = ReboundTools::star_vars$t_life_star,
  C_cap_star = ReboundTools::star_vars$C_cap_star,
  C_dot_om_star = ReboundTools::star_vars$C_dot_om_star,
  C_d_star = ReboundTools::star_vars$C_d_star,
  R_alpha_star = ReboundTools::star_vars$R_alpha_star,
  R_omega_star = ReboundTools::star_vars$R_omega_star,
  eta_star = ReboundTools::star_vars$eta_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  G_dot = ReboundTools::star_vars$G_dot,
  p_E_star = ReboundTools::star_vars$p_E_star,
  p_s_star = ReboundTools::star_vars$p_s_star,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  R_alpha_C_dot_cap_star = ReboundTools::star_vars$R_alpha_C_dot_cap_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  C_dot_s_star = ReboundTools::star_vars$C_dot_s_star,
  C_dot_d_star = ReboundTools::star_vars$C_dot_d_star,
  R_omega_C_dot_d_star = ReboundTools::star_vars$R_omega_C_dot_d_star,
  C_dot_omd_star = ReboundTools::star_vars$C_dot_omd_star,
  M_dot_star = ReboundTools::star_vars$M_dot_star,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  f_Cs_star = ReboundTools::star_vars$f_Cs_star,
  e_qs_ps_UC_star = ReboundTools::star_vars$e_qs_ps_UC_star,
  e_qo_ps_UC_star = ReboundTools::star_vars$e_qo_ps_UC_star,
  e_qs_ps_C_star = ReboundTools::star_vars$e_qs_ps_C_star,
  e_qo_ps_C_star = ReboundTools::star_vars$e_qo_ps_C_star,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star
)

Arguments

.orig_data

An optional data frame containing EEU base data and original data, likely calculated by calc_orig().

r, MJ_engr_unit, p_E

See ReboundTools::eeu_base_params.

R_alpha_orig, C_dot_omd_orig, C_dot_om_star, C_d_star, C_dot_d_star, R_omega_C_dot_d_star, C_dot_omd_star, eta_orig, E_dot_s_orig, q_dot_s_orig, M_dot_orig, C_dot_cap_orig, C_dot_o_orig, e_qs_ps_UC_orig, e_qo_ps_UC_orig, e_qs_ps_C_orig, e_qo_ps_C_orig

See ReboundTools::orig_vars.

R_alpha_star, R_omega_star, eta_engr_units_star, E_emb_star, t_life_star, C_cap_star, eta_star, eta_ratio, S_dot_dev, G_dot, p_s_star, q_dot_s_star, C_dot_cap_star, R_alpha_C_dot_cap_star, E_dot_emb_star, C_dot_s_star, M_dot_star, N_dot_star, C_dot_o_star, f_Cs_star, e_qs_ps_UC_star, e_qo_ps_UC_star, e_qs_ps_C_star, e_qo_ps_C_star, E_dot_s_star, p_E_star

See ReboundTools::star_vars.

Value

A list or data frame of derived rebound values for the star stage (after the emplacement effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star()
#> # A tibble: 3 × 75
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 68 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>, R_alpha_orig <dbl>,
#> #   R_omega_orig <dbl>, p_E <dbl>, p_E_orig <dbl>, eta_orig <dbl>, …