Calculate energy rebound data at the tilde stage

This function calculates energy rebound information for the tilde stage (immediately after the macro effect).

Usage

calc_tilde(
  .bar_data = NULL,
  p_E = ReboundTools::orig_vars$p_E,
  t_life_bar = ReboundTools::bar_vars$t_life_bar,
  R_alpha_bar = ReboundTools::bar_vars$R_alpha_bar,
  R_omega_bar = ReboundTools::bar_vars$R_omega_bar,
  eta_engr_units_bar = ReboundTools::bar_vars$eta_engr_units_bar,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_dot_om_bar = ReboundTools::bar_vars$C_dot_om_bar,
  C_d_bar = ReboundTools::bar_vars$C_d_bar,
  C_dot_d_bar = ReboundTools::bar_vars$C_dot_d_bar,
  C_dot_omd_bar = ReboundTools::bar_vars$C_dot_omd_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  e_qs_ps_UC_bar = ReboundTools::bar_vars$e_qs_ps_UC_bar,
  e_qo_ps_UC_bar = ReboundTools::bar_vars$e_qo_ps_UC_bar,
  e_qs_ps_C_bar = ReboundTools::bar_vars$e_qs_ps_C_bar,
  e_qo_ps_C_bar = ReboundTools::bar_vars$e_qo_ps_C_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar,
  t_life_tilde = ReboundTools::tilde_vars$t_life_tilde,
  R_alpha_tilde = ReboundTools::tilde_vars$R_alpha_tilde,
  R_omega_tilde = ReboundTools::tilde_vars$R_omega_tilde,
  eta_engr_units_tilde = ReboundTools::tilde_vars$eta_engr_units_tilde,
  eta_tilde = ReboundTools::tilde_vars$eta_tilde,
  p_E_tilde = ReboundTools::tilde_vars$p_E_tilde,
  p_s_tilde = ReboundTools::tilde_vars$p_s_tilde,
  C_dot_cap_tilde = ReboundTools::tilde_vars$C_dot_cap_tilde,
  R_alpha_C_dot_cap_tilde = ReboundTools::tilde_vars$R_alpha_C_dot_cap_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  M_dot_tilde = ReboundTools::tilde_vars$M_dot_tilde,
  q_dot_s_tilde = ReboundTools::tilde_vars$q_dot_s_tilde,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  C_dot_s_tilde = ReboundTools::tilde_vars$C_dot_s_tilde,
  C_dot_om_tilde = ReboundTools::tilde_vars$C_dot_om_tilde,
  C_d_tilde = ReboundTools::tilde_vars$C_d_tilde,
  C_dot_d_tilde = ReboundTools::tilde_vars$C_dot_d_tilde,
  R_omega_C_dot_d_tilde = ReboundTools::tilde_vars$R_omega_C_dot_d_tilde,
  C_dot_omd_tilde = ReboundTools::tilde_vars$C_dot_omd_tilde,
  C_dot_o_tilde = ReboundTools::tilde_vars$C_dot_o_tilde,
  f_Cs_tilde = ReboundTools::tilde_vars$f_Cs_tilde,
  e_qs_ps_UC_tilde = ReboundTools::tilde_vars$e_qs_ps_UC_tilde,
  e_qo_ps_UC_tilde = ReboundTools::tilde_vars$e_qo_ps_UC_tilde,
  e_qs_ps_C_tilde = ReboundTools::tilde_vars$e_qs_ps_C_tilde,
  e_qo_ps_C_tilde = ReboundTools::tilde_vars$e_qo_ps_C_tilde,
  N_dot_tilde = ReboundTools::tilde_vars$N_dot_tilde
)

Arguments

.bar_data

An optional data frame containing rebound calculations, original data, star data, hat data, and bar data, likely calculated by calc_bar().

p_E

See ReboundTools::orig_vars

t_life_bar, C_dot_om_bar, C_d_bar, C_dot_d_bar, C_dot_omd_bar, R_alpha_bar, R_omega_bar, eta_engr_units_bar, eta_bar, p_s_bar, C_dot_cap_bar, E_dot_emb_bar, M_dot_bar, q_dot_s_bar, E_dot_s_bar, C_dot_s_bar, C_dot_o_bar, e_qs_ps_UC_bar, e_qo_ps_UC_bar, e_qs_ps_C_bar, e_qo_ps_C_bar, N_dot_bar

See ReboundTools::bar_vars.

t_life_tilde, C_dot_om_tilde, C_d_tilde, C_dot_d_tilde, R_omega_C_dot_d_tilde, C_dot_omd_tilde, R_alpha_tilde, R_omega_tilde, eta_engr_units_tilde, eta_tilde, p_s_tilde, C_dot_cap_tilde, R_alpha_C_dot_cap_tilde, E_dot_emb_tilde, M_dot_tilde, q_dot_s_tilde, E_dot_s_tilde, C_dot_s_tilde, C_dot_o_tilde, f_Cs_tilde, e_qs_ps_UC_tilde, e_qo_ps_UC_tilde, e_qs_ps_C_tilde, e_qo_ps_C_tilde, N_dot_tilde, p_E_tilde

See ReboundTools::tilde_vars.

Value

A list or data frame of derived rebound values for the bar stage (after the income effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde()
#> # A tibble: 3 × 154
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 147 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>, R_alpha_orig <dbl>,
#> #   R_omega_orig <dbl>, p_E <dbl>, p_E_orig <dbl>, eta_orig <dbl>, …