Graph parameters

The list of graph parameters for drawing energy path graphs, cost path graphs, and preference path graphs in the ReboundTools package.

default_graph_params

Format

A string list with 54 entries.

lineend

The line end style.

linejoin

The line join style.

linejoin

The line join style.

dempl_colour

The colour for direct emplacment lines.

emb_colour

The colour for embodied energy lines.

cap_colour

The colour for capital cost lines.

md_colour

The colour for maintenance and disposal lines.

dsub_colour

The colour for direct substitution lines.

isub_colour

The colour for indirect substitution lines.

dinc_colour

The colour for direct income lines.

iinc_colour

The colour for indirect income lines.

prod_colour

The colour for productivity lines.

dempl_size

The size for direct emplacment lines.

emb_size

The size for embodied energy lines.

cap_size

The size for capital cost lines.

md_size

The size for maintenance and disposal lines.

dsub_size

The size for direct substitution lines.

isub_size

The size for indirect substitution lines.

dinc_size

The size for direct income lines.

iinc_size

The size for indirect income lines.

prod_size

The size for productivity lines.

dempl_linetyps

The linetype for direct emplacment lines.

emb_linetype

The linetype for embodied energy lines.

cap_linetype

The linetype for capital cost lines.

md_linetype

The linetype for maintenance and disposal lines.

dsub_linetype

The linetype for direct substitution lines.

isub_linetype

The linetype for indirect substitution lines.

dinc_linetype

The linetype for direct income lines.

iinc_linetype

The linetype for indirect income lines.

prod_linetype

The linetype for productivity lines.

energy_grid_colour

The colour for energy grid lines.

energy_rebound_lines_colour

The colour for energy rebound lines.

cost_grid_colour

The colour for cost grid lines.

cost_ray_colour

The colour for cost rays.

prefs_grid_colour

The colour for preferences grid lines.

prefs_indiff_curve_colour

The colour for indifference curve lines.

energy_grid_size

The size for energy grid lines.

energy_rebound_lines_size

The size for energy rebound lines.

cost_grid_size

The size for cost grid lines.

cost_ray_size

The size for cost rays.

prefs_grid_size

The size for preferences grid lines.

prefs_indiff_curve_size

The size for indifference curve lines.

energy_grid_linetype

The linetype for energy grid lines.

energy_rebound_lines_linetype

The linetype for energy rebound lines.

cost_grid_linetype

The linetype for cost grid lines.

cost_ray_linetype

The linetype for cost rays.

prefs_grid_linetype

The linetype for preferences grid lines.

prefs_indiff_curve_linetype

The linetype for indifference curve lines.

include_start_point

A boolean that tells whether to include the starting point.

start_point_size

The size of the start point.

start_point_shape

The shape of the start point.

include_end_arrow

A boolean that tells whether to include an ending arrow.

arrow_angle

The angle for the arrow head, in degrees.

arrow_length

The length of the arrow. See grid::unit.

arrow_type

The arrow type, "closed" (the default) or "open". See grid::arrow.

Details

This list is passed to several graphing functions. Callers may pass a modified version of this list to change graph appearance.

Examples

default_graph_params

#> $lineend
#> [1] "round"
#> 
#> $linejoin
#> [1] "round"
#> 
#> $dempl_colour
#> [1] "#F5C2C1FF"
#> 
#> $emb_colour
#> [1] "#F5C2C1FF"
#> 
#> $cap_colour
#> [1] "#F5C2C1FF"
#> 
#> $md_colour
#> [1] "#F5C2C1FF"
#> 
#> $dsub_colour
#> [1] "#FAE0C3FF"
#> 
#> $isub_colour
#> [1] "#FAE0C3FF"
#> 
#> $dinc_colour
#> [1] "#CDFDC5FF"
#> 
#> $iinc_colour
#> [1] "#CDFDC5FF"
#> 
#> $prod_colour
#> [1] "#BFC0FAFF"
#> 
#> $dempl_size
#> [1] 1
#> 
#> $emb_size
#> [1] 1
#> 
#> $cap_size
#> [1] 1
#> 
#> $md_size
#> [1] 1
#> 
#> $dsub_size
#> [1] 1
#> 
#> $isub_size
#> [1] 1
#> 
#> $dinc_size
#> [1] 1
#> 
#> $iinc_size
#> [1] 1
#> 
#> $prod_size
#> [1] 1
#> 
#> $dempl_linetype
#> [1] "solid"
#> 
#> $emb_linetype
#> [1] "solid"
#> 
#> $cap_linetype
#> [1] "solid"
#> 
#> $md_linetype
#> [1] "solid"
#> 
#> $dsub_linetype
#> [1] "solid"
#> 
#> $isub_linetype
#> [1] "solid"
#> 
#> $dinc_linetype
#> [1] "solid"
#> 
#> $iinc_linetype
#> [1] "solid"
#> 
#> $prod_linetype
#> [1] "solid"
#> 
#> $energy_grid_colour
#> [1] "gray"
#> 
#> $energy_rebound_lines_colour
#> [1] "gray"
#> 
#> $cost_grid_colour
#> [1] "gray"
#> 
#> $prefs_grid_colour
#> [1] "gray"
#> 
#> $prefs_ray_colour
#> [1] "gray"
#> 
#> $prefs_indiff_grid_colour
#> [1] "gray"
#> 
#> $energy_grid_size
#> [1] 0.1
#> 
#> $energy_rebound_lines_size
#> [1] 0.1
#> 
#> $cost_grid_size
#> [1] 0.1
#> 
#> $prefs_grid_size
#> [1] 0.1
#> 
#> $prefs_ray_size
#> [1] 0.1
#> 
#> $prefs_indiff_grid_size
#> [1] 0.1
#> 
#> $energy_grid_linetype
#> [1] "solid"
#> 
#> $energy_rebound_lines_linetype
#> [1] "solid"
#> 
#> $cost_grid_linetype
#> [1] "solid"
#> 
#> $prefs_grid_linetype
#> [1] "solid"
#> 
#> $prefs_ray_linetype
#> [1] "solid"
#> 
#> $prefs_indiff_grid_linetype
#> [1] "solid"
#> 
#> $include_start_point
#> [1] TRUE
#> 
#> $start_point_size
#> [1] 3
#> 
#> $start_point_shape
#> [1] 16
#> 
#> $include_end_arrow
#> [1] TRUE
#> 
#> $arrow_angle
#> [1] 20
#> 
#> $arrow_length
#> [1] 0.1inches
#> 
#> $arrow_type
#> [1] "closed"
#>