EEU base parameters

This is the list of the required parameters for a rebound analysis. These parameters must be supplied for a rebound analysis to commence. The names in this list are the assumed names throughout this package, but they can always be overridden by function arguments.

Usage

eeu_base_params

Format

A string list with 14 entries.

reference

A string for a reference for this case, e.g. bibliographic entry for a paper in which this case appears.

case

A string to identify the case being analyzed, e.g., "Lamp".

original

A string to identify the original device, e.g., "Incandescent".

upgrade

A string to identify the upgraded device, e.g., "LED".

r

The discount rate, in units of 1/year. E.g., 3.0% should be entered as 0.03.

service_unit

A string to identify the unit of the energy service, e.g., "miles" in "miles/gal" or "lm-hr" in "lm-hr/kW-hr".

energy_engr_unit

A string to identify the energy units of the service, e.g., "gal" in "miles/gal" or "kW-hr" in "lm-hr/kW-hr".

MJ_engr_unit

A unit conversion factor: the number of MJ per engineering unit for the service efficiency. For example, if the service efficiency is given in miles/gallon, MJ_engr_unit should be 126.6 MJ/gallon. This unit conversion number is used in calculating the actual service efficiency.

I_E

The energy intensity of the economy [MJ/$].

k

The macro effect factor [--].

p_E_engr_units

The price of energy in engineering units, e.g., $/gal or $/kW-hr [$/energy_engr_unit].

e_qs_ps_UC_orig

The original uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [--].

e_qs_M

The income ("M") elasticity ("e") of energy service ("qs") consumption [--].

e_qo_M

The income ("M") elasticity ("e") of other goods ("qo") consumption [--].

Details

In the descriptions for each parameter, typical units are provided for energy efficiency upgrades for cars and electric lamps.

Note the following meanings for the default parameter names:

orig

The original device, prior to the EEU.

star

After the emplacement effect and before the substitution effect.

hat

After the substitution effect and before the income effect.

bar

After the income effect and before the macro effect.

tilde

After the macro effect.

I

An economic intensity (per $).

p

Price.

E

Energy, typically final energy.

s

Energy service.

o

Other goods.

eps

Elasticity.

UC

Uncompensated.

dot

Signifies a rate, typically per yr.

q

Quantity of energy service or other goods consumed.

M

Income.

C

Cost.

t

Expected device lifetime.

Examples

eeu_base_params
#> $reference
#> [1] "Reference"
#> 
#> $case
#> [1] "Case"
#> 
#> $original
#> [1] "Original"
#> 
#> $upgrade
#> [1] "Upgrade"
#> 
#> $r
#> [1] "r"
#> 
#> $service_unit
#> [1] "service_unit"
#> 
#> $energy_engr_unit
#> [1] "energy_engr_unit"
#> 
#> $MJ_engr_unit
#> [1] "MJ/energy_engr_unit"
#> 
#> $I_E
#> [1] "I_E"
#> 
#> $k
#> [1] "k"
#> 
#> $p_E_engr_units
#> [1] "p_E_engr_units"
#> 
#> $e_qs_ps_UC_orig
#> [1] "e_qs_ps_UC_orig"
#> 
#> $e_qs_M
#> [1] "e_qs_M"
#> 
#> $e_qo_M
#> [1] "e_qo_M"
#>