A data frame of energy rebound paths

Make a data frame of segments for an energy rebound graph. Each stage of the rebound process is represented in the data frame.

Usage

energy_paths(
  .rebound_data,
  indexed = FALSE,
  graph_params = ReboundTools::path_graph_params,
  rebound_segments = ReboundTools::rebound_segments,
  graph_type = ReboundTools::graph_types$energy,
  k = ReboundTools::eeu_base_params$k,
  I_E = ReboundTools::eeu_base_params$I_E,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars$Delta_E_dot_emb_star,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars$Delta_E_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  Delta_E_dot_s_bar = ReboundTools::Delta_vars$Delta_E_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  graph_df_colnames = ReboundTools::graph_df_colnames
)

Arguments

.rebound_data

A data frame of rebound analysis results, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_params

See ReboundTools::graph_params.

rebound_segments

See ReboundTools::rebound_segments.

graph_type

See ReboundTools::graph_types.

k, I_E

See ReboundTools::eeu_base_params.

E_dot_s_orig, E_dot_emb_orig, C_dot_md_orig, C_dot_o_orig

See ReboundTools::orig_vars.

S_dot_dev

See ReboundTools::star_vars.

Delta_E_dot_emb_star, Delta_C_dot_md_star, Delta_E_dot_s_hat, Delta_C_dot_o_hat, Delta_E_dot_s_bar, Delta_C_dot_o_bar, N_dot_hat

See ReboundTools::Delta_vars.

graph_df_colnames

See ReboundTools::graph_df_colnames.

Value

A data frame with energy rebound path segments.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  energy_paths()
#> # A tibble: 16 × 14
#>    Reference Case  Origi…¹ Upgrade graph…² line_…³ colour linew…⁴ linet…⁵      x
#>    <chr>     <chr> <chr>   <chr>   <chr>   <chr>   <chr>    <dbl> <chr>    <dbl>
#>  1 None yet  Car   Ford F… Ford F… Energy  dempl   #1507…     1   solid   7.31e4
#>  2 None yet  Lamp  Incand… LED     Energy  dempl   #1507…     1   solid   2.37e2
#>  3 None yet  Car   Ford F… Ford F… Energy  emb     #1507…     1.5 11      4.35e4
#>  4 None yet  Lamp  Incand… LED     Energy  emb     #1507…     1.5 11      2.55e1
#>  5 None yet  Car   Ford F… Ford F… Energy  md      #1507…     1   solid   4.35e4
#>  6 None yet  Lamp  Incand… LED     Energy  md      #1507…     1   solid   2.55e1
#>  7 None yet  Car   Ford F… Ford F… Energy  isub    #9914…     1   solid   4.35e4
#>  8 None yet  Lamp  Incand… LED     Energy  isub    #9914…     1   solid   2.55e1
#>  9 None yet  Car   Ford F… Ford F… Energy  dsub    #9914…     1   solid   4.35e4
#> 10 None yet  Lamp  Incand… LED     Energy  dsub    #9914…     1   solid   2.55e1
#> 11 None yet  Car   Ford F… Ford F… Energy  dinc    #E76F…     1   solid   4.44e4
#> 12 None yet  Lamp  Incand… LED     Energy  dinc    #E76F…     1   solid   6.22e1
#> 13 None yet  Car   Ford F… Ford F… Energy  iinc    #E76F…     1   solid   4.59e4
#> 14 None yet  Lamp  Incand… LED     Energy  iinc    #E76F…     1   solid   6.22e1
#> 15 None yet  Car   Ford F… Ford F… Energy  macro   #F7E2…     1   solid   4.59e4
#> 16 None yet  Lamp  Incand… LED     Energy  macro   #F7E2…     1   solid   6.22e1
#> # … with 4 more variables: y <dbl>, xend <dbl>, yend <dbl>, end_arrow <lgl>,
#> #   and abbreviated variable names ¹​Original, ²​graph_type, ³​line_name,
#> #   ⁴​linewidth, ⁵​linetype