A data frame of expenditure path segments

Makes a data frame of segments for an expenditure path graph. Each stage of the rebound process is represented in the data frame.

expenditure_paths(
  .rebound_data,
  indexed = FALSE,
  graph_params = ReboundTools::path_graph_params,
  rebound_segments = ReboundTools::rebound_segments,
  graph_type = ReboundTools::graph_types$expenditure,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  Delta_C_dot_cap_star = ReboundTools::Delta_vars$Delta_C_dot_cap_star,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_C_dot_s_hat = ReboundTools::Delta_vars$Delta_C_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_C_dot_s_bar = ReboundTools::Delta_vars$Delta_C_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  graph_df_colnames = ReboundTools::graph_df_colnames
)

Arguments

.rebound_data

A data frame of rebound analysis results, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_params

See ReboundTools::graph_params.

rebound_segments

See ReboundTools::rebound_segments.

graph_type

See ReboundTools::graph_types.

C_dot_s_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig

See ReboundTools::orig_vars.

G_dot

See ReboundTools::star_vars.

Delta_C_dot_cap_star, Delta_C_dot_md_star, Delta_C_dot_s_hat, Delta_C_dot_o_hat, Delta_C_dot_s_bar, Delta_C_dot_o_bar

See ReboundTools::Delta_vars.

graph_df_colnames

See ReboundTools::graph_df_colnames.

Value

A data frame with cost rebound path segments.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  expenditure_paths()
#> # A tibble: 14 × 14
#>    Refere…¹ Case  Origi…² Upgrade graph…³ line_…⁴ colour linew…⁵ linet…⁶       x
#>    <chr>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>    <dbl> <chr>     <dbl>
#>  1 None yet Car   Ford F… Ford F… Expend… dempl   #1507…     1   solid   1.28e+3
#>  2 None yet Lamp  Incand… LED     Expend… dempl   #1507…     1   solid   8.90e+0
#>  3 None yet Car   Ford F… Ford F… Expend… cap     #1507…     1.5 11      7.59e+2
#>  4 None yet Lamp  Incand… LED     Expend… cap     #1507…     1.5 11      9.61e-1
#>  5 None yet Car   Ford F… Ford F… Expend… md      #1507…     1   solid   7.59e+2
#>  6 None yet Lamp  Incand… LED     Expend… md      #1507…     1   solid   9.61e-1
#>  7 None yet Car   Ford F… Ford F… Expend… isub    #9914…     1   solid   7.59e+2
#>  8 None yet Lamp  Incand… LED     Expend… isub    #9914…     1   solid   9.61e-1
#>  9 None yet Car   Ford F… Ford F… Expend… dsub    #9914…     1   solid   7.59e+2
#> 10 None yet Lamp  Incand… LED     Expend… dsub    #9914…     1   solid   9.61e-1
#> 11 None yet Car   Ford F… Ford F… Expend… dinc    #E76F…     1   solid   7.74e+2
#> 12 None yet Lamp  Incand… LED     Expend… dinc    #E76F…     1   solid   2.34e+0
#> 13 None yet Car   Ford F… Ford F… Expend… iinc    #E76F…     1   solid   8.02e+2
#> 14 None yet Lamp  Incand… LED     Expend… iinc    #E76F…     1   solid   2.34e+0
#> # … with 4 more variables: y <dbl>, xend <dbl>, yend <dbl>, end_arrow <lgl>,
#> #   and abbreviated variable names ¹​Reference, ²​Original, ³​graph_type,
#> #   ⁴​line_name, ⁵​linewidth, ⁶​linetype