EEU hat data

This is the list of the derived variables at the hat stage (after substitution effect) of a rebound analysis.

Usage

hat_vars

Format

A string list with 28 entries.

tau_alpha_hat

The post-substitution effect discount multiplier for beginning-of-life expenses, such as capital cost. Same as tau_alpha_star.

tau_omega_hat

The post-substitution effect discount multiplier for end-of-life expenses, such as disposal cost. Same as tau_omega_star.

eta_hat

Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], exactly eta_star.

p_s_hat

The energy service price after the substitution effect [$/service], exactly p_s_star.

C_dot_cap_hat

The capital expenditure rate of the device without discounting [$/yr], calculated by C_cap/t_life.

tau_alpha_C_dot_cap_hat

The product of tau_alpha and the undiscounted capital cost rate, itself annualized and discounted in [$/yr].

C_dot_om_hat

The operations and maintenance expenditure rate of the device [$/yr].

C_d_hat

The disposal cost for the device [$/yr].

C_dot_d_hat

The disposal cost rate of the device without discounting [$/yr].

tau_omega_C_dot_d_hat

The product of tau_omega and the undiscounted disposal cost rate, itself annualized and discounted in [$/yr].

C_dot_omd_hat

The operations, maintenance, and disposal expenditure rate of the device [$/yr].

E_dot_emb_hat

The embodied energy rate after the substitution effect [MJ/yr], exactly E_dot_emb_star.

M_dot_hat

Real income after the substitution effect [MJ/yr], exactly M_dot_star.

q_dot_s_hat

The rate of energy service consumption after the substitution effect\ [service/yr], calculated by q_dot_s_star * eta_ratio^(-e_qs_ps_C).

p_E_hat

The price of energy [$/MJ].

E_dot_s_hat

The rate of energy consumption after the substitution effect\ [service/yr], calculated by q_dot_s_hat / eta_hat.

C_dot_g_hat

The rate of other goods expenditures after the substitution effect [$/yr], calculated by C_dot_g_star * eta_ratio^(-e_qo_ps_C).

f_Cs_hat

The post-substitution effect fraction of the energy and other budget spent on the energy service [–], calculated by C_dot_s_star / (C_dot_s_star + C_dot_g_star).

e_qs_ps_C_hat

The post-substitution effect compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [–], calculated by e_qs_ps_UC_orig + f_Cs_orig*e_qs_M.

e_qg_ps_C_hat

The post-substitution effect compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [–], calculated by f_Cs_orig*(f_Cs_orig + e_qs_ps_UC_orig) / (f_Cs_orig - 1).

e_qs_ps_UC_hat

The post-substitution effect uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [–].

e_qg_ps_UC_hat

The post-substitution effect uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of other goods ("qo") consumption (cross-price elasticity) [–].

N_dot_hat

The freed cash rate [$/yr], calculated by G_dot - (C_dot_cap_star - C_dot_cap_orig) - (C_dot_md_star - C_dot_md_orig).

M_dot_hat_prime

Modified M_dot for the income effect [$/yr], calculated by M_dot_hat - C_dot_cap_orig - C_dot_md_orig - G_dot + p_E*(E_dot_s_hat - E_dot_s_star) + (C_dot_g_hat - C_dot_g_star)

[service/yr]: R:service/yr%5C [service/yr]: R:service/yr%5C

Examples

hat_vars
#> $t_life_hat
#> [1] "t_life_hat"
#> 
#> $tau_alpha_hat
#> [1] "tau_alpha_hat"
#> 
#> $tau_omega_hat
#> [1] "tau_omega_hat"
#> 
#> $eta_engr_units_hat
#> [1] "eta_engr_units_hat"
#> 
#> $eta_hat
#> [1] "eta_hat"
#> 
#> $p_s_hat
#> [1] "p_s_hat"
#> 
#> $C_cap_hat
#> [1] "C_cap_hat"
#> 
#> $C_dot_cap_hat
#> [1] "C_dot_cap_hat"
#> 
#> $tau_alpha_C_dot_cap_hat
#> [1] "tau_alpha_C_dot_cap_hat"
#> 
#> $E_dot_emb_hat
#> [1] "E_dot_emb_hat"
#> 
#> $M_dot_hat
#> [1] "M_dot_hat"
#> 
#> $q_dot_s_hat
#> [1] "q_dot_s_hat"
#> 
#> $p_E_hat
#> [1] "p_E_hat"
#> 
#> $E_dot_s_hat
#> [1] "E_dot_s_hat"
#> 
#> $C_dot_s_hat
#> [1] "C_dot_s_hat"
#> 
#> $C_dot_om_hat
#> [1] "C_dot_om_hat"
#> 
#> $C_d_hat
#> [1] "C_d_hat"
#> 
#> $C_dot_d_hat
#> [1] "C_dot_d_hat"
#> 
#> $tau_omega_C_dot_d_hat
#> [1] "tau_omega_C_dot_d_hat"
#> 
#> $C_dot_omd_hat
#> [1] "C_dot_omd_hat"
#> 
#> $C_dot_g_hat
#> [1] "C_dot_g_hat"
#> 
#> $f_Cs_hat
#> [1] "f_Cs_hat"
#> 
#> $e_qs_ps_C_hat
#> [1] "e_qs_ps_C_hat"
#> 
#> $e_qs_ps_UC_hat
#> [1] "e_qs_ps_UC_hat"
#> 
#> $e_qg_ps_C_hat
#> [1] "e_qg_ps_C_hat"
#> 
#> $e_qg_ps_UC_hat
#> [1] "e_qg_ps_UC_hat"
#> 
#> $N_dot_hat
#> [1] "N_dot_hat"
#> 
#> $M_dot_hat_prime
#> [1] "M_dot_hat_prime"
#>