Indifference curve lines

This function calculates indifference curves to be drawn on the consumption path graph.

Usage

indifference_lines(
  .rebound_data,
  graph_type = ReboundTools::graph_types$consumption,
  graph_params = ReboundTools::path_graph_params,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  sigma = ReboundTools::orig_vars$sigma,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  C_dot_g_bar = ReboundTools::bar_vars$C_dot_g_bar
)

Arguments

.rebound_data

A data frame of rebound information, likely created by rebound_analysis().

graph_type

See ReboundTools::graph_types. Default is ReboundTools::graph_types$consumption.

graph_params

Parameters that control the appearance of the graph. Default is ReboundTools::path_graph_params.

q_dot_s_orig, C_dot_g_orig, f_Cs_orig, sigma

See ReboundTools::orig_vars.

q_dot_s_hat

See ReboundTools::hat_vars.

q_dot_s_bar, C_dot_g_bar

See ReboundTools::bar_vars.

Value

A data frame of indifference curves for a consumption path graph.

Details

The consumption path graph is always indexed, so there is no indexed argument.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  indifference_lines()
#> # A tibble: 835 × 15
#>    Reference Case  Original    Upgrade     graph_type line_name colour linewidth
#>    <chr>     <chr> <chr>       <chr>       <chr>      <chr>     <chr>      <dbl>
#>  1 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  2 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  3 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  4 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  5 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  6 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  7 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  8 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#>  9 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#> 10 None yet  Car   Ford Fusion Ford Fusio… Consumpti… orig      black        0.5
#> # ℹ 825 more rows
#> # ℹ 7 more variables: linetype <chr>, qs1_qs0 <dbl>, Cg1_Cg0 <dbl>,
#> #   f_Cs_orig <dbl>, sigma <dbl>, x <dbl>, y <dbl>