Create iso-budget (expenditure) lines for a preferences graph

This function creates a data frame of constant-budget (expenditure) lines at various rebound stages. The iso lines are described by slope and intercept.

iso_budget_lines_prefs(
  .rebound_data,
  graph_type = ReboundTools::graph_types$preferences,
  graph_params = ReboundTools::path_graph_params,
  p_s_orig = ReboundTools::orig_vars$p_s_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  p_s_star = ReboundTools::star_vars$p_s_star,
  G_dot = ReboundTools::star_vars$G_dot,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  Delta_q_dot_s_hat = ReboundTools::Delta_vars$Delta_q_dot_s_hat,
  Delta_C_dot_cap_star = ReboundTools::Delta_vars$Delta_C_dot_cap_star,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat
)

Arguments

.rebound_data
A data frame of rebound information, likely created by rebound_analysis().

graph_type
See ReboundTools::graph_types.

graph_params
See ReboundTools::graph_params.

p_s_orig, q_dot_s_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig, M_dot_orig
See ReboundTools::orig_vars.

p_s_star, G_dot
See ReboundTools::star_vars.

q_dot_s_hat, C_dot_o_hat
See ReboundTools::hat_vars.

Delta_q_dot_s_hat, Delta_C_dot_cap_star, Delta_C_dot_md_star, Delta_C_dot_o_hat
See ReboundTools::Delta_vars.

Value

A data frame of iso-budget lines for a preferences graph.

Details

The preferences graph is always indexed, so there is no indexed argument.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  iso_budget_lines_prefs()

#> # A tibble: 10 x 11
#>    Reference Case  Original  Upgrade  graph_type line_name colour  size linetype
#>    <chr>     <chr> <chr>     <chr>    <chr>      <chr>     <chr>  <dbl> <chr>   
#>  1 None yet  Car   Ford Fus… Ford Fu… Preferenc… orig      black    0.1 solid   
#>  2 None yet  Lamp  Incandes… LED      Preferenc… orig      black    0.1 solid   
#>  3 None yet  Car   Ford Fus… Ford Fu… Preferenc… star      black    0.1 solid   
#>  4 None yet  Lamp  Incandes… LED      Preferenc… star      black    0.1 solid   
#>  5 None yet  Car   Ford Fus… Ford Fu… Preferenc… hat       black    0.1 solid   
#>  6 None yet  Lamp  Incandes… LED      Preferenc… hat       black    0.1 solid   
#>  7 None yet  Car   Ford Fus… Ford Fu… Preferenc… bar       black    0.1 solid   
#>  8 None yet  Lamp  Incandes… LED      Preferenc… bar       black    0.1 solid   
#>  9 None yet  Car   Ford Fus… Ford Fu… Preferenc… ray       black    0.1 solid   
#> 10 None yet  Lamp  Incandes… LED      Preferenc… ray       black    0.1 solid   
#> # … with 2 more variables: slope <dbl>, intercept <dbl>