Iso-energy lines for energy rebound path graphs

This function creates a data frame of constant-energy lines at various rebound stages. The iso lines are described by slope and intercept.

Usage

iso_energy_lines(
  .rebound_data,
  indexed = FALSE,
  graph_type = ReboundTools::graph_types$energy,
  graph_params = ReboundTools::path_graph_params,
  I_E = ReboundTools::eeu_base_params$I_E,
  k = ReboundTools::eeu_base_params$k,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  C_dot_omd_orig = ReboundTools::orig_vars$C_dot_omd_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  C_dot_omd_star = ReboundTools::star_vars$C_dot_omd_star,
  C_dot_g_star = ReboundTools::star_vars$C_dot_g_star,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  C_dot_omd_hat = ReboundTools::hat_vars$C_dot_omd_hat,
  C_dot_g_hat = ReboundTools::hat_vars$C_dot_g_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  C_dot_omd_bar = ReboundTools::bar_vars$C_dot_omd_bar,
  C_dot_g_bar = ReboundTools::bar_vars$C_dot_g_bar,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  C_dot_omd_tilde = ReboundTools::tilde_vars$C_dot_omd_tilde,
  C_dot_g_tilde = ReboundTools::tilde_vars$C_dot_g_tilde
)

Arguments

.rebound_data

A data frame of rebound information, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_type

See ReboundTools::graph_types.

graph_params

See ReboundTools::path_graph_params.

I_E, k

See ReboundTools::eeu_base_params.

E_dot_s_orig, E_dot_emb_orig, C_dot_omd_orig, C_dot_g_orig

See ReboundTools::orig_vars.

S_dot_dev, E_dot_s_star, E_dot_emb_star, C_dot_omd_star, C_dot_g_star

See ReboundTools::star_vars.

E_dot_s_hat, E_dot_emb_hat, C_dot_omd_hat, C_dot_g_hat, N_dot_hat

See ReboundTools::hat_vars.

E_dot_s_bar, E_dot_emb_bar, C_dot_omd_bar, C_dot_g_bar

See ReboundTools::bar_vars.

E_dot_s_tilde, E_dot_emb_tilde, C_dot_omd_tilde, C_dot_g_tilde

See ReboundTools::tilde_vars.

Value

A data frame containing iso-energy lines to be drawn on an energy rebound path graph.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  iso_energy_lines()
#> # A tibble: 18 × 11
#>    Reference Case       Original   Upgrade graph_type line_name colour linewidth
#>    <chr>     <chr>      <chr>      <chr>   <chr>      <chr>     <chr>      <dbl>
#>  1 None yet  Car        Ford Fusi… Ford F… Energy     Re = 100% black        0.3
#>  2 None yet  Lamp       Incandesc… LED     Energy     Re = 100% black        0.3
#>  3 None yet  Car, r = 0 Ford Fusi… Ford F… Energy     Re = 100% black        0.3
#>  4 None yet  Car        Ford Fusi… Ford F… Energy     Re = 0%   black        0.3
#>  5 None yet  Lamp       Incandesc… LED     Energy     Re = 0%   black        0.3
#>  6 None yet  Car, r = 0 Ford Fusi… Ford F… Energy     Re = 0%   black        0.3
#>  7 None yet  Car        Ford Fusi… Ford F… Energy     Re_empl   black        0.1
#>  8 None yet  Lamp       Incandesc… LED     Energy     Re_empl   black        0.1
#>  9 None yet  Car, r = 0 Ford Fusi… Ford F… Energy     Re_empl   black        0.1
#> 10 None yet  Car        Ford Fusi… Ford F… Energy     Re_sub    black        0.1
#> 11 None yet  Lamp       Incandesc… LED     Energy     Re_sub    black        0.1
#> 12 None yet  Car, r = 0 Ford Fusi… Ford F… Energy     Re_sub    black        0.1
#> 13 None yet  Car        Ford Fusi… Ford F… Energy     Re_inc    black        0.1
#> 14 None yet  Lamp       Incandesc… LED     Energy     Re_inc    black        0.1
#> 15 None yet  Car, r = 0 Ford Fusi… Ford F… Energy     Re_inc    black        0.1
#> 16 None yet  Car        Ford Fusi… Ford F… Energy     Re_tot    black        0.1
#> 17 None yet  Lamp       Incandesc… LED     Energy     Re_tot    black        0.1
#> 18 None yet  Car, r = 0 Ford Fusi… Ford F… Energy     Re_tot    black        0.1
#> # ℹ 3 more variables: linetype <chr>, slope <dbl>, intercept <dbl>