Create a data frame of iso-expenditure lines

This function creates a data frame of constant expenditure lines at various rebound stages. The iso lines are described by slope and intercept.

Usage

iso_expenditure_lines(
  .rebound_data,
  indexed = FALSE,
  graph_type = ReboundTools::graph_types$expenditure,
  graph_params = ReboundTools::path_graph_params,
  tau_alpha_orig = ReboundTools::orig_vars$tau_alpha_orig,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_omd_orig = ReboundTools::orig_vars$C_dot_omd_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  C_dot_s_hat = ReboundTools::hat_vars$C_dot_s_hat,
  tau_alpha_star = ReboundTools::star_vars$tau_alpha_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  C_dot_omd_star = ReboundTools::star_vars$C_dot_omd_star,
  C_dot_g_star = ReboundTools::star_vars$C_dot_g_star,
  tau_alpha_hat = ReboundTools::hat_vars$tau_alpha_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_omd_hat = ReboundTools::hat_vars$C_dot_omd_hat,
  C_dot_g_hat = ReboundTools::hat_vars$C_dot_g_hat
)

Arguments

.rebound_data

A data frame of rebound information, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_type

See ReboundTools::graph_types.

graph_params

See ReboundTools::graph_params.

tau_alpha_orig, C_dot_s_orig, C_dot_cap_orig, C_dot_omd_orig, C_dot_g_orig

See ReboundTools::orig_vars.

G_dot

See ReboundTools::star_vars.

tau_alpha_star, C_dot_cap_star, C_dot_omd_star, C_dot_g_star

See ReboundTools::star_vars.

tau_alpha_hat, C_dot_s_hat, C_dot_cap_hat, C_dot_omd_hat, C_dot_g_hat

See ReboundTools::hat_vars.

Value

A data frame of iso-expenditure lines for an expenditure path graph.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  iso_expenditure_lines()
#> # A tibble: 12 × 11
#>    Reference Case       Original   Upgrade graph_type line_name colour linewidth
#>    <chr>     <chr>      <chr>      <chr>   <chr>      <chr>     <chr>      <dbl>
#>  1 None yet  Car        Ford Fusi… Ford F… Expenditu… orig      black        0.3
#>  2 None yet  Lamp       Incandesc… LED     Expenditu… orig      black        0.3
#>  3 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… orig      black        0.3
#>  4 None yet  Car        Ford Fusi… Ford F… Expenditu… G_dot     black        0.3
#>  5 None yet  Lamp       Incandesc… LED     Expenditu… G_dot     black        0.3
#>  6 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… G_dot     black        0.3
#>  7 None yet  Car        Ford Fusi… Ford F… Expenditu… star      black        0.3
#>  8 None yet  Lamp       Incandesc… LED     Expenditu… star      black        0.3
#>  9 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… star      black        0.3
#> 10 None yet  Car        Ford Fusi… Ford F… Expenditu… hat       black        0.3
#> 11 None yet  Lamp       Incandesc… LED     Expenditu… hat       black        0.3
#> 12 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… hat       black        0.3
#> # ℹ 3 more variables: linetype <chr>, slope <dbl>, intercept <dbl>