Key analysis variables

This is the list of key rebound analysis variables. These variables are differenced later in the analysis.

Usage

key_analysis_vars

Format

A string list with 21 entries.

t_life

The expected lifetime of the device.

tau_alpha

A parameter that accounts for discounting of beginning-of-life purchases.

tau_omega

A parameter that accounts for discounting of end-of-life purchases.

eta_engr_units

Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service, in engineering units.

eta

Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service.

p_s

Energy service price [service/MJ], calculated by p_E/eta.

q_dot_s

The rate of energy service consumption [service/yr], calculated by eta*E_dot_s.

p_E

The price of energy [$/MJ]

E_dot_s

The rate of final energy consumption by the energy conversion device [MJ/yr].

E_dot_emb

The rate of embodied energy demand by the energy conversion device [MJ/yr], calculated by E_emb/t_life.

C_dot_s

The expenditure rate of energy consumption by the device [$/yr], calculated by p_s*q_dot_s.

C_dot_cap

The capital expenditure rate of the device without discounting [$/yr], calculated by C_cap/t_life.

C_dot_om

The operations and maintenance expenditure rate of the device [$/yr].

C_d

The disposal cost for the device [$/yr].

C_dot_d

The disposal cost rate of the device without discounting [$/yr].

tau_omega_C_dot_d

A parameter that accounts for discounting of disposal costs.

C_dot_omd

The operations, maintenance, and disposal expenditure rate of the device [$/yr].

C_dot_g

The other goods consumption rate [$/yr], calculated, initially, as a residual of the budget constraint.

N_dot

Net income [$/yr].

M_dot

Real income [$/yr].

Examples

key_analysis_vars
#>                 M_dot                   p_E                t_life 
#>               "M_dot"                 "p_E"              "t_life" 
#>             tau_alpha             tau_omega        eta_engr_units 
#>           "tau_alpha"           "tau_omega"      "eta_engr_units" 
#>                   eta                   p_s             E_dot_emb 
#>                 "eta"                 "p_s"           "E_dot_emb" 
#>             C_dot_cap   tau_alpha_C_dot_cap              C_dot_om 
#>           "C_dot_cap" "tau_alpha_C_dot_cap"            "C_dot_om" 
#>                   C_d               C_dot_d     tau_omega_C_dot_d 
#>                 "C_d"             "C_dot_d"   "tau_omega_C_dot_d" 
#>             C_dot_omd               E_dot_s               C_dot_s 
#>           "C_dot_omd"             "E_dot_s"             "C_dot_s" 
#>                 N_dot               q_dot_s               C_dot_g 
#>               "N_dot"             "q_dot_s"             "C_dot_g"