Load EEU data from a sheet in an Excel workbook

Energy efficiency upgrade (EEU) data are often stored in an Excel workbook. This function reads those data and confirms that all required columns are present. If not all columns are present, an error is thrown.

load_eeu_data(
  path = sample_eeu_data_path(),
  sheet = ReboundTools::eeu_data_table$eeu_data_sheet,
  expected_col_names = ReboundTools::eeu_base_params
)

Arguments

path

The path to the sample data spreadsheet. Default is sample_eeu_data_path().

sheet

The name of the tab in the sample data spreadsheet. See ReboundTools::eeu_data_table.

expected_col_names

Names of columns that must be present in the target worksheet.

Value

A data frame of EEU data.

Examples

load_eeu_data()
#> # A tibble: 2 × 27
#>   Reference Case  Original   Upgrade servi…¹ energ…² MJ/en…³   I_E     k p_E_e…⁴
#>   <chr>     <chr> <chr>      <chr>   <chr>   <chr>     <dbl> <dbl> <dbl>   <dbl>
#> 1 None yet  Car   Ford Fusi… Ford F… mile    gal     1.27e+2  3.39     1 2.21e+0
#> 2 None yet  Lamp  Incandesc… LED     lm-hr   W-hr    3.6 e-3  3.39     1 1.35e-4
#> # … with 17 more variables: e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   and abbreviated variable names ¹​service_unit, ²​energy_engr_unit,
#> #   ³​`MJ/energy_engr_unit`, ⁴​p_E_engr_units