Run parametric analyses

After defining a base case for rebound analyses, parametric analyses and sensitivity studies are often helpful. This function provides a convenient way to do parametric studies.

Usage

parametric_analysis(
  original_cases,
  parameterization,
  include_orig_point = TRUE,
  case_colname = ReboundTools::eeu_base_params$case,
  point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
  orig = ReboundTools::parametric_analysis_point_types$orig,
  sweep = ReboundTools::parametric_analysis_point_types$sweep
)

Arguments

original_cases

A data frame with baseline data from which parametric studies are desired. This data frame is expected to contain 1 row, a single case.

parameterization

A named list of lists. See details.

include_orig_point

A boolean that tells whether the original point for each case (from original_cases) should be included in the output. Default is TRUE.

case_colname

See ReboundTools::eeu_base_params.

point_type_colname, orig, sweep

See ReboundTools::parametric_analysis_point_types.

Value

A data frame containing results of parametric studies.

Details

original_cases must have the same structure as the example data frames in this package. See the file at sample_eeu_data_path(). Failures will occur if original_cases contains columns calculated by rebound_analysis().

parameterization is a list and must have the following structure. Top-level elements are named for cases. Case names must match the case column in original_cases. Not every case in original_cases must be present. Each top-level element in parameterization must itself also be a list. Each 2nd-level list must contain named vectors, with each name corresponding to a parameter in original_cases that will be swept in the parametric study. If more than one entry is included in a 2nd-level list, all combinations of values will be used, via expand.grid(). Note that the original value of each parameter (in original_cases) will not be inserted into the parameterization lists, so be sure to include that value, if desired.

Note that the case names in parameterization (i.e., the names of the top-level entries) are filtering toward original_cases. I.e., the output of this function will contain only those cases in parameterization.

Examples

car_case <- load_eeu_data() %>% 
  dplyr::filter(.data[[ReboundTools::eeu_base_params$case]] == "Car")
# This example creates a parametric study for the "Car" case only,
# sweeping through values of both `k` and `p_E_engr_units`.
params <- list(Car = list(k = seq(0, 2, by = 1), 
                          p_E_engr_units = seq(1.5, 2.5, by = 0.25)))
res <- parametric_analysis(car_case, params)
dplyr::glimpse(res)
#> Rows: 17
#> Columns: 257
#> $ Reference                       <chr> "None yet", "None yet", "None yet", "N…
#> $ Case                            <chr> "Car", "Car", "Car", "Car", "Car", "Ca…
#> $ Original                        <chr> "Ford Fusion", "Ford Fusion", "Ford Fu…
#> $ Upgrade                         <chr> "Ford Fusion Hybrid", "Ford Fusion Hyb…
#> $ r                               <dbl> 0.03, 0.03, 0.03, 0.03, 0.03, 0.03, 0.…
#> $ service_unit                    <chr> "mile", "mile", "mile", "mile", "mile"…
#> $ energy_engr_unit                <chr> "gal", "gal", "gal", "gal", "gal", "ga…
#> $ `MJ/energy_engr_unit`           <dbl> 126.6216, 126.6216, 126.6216, 126.6216…
#> $ I_E                             <dbl> 3.389339, 3.389339, 3.389339, 3.389339…
#> $ k                               <dbl> 1, 1, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1,…
#> $ p_E_engr_units                  <dbl> 2.21, 2.21, 1.50, 1.50, 1.50, 1.75, 1.…
#> $ e_qs_ps_UC_orig                 <dbl> -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0…
#> $ e_qs_M                          <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ e_qg_M                          <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ eta_engr_units_orig             <dbl> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25…
#> $ eta_engr_units_star             <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42…
#> $ q_dot_s_orig                    <dbl> 14425, 14425, 14425, 14425, 14425, 144…
#> $ M_dot_orig                      <dbl> 27401.28, 27401.28, 27401.28, 27401.28…
#> $ C_cap_orig                      <dbl> 28216.1, 28216.1, 28216.1, 28216.1, 28…
#> $ C_cap_star                      <dbl> 27523.4, 27523.4, 27523.4, 27523.4, 27…
#> $ C_dot_om_orig                   <dbl> 2861.134, 2861.134, 2861.134, 2861.134…
#> $ C_d_orig                        <dbl> 100, 100, 100, 100, 100, 100, 100, 100…
#> $ C_dot_om_star                   <dbl> 2774.668, 2774.668, 2774.668, 2774.668…
#> $ C_d_star                        <dbl> 101, 101, 101, 101, 101, 101, 101, 101…
#> $ E_emb_orig                      <dbl> 34000, 34000, 34000, 34000, 34000, 340…
#> $ t_life_orig                     <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14…
#> $ E_emb_star                      <dbl> 40000, 40000, 40000, 40000, 40000, 400…
#> $ t_life_star                     <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14…
#> $ point_type                      <chr> "orig", "sweep", "sweep", "sweep", "sw…
#> $ tau_alpha_orig                  <dbl> 1.203271, 1.203271, 1.203271, 1.203271…
#> $ tau_omega_orig                  <dbl> 0.7955036, 0.7955036, 0.7955036, 0.795…
#> $ p_E                             <dbl> 0.01745357, 0.01745357, 0.01184632, 0.…
#> $ p_E_orig                        <dbl> 0.01745357, 0.01745357, 0.01184632, 0.…
#> $ eta_orig                        <dbl> 0.1974386, 0.1974386, 0.1974386, 0.197…
#> $ E_dot_s_orig                    <dbl> 73060.68, 73060.68, 73060.68, 73060.68…
#> $ C_dot_cap_orig                  <dbl> 2015.436, 2015.436, 2015.436, 2015.436…
#> $ tau_alpha_C_dot_cap_orig        <dbl> 2425.115, 2425.115, 2425.115, 2425.115…
#> $ p_s_orig                        <dbl> 0.0884, 0.0884, 0.0600, 0.0600, 0.0600…
#> $ C_dot_s_orig                    <dbl> 1275.17, 1275.17, 865.50, 865.50, 865.…
#> $ C_dot_d_orig                    <dbl> 7.142857, 7.142857, 7.142857, 7.142857…
#> $ tau_omega_C_dot_d_orig          <dbl> 5.682169, 5.682169, 5.682169, 5.682169…
#> $ C_dot_omd_orig                  <dbl> 2866.816, 2866.816, 2866.816, 2866.816…
#> $ C_dot_g_orig                    <dbl> 20834.18, 20834.18, 21243.85, 21243.85…
#> $ f_Cs_orig                       <dbl> 0.05767561, 0.05767561, 0.03914634, 0.…
#> $ e_qg_ps_UC_orig                 <dbl> -0.05508511, -0.05508511, -0.03666709,…
#> $ e_qs_ps_C_orig                  <dbl> -0.04232439, -0.04232439, -0.06085366,…
#> $ e_qg_ps_C_orig                  <dbl> 0.002590493, 0.002590493, 0.002479252,…
#> $ sigma                           <dbl> 0.04491489, 0.04491489, 0.06333291, 0.…
#> $ rho                             <dbl> -21.26433, -21.26433, -14.78958, -14.7…
#> $ E_dot_emb_orig                  <dbl> 2428.571, 2428.571, 2428.571, 2428.571…
#> $ N_dot_orig                      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ tau_alpha_star                  <dbl> 1.203271, 1.203271, 1.203271, 1.203271…
#> $ tau_omega_star                  <dbl> 0.7955036, 0.7955036, 0.7955036, 0.795…
#> $ eta_star                        <dbl> 0.3316969, 0.3316969, 0.3316969, 0.331…
#> $ eta_ratio                       <dbl> 1.68, 1.68, 1.68, 1.68, 1.68, 1.68, 1.…
#> $ S_dot_dev                       <dbl> 29572.18, 29572.18, 29572.18, 29572.18…
#> $ G_dot                           <dbl> 516.1402, 516.1402, 350.3214, 350.3214…
#> $ p_E_star                        <dbl> 0.01745357, 0.01745357, 0.01184632, 0.…
#> $ p_s_star                        <dbl> 0.05261905, 0.05261905, 0.03571429, 0.…
#> $ q_dot_s_star                    <dbl> 14425, 14425, 14425, 14425, 14425, 144…
#> $ C_dot_cap_star                  <dbl> 1965.957, 1965.957, 1965.957, 1965.957…
#> $ tau_alpha_C_dot_cap_star        <dbl> 2365.578, 2365.578, 2365.578, 2365.578…
#> $ E_dot_emb_star                  <dbl> 2857.143, 2857.143, 2857.143, 2857.143…
#> $ C_dot_s_star                    <dbl> 759.0298, 759.0298, 515.1786, 515.1786…
#> $ C_dot_d_star                    <dbl> 7.214286, 7.214286, 7.214286, 7.214286…
#> $ tau_omega_C_dot_d_star          <dbl> 5.738991, 5.738991, 5.738991, 5.738991…
#> $ C_dot_omd_star                  <dbl> 2780.407, 2780.407, 2780.407, 2780.407…
#> $ M_dot_star                      <dbl> 27401.28, 27401.28, 27401.28, 27401.28…
#> $ N_dot_star                      <dbl> 662.0857, 662.0857, 496.2669, 496.2669…
#> $ C_dot_g_star                    <dbl> 20834.18, 20834.18, 21243.85, 21243.85…
#> $ E_dot_s_star                    <dbl> 43488.5, 43488.5, 43488.5, 43488.5, 43…
#> $ f_Cs_star                       <dbl> 0.03515132, 0.03515132, 0.02367655, 0.…
#> $ e_qs_ps_UC_star                 <dbl> -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0…
#> $ e_qg_ps_UC_star                 <dbl> -0.05508511, -0.05508511, -0.03666709,…
#> $ e_qs_ps_C_star                  <dbl> -0.04232439, -0.04232439, -0.06085366,…
#> $ e_qg_ps_C_star                  <dbl> 0.002590493, 0.002590493, 0.002479252,…
#> $ t_life_hat                      <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14…
#> $ tau_alpha_hat                   <dbl> 1.203271, 1.203271, 1.203271, 1.203271…
#> $ tau_omega_hat                   <dbl> 0.7955036, 0.7955036, 0.7955036, 0.795…
#> $ eta_engr_units_hat              <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42…
#> $ eta_hat                         <dbl> 0.3316969, 0.3316969, 0.3316969, 0.331…
#> $ p_E_hat                         <dbl> 0.01745357, 0.01745357, 0.01184632, 0.…
#> $ p_s_hat                         <dbl> 0.05261905, 0.05261905, 0.03571429, 0.…
#> $ C_dot_cap_hat                   <dbl> 1965.957, 1965.957, 1965.957, 1965.957…
#> $ tau_alpha_C_dot_cap_hat         <dbl> 2365.578, 2365.578, 2365.578, 2365.578…
#> $ E_dot_emb_hat                   <dbl> 2857.143, 2857.143, 2857.143, 2857.143…
#> $ M_dot_hat                       <dbl> 27401.28, 27401.28, 27401.28, 27401.28…
#> $ q_dot_s_hat                     <dbl> 14749.25, 14749.25, 14891.54, 14891.54…
#> $ E_dot_s_hat                     <dbl> 44466.06, 44466.06, 44895.02, 44895.02…
#> $ C_dot_s_hat                     <dbl> 776.0917, 776.0917, 531.8406, 531.8406…
#> $ C_d_hat                         <dbl> 101, 101, 101, 101, 101, 101, 101, 101…
#> $ C_dot_d_hat                     <dbl> 7.214286, 7.214286, 7.214286, 7.214286…
#> $ tau_omega_C_dot_d_hat           <dbl> 5.738991, 5.738991, 5.738991, 5.738991…
#> $ C_dot_om_hat                    <dbl> 2774.668, 2774.668, 2774.668, 2774.668…
#> $ C_dot_omd_hat                   <dbl> 2780.407, 2780.407, 2780.407, 2780.407…
#> $ C_dot_g_hat                     <dbl> 20811.86, 20811.86, 21222.05, 21222.05…
#> $ f_Cs_hat                        <dbl> 0.03595023, 0.03595023, 0.02444807, 0.…
#> $ e_qs_ps_C_hat                   <dbl> -0.04330019, -0.04330019, -0.06178454,…
#> $ e_qg_ps_C_hat                   <dbl> 0.001614700, 0.001614700, 0.001548367,…
#> $ e_qs_ps_UC_hat                  <dbl> -0.07925041, -0.07925041, -0.08623261,…
#> $ e_qg_ps_UC_hat                  <dbl> -0.03433553, -0.03433553, -0.02289970,…
#> $ N_dot_hat                       <dbl> 667.3440, 667.3440, 501.4001, 501.4001…
#> $ M_dot_hat_prime                 <dbl> 21587.95, 21587.95, 21753.89, 21753.89…
#> $ t_life_bar                      <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14…
#> $ tau_alpha_bar                   <dbl> 1.203271, 1.203271, 1.203271, 1.203271…
#> $ tau_omega_bar                   <dbl> 0.7955036, 0.7955036, 0.7955036, 0.795…
#> $ eta_engr_units_bar              <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42…
#> $ eta_bar                         <dbl> 0.3316969, 0.3316969, 0.3316969, 0.331…
#> $ p_E_bar                         <dbl> 0.01745357, 0.01745357, 0.01184632, 0.…
#> $ p_s_bar                         <dbl> 0.05261905, 0.05261905, 0.03571429, 0.…
#> $ C_dot_cap_bar                   <dbl> 1965.957, 1965.957, 1965.957, 1965.957…
#> $ tau_alpha_C_dot_cap_bar         <dbl> 2365.578, 2365.578, 2365.578, 2365.578…
#> $ E_dot_emb_bar                   <dbl> 2857.143, 2857.143, 2857.143, 2857.143…
#> $ M_dot_bar                       <dbl> 27401.28, 27401.28, 27401.28, 27401.28…
#> $ q_dot_s_bar                     <dbl> 15205.19, 15205.19, 15234.77, 15234.77…
#> $ E_dot_s_bar                     <dbl> 45840.63, 45840.63, 45929.79, 45929.79…
#> $ C_dot_s_bar                     <dbl> 800.0828, 800.0828, 544.0989, 544.0989…
#> $ C_d_bar                         <dbl> 101, 101, 101, 101, 101, 101, 101, 101…
#> $ C_dot_d_bar                     <dbl> 7.214286, 7.214286, 7.214286, 7.214286…
#> $ tau_omega_C_dot_d_bar           <dbl> 5.738991, 5.738991, 5.738991, 5.738991…
#> $ C_dot_om_bar                    <dbl> 2774.668, 2774.668, 2774.668, 2774.668…
#> $ C_dot_omd_bar                   <dbl> 2780.407, 2780.407, 2780.407, 2780.407…
#> $ C_dot_g_bar                     <dbl> 21455.21, 21455.21, 21711.19, 21711.19…
#> $ f_Cs_bar                        <dbl> 0.03595023, 0.03595023, 0.02444807, 0.…
#> $ e_qs_ps_UC_bar                  <dbl> -0.07925041, -0.07925041, -0.08623261,…
#> $ e_qg_ps_UC_bar                  <dbl> -0.03433553, -0.03433553, -0.02289970,…
#> $ e_qs_ps_C_bar                   <dbl> -0.04330019, -0.04330019, -0.06178454,…
#> $ e_qg_ps_C_bar                   <dbl> 0.001614700, 0.001614700, 0.001548367,…
#> $ N_dot_bar                       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ t_life_tilde                    <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14…
#> $ tau_alpha_tilde                 <dbl> 1.203271, 1.203271, 1.203271, 1.203271…
#> $ tau_omega_tilde                 <dbl> 0.7955036, 0.7955036, 0.7955036, 0.795…
#> $ eta_engr_units_tilde            <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42…
#> $ eta_tilde                       <dbl> 0.3316969, 0.3316969, 0.3316969, 0.331…
#> $ p_E_tilde                       <dbl> 0.01745357, 0.01745357, 0.01184632, 0.…
#> $ p_s_tilde                       <dbl> 0.05261905, 0.05261905, 0.03571429, 0.…
#> $ C_dot_cap_tilde                 <dbl> 1965.957, 1965.957, 1965.957, 1965.957…
#> $ tau_alpha_C_dot_cap_tilde       <dbl> 2365.578, 2365.578, 2365.578, 2365.578…
#> $ E_dot_emb_tilde                 <dbl> 2857.143, 2857.143, 2857.143, 2857.143…
#> $ M_dot_tilde                     <dbl> 27401.28, 27401.28, 27401.28, 27401.28…
#> $ q_dot_s_tilde                   <dbl> 15205.19, 15205.19, 15234.77, 15234.77…
#> $ E_dot_s_tilde                   <dbl> 45840.63, 45840.63, 45929.79, 45929.79…
#> $ C_dot_s_tilde                   <dbl> 800.0828, 800.0828, 544.0989, 544.0989…
#> $ C_dot_om_tilde                  <dbl> 2774.668, 2774.668, 2774.668, 2774.668…
#> $ C_d_tilde                       <dbl> 101, 101, 101, 101, 101, 101, 101, 101…
#> $ C_dot_d_tilde                   <dbl> 7.214286, 7.214286, 7.214286, 7.214286…
#> $ tau_omega_C_dot_d_tilde         <dbl> 5.738991, 5.738991, 5.738991, 5.738991…
#> $ C_dot_omd_tilde                 <dbl> 2780.407, 2780.407, 2780.407, 2780.407…
#> $ C_dot_g_tilde                   <dbl> 21455.21, 21455.21, 21711.19, 21711.19…
#> $ f_Cs_tilde                      <dbl> 0.03595023, 0.03595023, 0.02444807, 0.…
#> $ e_qs_ps_UC_tilde                <dbl> -0.07925041, -0.07925041, -0.08623261,…
#> $ e_qg_ps_UC_tilde                <dbl> -0.03433553, -0.03433553, -0.02289970,…
#> $ e_qs_ps_C_tilde                 <dbl> -0.04330019, -0.04330019, -0.06178454,…
#> $ e_qg_ps_C_tilde                 <dbl> 0.001614700, 0.001614700, 0.001548367,…
#> $ N_dot_tilde                     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_M_dot_star                <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_E_star                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_t_life_star               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_star            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_star            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_engr_units_star       <dbl> 17, 17, 17, 17, 17, 17, 17, 17, 17, 17…
#> $ Delta_eta_star                  <dbl> 0.1342583, 0.1342583, 0.1342583, 0.134…
#> $ Delta_p_s_star                  <dbl> -0.03578095, -0.03578095, -0.02428571,…
#> $ Delta_E_dot_emb_star            <dbl> 428.5714, 428.5714, 428.5714, 428.5714…
#> $ Delta_C_dot_cap_star            <dbl> -49.47857, -49.47857, -49.47857, -49.4…
#> $ Delta_tau_alpha_C_dot_cap_star  <dbl> -59.53611, -59.53611, -59.53611, -59.5…
#> $ Delta_C_dot_om_star             <dbl> -86.46614, -86.46614, -86.46614, -86.4…
#> $ Delta_C_d_star                  <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,…
#> $ Delta_C_dot_d_star              <dbl> 0.07142857, 0.07142857, 0.07142857, 0.…
#> $ Delta_tau_omega_C_dot_d_star    <dbl> 0.05682169, 0.05682169, 0.05682169, 0.…
#> $ Delta_C_dot_omd_star            <dbl> -86.40931, -86.40931, -86.40931, -86.4…
#> $ Delta_E_dot_s_star              <dbl> -29572.18, -29572.18, -29572.18, -2957…
#> $ Delta_C_dot_s_star              <dbl> -516.1402, -516.1402, -350.3214, -350.…
#> $ Delta_N_dot_star                <dbl> 662.0857, 662.0857, 496.2669, 496.2669…
#> $ Delta_q_dot_s_star              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_g_star              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_M_dot_hat                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_E_hat                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_t_life_hat                <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_hat             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_hat             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_engr_units_hat        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_hat                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_s_hat                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_E_dot_emb_hat             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_cap_hat             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_C_dot_cap_hat   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_om_hat              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_d_hat                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_d_hat               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_C_dot_d_hat     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_omd_hat             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_E_dot_s_hat               <dbl> 977.5597, 977.5597, 1406.5154, 1406.51…
#> $ Delta_C_dot_s_hat               <dbl> 17.06191, 17.06191, 16.66203, 16.66203…
#> $ Delta_N_dot_hat                 <dbl> 5.258380, 5.258380, 5.133270, 5.133270…
#> $ Delta_q_dot_s_hat               <dbl> 324.2535, 324.2535, 466.5368, 466.5368…
#> $ Delta_C_dot_g_hat               <dbl> -22.32029, -22.32029, -21.79530, -21.7…
#> $ Delta_M_dot_bar                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_E_bar                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_t_life_bar                <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_bar             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_bar             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_engr_units_bar        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_bar                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_s_bar                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_E_dot_emb_bar             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_cap_bar             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_C_dot_cap_bar   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_om_bar              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_d_bar                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_d_bar               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_C_dot_d_bar     <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_omd_bar             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_E_dot_s_bar               <dbl> 1374.571, 1374.571, 1034.774, 1034.774…
#> $ Delta_C_dot_s_bar               <dbl> 23.99117, 23.99117, 12.25826, 12.25826…
#> $ Delta_N_dot_bar                 <dbl> -667.3440, -667.3440, -501.4001, -501.…
#> $ Delta_q_dot_s_bar               <dbl> 455.9408, 455.9408, 343.2314, 343.2314…
#> $ Delta_C_dot_g_bar               <dbl> 643.3529, 643.3529, 489.1419, 489.1419…
#> $ Delta_M_dot_tilde               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_E_tilde                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_t_life_tilde              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_tilde           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_tilde           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_engr_units_tilde      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_eta_tilde                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_p_s_tilde                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_E_dot_emb_tilde           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_cap_tilde           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_alpha_C_dot_cap_tilde <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_om_tilde            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_d_tilde                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_d_tilde             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_tau_omega_C_dot_d_tilde   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_omd_tilde           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_E_dot_s_tilde             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_s_tilde             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_N_dot_tilde               <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_q_dot_s_tilde             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Delta_C_dot_g_tilde             <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Re_dempl                        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,…
#> $ Re_emb                          <dbl> 0.01449239, 0.01449239, 0.01449239, 0.…
#> $ Re_cap                          <dbl> -0.09553009, -0.09553009, -0.09553009,…
#> $ Re_om                           <dbl> -0.009910093, -0.009910093, -0.0099100…
#> $ Re_d                            <dbl> 6.512471e-06, 6.512471e-06, 6.512471e-…
#> $ Re_omd                          <dbl> -0.00990358, -0.00990358, -0.00990358,…
#> $ Re_empl                         <dbl> 0.004588805, 0.004588805, 0.004588805,…
#> $ Re_isub                         <dbl> -0.002558182, -0.002558182, -0.0024980…
#> $ Re_dsub                         <dbl> 0.03305673, 0.03305673, 0.04756211, 0.…
#> $ Re_sub                          <dbl> 0.03049855, 0.03049855, 0.04506410, 0.…
#> $ Re_dinc                         <dbl> 0.04648188, 0.04648188, 0.03499148, 0.…
#> $ Re_iinc                         <dbl> 0.07373623, 0.07373623, 0.05606173, 0.…
#> $ Re_inc                          <dbl> 0.12021811, 0.12021811, 0.09105321, 0.…
#> $ Re_micro                        <dbl> 0.1553055, 0.1553055, 0.1407061, 0.140…
#> $ Re_macro                        <dbl> 0.07588324, 0.07588324, 0.00000000, 0.…
#> $ Re_dir                          <dbl> 0.07953862, 0.07953862, 0.08255359, 0.…
#> $ Re_indir                        <dbl> 0.15165009, 0.15165009, 0.05815252, 0.…
#> $ Re_tot                          <dbl> 0.2311887, 0.2311887, 0.1407061, 0.197…