Perform a complete rebound analysis

This function calls all rebound analysis functions in the correct order

Usage

rebound_analysis(.eeu_data)

Arguments

.eeu_data

Energy efficiency upgrade information in a data frame. See load_eeu_data() for an example data frame.

Value

.eeu_data with all rebound terms added as columns to the right.

Examples

complicated <- load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde() %>% 
  calc_Deltas() %>% 
  calc_rebound()
simple <- load_eeu_data() %>% 
  rebound_analysis()
all.equal(complicated, simple)
#> [1] TRUE