Information for rebound units

A list of names and constants for rebound units.

rebound_units

Format

A list with 9 entries.

energy_si

The SI energy unit ("MJ").

time_unit

The time unit ("year").

currency_unit

The default currency unit ("$").

currency_unit_latex

The currency unit in LaTeX format ("\$").

unitless

The identifier for unitless variables ("-").

unitless_latex

The identifier for unitless variables in LaTeX format ("--").

leading_delta_pattern

A regex pattern that identifies a leading Delta ("^Delta_").

surround_left

The left delimiter for units ("["]).

surround_right

The right delimiter for units ("]").

Examples

rebound_units

#> $energy_si
#> [1] "MJ"
#> 
#> $time_unit
#> [1] "year"
#> 
#> $currency_unit
#> [1] "$"
#> 
#> $currency_unit_latex
#> [1] "\\$"
#> 
#> $unitless
#> [1] "-"
#> 
#> $unitless_latex
#> [1] "--"
#> 
#> $leading_delta_pattern
#> [1] "^Delta_"
#> 
#> $surround_left
#> [1] "["
#> 
#> $surround_right
#> [1] "]"
#>