Sensitivity graph parameters

The list of graph parameters for drawing sensitivity graphs in the ReboundTools package.

sens_graph_params

Format

A string list with 59 entries.

orig_point_shape

The shape for points between rebound effects. Default is 1, an open circle..

orig_point_size

The size for original points. Default is 2.

orig_point_stroke

The size of the line surrounding points between rebound effects. Default is 0.5.

orig_point_colour

The outline colour for original points. Default is "black".

orig_point_fill

The fill colour for original points. Default is "black".

dempl_colour

The colour for direct emplacment lines.

emb_colour

The colour for embodied energy lines.

cap_colour

The colour for capital expenditure lines.

md_colour

The colour for maintenance and disposal lines.

empl_colour

The colour for emplacement lines.

dsub_colour

The colour for direct substitution lines.

isub_colour

The colour for indirect substitution lines.

sub_colour

The colour for substitution lines.

dinc_colour

The colour for direct income lines.

iinc_colour

The colour for indirect income lines.

inc_colour

The colour for income lines.

macro_colour

The colour for macro lines.

dir_colour

The colour for direct lines.

indir_colour

The colour for indirect lines.

tot_colour

The colour for total rebound lines.

dempl_size

The size for direct emplacment lines.

emb_size

The size for embodied energy lines.

cap_size

The size for capital expenditure lines.

md_size

The size for maintenance and disposal lines.

empl_size

The size for emplacement lines.

dsub_size

The size for direct substitution lines.

isub_size

The size for indirect substitution lines.

sub_size

The size for substitution lines.

dinc_size

The size for direct income lines.

iinc_size

The size for indirect income lines.

inc_size

The size for income lines.

macro_size

The size for macro lines.

dir_size

The size for direct lines.

indir_size

The size for indirect lines.

tot_size

The size for total rebound lines.

dempl_linetyps

The linetype for direct emplacment lines.

emb_linetype

The linetype for embodied energy lines.

cap_linetype

The linetype for capital expenditure lines.

md_linetype

The linetype for maintenance and disposal lines.

empl_linetype

The linetype for emplacement lines.

dsub_linetype

The linetype for direct substitution lines.

isub_linetype

The linetype for indirect substitution lines.

sub_linetype

The linetype for substitution lines.

dinc_linetype

The linetype for direct income lines.

iinc_linetype

The linetype for indirect income lines.

inc_linetype

The linetype for income lines.

macro_linetype

The linetype for macro lines.

dir_linetype

The linetype for direct lines.

indir_linetype

The linetype for indirect lines.

tot_linetype

The linetype for total rebound lines.

lineend

The line end style.

linejoin

The line join style.

include_base_condition_points

Tells whether to include base conditions points. Default is TRUE.

points_atop_paths

Tells whether to draw points above paths (TRUE) or beneath paths (FALSE). Default is TRUE.

use_latex_legend

Tells whether to convert legend text to LaTeX for a nicely-formatted legend. Default is FALSE.

include_x_axis

Tells whether to add an x-axis at y = 0. Default is FALSE.

Details

This list is passed to several graphing functions. Callers may pass a modified version of this list to change sensitivity graph appearance.

Examples

sens_graph_params
#> $orig_point_shape
#> [1] 16
#> 
#> $orig_point_size
#> [1] 2
#> 
#> $orig_point_stroke
#> [1] 0.5
#> 
#> $orig_point_colour
#> [1] "black"
#> 
#> $orig_point_fill
#> [1] "black"
#> 
#> $dempl_colour
#> [1] "#150789FF"
#> 
#> $emb_colour
#> [1] "#150789FF"
#> 
#> $cap_colour
#> [1] "#150789FF"
#> 
#> $md_colour
#> [1] "#150789FF"
#> 
#> $empl_colour
#> [1] "#150789FF"
#> 
#> $dsub_colour
#> [1] "#99149FFF"
#> 
#> $isub_colour
#> [1] "#99149FFF"
#> 
#> $sub_colour
#> [1] "#99149FFF"
#> 
#> $dinc_colour
#> [1] "#E76F5AFF"
#> 
#> $iinc_colour
#> [1] "#E76F5AFF"
#> 
#> $inc_colour
#> [1] "#E76F5AFF"
#> 
#> $micro_colour
#> [1] "#F7E225FF"
#> 
#> $macro_colour
#> [1] "#F7E225FF"
#> 
#> $dir_colour
#> [1] "black"
#> 
#> $indir_colour
#> [1] "black"
#> 
#> $tot_colour
#> [1] "black"
#> 
#> $dempl_linewidth
#> [1] 0.5
#> 
#> $emb_linewidth
#> [1] 0.5
#> 
#> $cap_linewidth
#> [1] 0.5
#> 
#> $md_linewidth
#> [1] 0.5
#> 
#> $empl_linewidth
#> [1] 0.5
#> 
#> $dsub_linewidth
#> [1] 0.5
#> 
#> $isub_linewidth
#> [1] 0.5
#> 
#> $sub_linewidth
#> [1] 0.5
#> 
#> $dinc_linewidth
#> [1] 0.5
#> 
#> $iinc_linewidth
#> [1] 0.5
#> 
#> $inc_linewidth
#> [1] 0.5
#> 
#> $micro_linewidth
#> [1] 0.5
#> 
#> $macro_linewidth
#> [1] 0.5
#> 
#> $dir_linewidth
#> [1] 0.5
#> 
#> $indir_linewidth
#> [1] 0.5
#> 
#> $tot_linewidth
#> [1] 1
#> 
#> $dempl_linetype
#> [1] "solid"
#> 
#> $emb_linetype
#> [1] "22"
#> 
#> $cap_linetype
#> [1] "22"
#> 
#> $md_linetype
#> [1] "41"
#> 
#> $empl_linetype
#> [1] "solid"
#> 
#> $dsub_linetype
#> [1] "solid"
#> 
#> $isub_linetype
#> [1] "22"
#> 
#> $sub_linetype
#> [1] "solid"
#> 
#> $dinc_linetype
#> [1] "solid"
#> 
#> $iinc_linetype
#> [1] "22"
#> 
#> $inc_linetype
#> [1] "solid"
#> 
#> $micro_linetype
#> [1] "solid"
#> 
#> $macro_linetype
#> [1] "solid"
#> 
#> $dir_linetype
#> [1] "solid"
#> 
#> $indir_linetype
#> [1] "22"
#> 
#> $tot_linetype
#> [1] "solid"
#> 
#> $lineend
#> [1] "round"
#> 
#> $linejoin
#> [1] "round"
#> 
#> $include_base_condition_points
#> [1] TRUE
#> 
#> $points_atop_paths
#> [1] TRUE
#> 
#> $include_x_axis
#> [1] FALSE
#> 
#> $use_latex_legend
#> [1] FALSE
#>