A visibility mask for the stages_table()

This data frame contains a boolean column ("Visible") that tells whether values are shown in the stages_table().

Usage

stages_table_visibility_mask

Format

A data frame with three columns, "Variable", "Stage", and "Visible".

.name

The name of the variable in the stages_table().

.stage

The name of the rebound in the stages_table().

Visible

A boolean that tells whether this combination of .name and .stage should be visible.

Examples

stages_table_visibility_mask
#> # A tibble: 105 × 3
#>    .name .stage Visible
#>    <fct> <fct>  <lgl>  
#>  1 M_dot orig   TRUE   
#>  2 M_dot star   FALSE  
#>  3 M_dot hat    FALSE  
#>  4 M_dot bar    FALSE  
#>  5 M_dot tilde  FALSE  
#>  6 p_E   orig   TRUE   
#>  7 p_E   star   FALSE  
#>  8 p_E   hat    FALSE  
#>  9 p_E   bar    FALSE  
#> 10 p_E   tilde  FALSE  
#> # ℹ 95 more rows