EEU star data

This is the list of the derived variables at the star (post-emplacement) stage of a rebound analysis.

Usage

star_vars

Format

A string list with 31 entries.

tau_alpha_star

The post-emplacement discount multiplier for beginning-of-life expenses, such as capital cost.

tau_omega_star

The post-emplacement discount multiplier for end-of-life expenses, such as disposal cost.

C_cap_star

The net capital expenditure of the upgraded device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].

C_dot_md_star

The upgraded (post-EEU) maintenance and disposal expenditure rate [$/yr].

E_emb_star

The embodied energy of the upgraded (post-EEU) device [MJ].

t_life_star

The expected lifetime of the upgraded (post-EEU) device [yr].

eta_engr_units_star

The upgraded (post-EEU) energy service efficiency. This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of eta_engr_units_orig is assumed to be the same as the denominator unit of MJ_engr_unit.

eta_star

Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], calculated by eta_tilde_engr_units / MJ_engr_unit.

eta_ratio

The energy service efficiency ratio [–], calculated by eta_star/eta_orig.

S_dot_dev

The expected device-level energy savings rate [MJ/yr], calculated by (eta_ratio - 1) * (1/eta_ratio) * E_dot_s_orig.

G_dot

The expected device-level energy gross cost savings rate [MJ/yr], calculated by p_E * S_dot_dev.

p_s_star

The upgraded (post-EEU) energy service price [$/service], calculated by p_E / eta_star = p_E / eta_tilde.

q_dot_s_star

The upgraded (post-EEU) energy service consumption rate [service/yr], same as q_dot_s_orig.

C_dot_cap_star

The capital expenditure rate of the device without discounting [$/yr], calculated by C_cap/t_life.

tau_alpha_C_dot_cap_star

The product of tau_alpha and the undiscounted capital cost rate, itself annualized and discounted in [$/yr].

C_dot_om_star

The operations and maintenance expenditure rate of the device [$/yr].

C_d_star

The disposal cost for the device [$/yr].

C_dot_d_star

The disposal cost rate of the device without discounting [$/yr].

tau_omega_C_dot_d_star

The product of tau_omega and the undiscounted disposal cost rate, itself annualized and discounted in [$/yr].

C_dot_omd_star

The operations, maintenance, and disposal expenditure rate of the device [$/yr].

E_dot_emb_star

The upgraded (post-EEU) embodied energy rate [MJ/yr], calculated by E_emb_star / t_star.

C_dot_s_star

The upgraded (post-EEU) energy expenditure rate [$/yr], calculated by p_s_star * q_dot_s_star.

M_dot_star

The disposable income rate, exclusive of taxes and savings [$/yr], exactly M_dot_orig.

N_dot_star

The freed cash rate [$/yr], calculated by G_dot - (C_dot_cap_star - C_dot_cap_orig) - (C_dot_md_star - C_dot_md_orig).

C_dot_g_star

The upgraded (post-EEU) other goods expenditure rate [$/yr], exactly C_dot_g_orig.

f_Cs_star

The upgraded (post-EEU) fraction of the energy and other budget spent on the energy service [–], calculated by C_dot_s_star / (C_dot_s_star + C_dot_g_star).

e_qs_ps_C_star

The upgraded (post-EEU) compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [–], calculated by e_qs_ps_UC_orig + f_Cs_orig*e_qs_M.

e_qg_ps_C_star

The upgraded (post-EEU) compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [–], calculated by f_Cs_orig*(f_Cs_orig + e_qs_ps_UC_orig) / (f_Cs_orig - 1).

e_qs_ps_UC_star

The upgraded (post-EEU) uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [–].

e_qg_ps_UC_star

The upgraded (post-EEU) uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of other goods ("qo") consumption (cross-price elasticity) [–].

p_E_star

The price of energy [$/MJ],

E_dot_s_star

The upgraded (post-EEU) energy consumption rate [MJ/yr], calculated by q_dot_s_star / eta_star.

Examples

star_vars
#> $tau_alpha_star
#> [1] "tau_alpha_star"
#> 
#> $tau_omega_star
#> [1] "tau_omega_star"
#> 
#> $C_cap_star
#> [1] "C_cap_star"
#> 
#> $C_dot_om_star
#> [1] "C_dot_om_star"
#> 
#> $C_d_star
#> [1] "C_d_star"
#> 
#> $E_emb_star
#> [1] "E_emb_star"
#> 
#> $t_life_star
#> [1] "t_life_star"
#> 
#> $eta_engr_units_star
#> [1] "eta_engr_units_star"
#> 
#> $eta_star
#> [1] "eta_star"
#> 
#> $eta_ratio
#> [1] "eta_ratio"
#> 
#> $S_dot_dev
#> [1] "S_dot_dev"
#> 
#> $G_dot
#> [1] "G_dot"
#> 
#> $p_s_star
#> [1] "p_s_star"
#> 
#> $q_dot_s_star
#> [1] "q_dot_s_star"
#> 
#> $C_dot_cap_star
#> [1] "C_dot_cap_star"
#> 
#> $tau_alpha_C_dot_cap_star
#> [1] "tau_alpha_C_dot_cap_star"
#> 
#> $E_dot_emb_star
#> [1] "E_dot_emb_star"
#> 
#> $C_dot_s_star
#> [1] "C_dot_s_star"
#> 
#> $M_dot_star
#> [1] "M_dot_star"
#> 
#> $N_dot_star
#> [1] "N_dot_star"
#> 
#> $C_dot_d_star
#> [1] "C_dot_d_star"
#> 
#> $tau_omega_C_dot_d_star
#> [1] "tau_omega_C_dot_d_star"
#> 
#> $C_dot_omd_star
#> [1] "C_dot_omd_star"
#> 
#> $C_dot_g_star
#> [1] "C_dot_g_star"
#> 
#> $f_Cs_star
#> [1] "f_Cs_star"
#> 
#> $e_qs_ps_C_star
#> [1] "e_qs_ps_C_star"
#> 
#> $e_qs_ps_UC_star
#> [1] "e_qs_ps_UC_star"
#> 
#> $e_qg_ps_C_star
#> [1] "e_qg_ps_C_star"
#> 
#> $e_qg_ps_UC_star
#> [1] "e_qg_ps_UC_star"
#> 
#> $p_E_star
#> [1] "p_E_star"
#> 
#> $E_dot_s_star
#> [1] "E_dot_s_star"
#>