EEU tilde data

This is the list of the derived variables at the tilde stage (after macro effect) of a rebound analysis.

Usage

tilde_vars

Format

A string list with 17 entries.

eta_tilde

Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], exactly eta_bar.

p_s_tilde

The energy service price after the macro effect [$/service], exactly p_s_bar.

C_dot_cap_tilde

The capital expenditure rate after the macro effect [$/yr], exactly C_dot_cap_bar.

C_dot_md_tilde

The maintenance and disposal expenditure rate after the macro effect [$/yr], exactly C_dot_md_bar.

E_dot_emb_tilde

The embodied energy rate after the macro effect [MJ/yr], exactly E_dot_emb_bar.

M_dot_tilde

Real income after the macro effect [MJ/yr], exactly M_dot_bar.

q_dot_s_tilde

The rate of energy service consumption after the macro effect\ [service/yr], exactly q_dot_s_bar.

C_dot_o_tilde

The rate of other goods expenditures after the macro effect [$/yr], exactly C_dot_o_bar.

f_Cs_tilde

The post-macro effect fraction of the energy and other budget spent on the energy service [--], calculated by C_dot_s_star / (C_dot_s_star + C_dot_o_star).

e_qs_ps_C_tilde

The post-macro effect compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [--], calculated by e_qs_ps_UC_orig + f_Cs_orig*e_qs_M.

e_qo_ps_C_tilde

The post-macro effect compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [--], calculated by f_Cs_orig*(f_Cs_orig + e_qs_ps_UC_orig) / (f_Cs_orig - 1).

e_qs_ps_UC_tilde

The post-macro effect uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of energy service ("qs") consumption (own-price elasticity) [--].

e_qo_ps_UC_tilde

The post-macro effect uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of other goods ("qo") consumption (cross-price elasticity) [--].

N_dot_tilde

The freed cash rate after the macro effect [$/yr], exactly 0.

[service/yr]: R:service/yr%5C

Examples

tilde_vars
#> $eta_engr_units_tilde
#> [1] "eta_engr_units_tilde"
#> 
#> $eta_tilde
#> [1] "eta_tilde"
#> 
#> $p_s_tilde
#> [1] "p_s_tilde"
#> 
#> $C_dot_cap_tilde
#> [1] "C_dot_cap_tilde"
#> 
#> $C_dot_md_tilde
#> [1] "C_dot_md_tilde"
#> 
#> $E_dot_emb_tilde
#> [1] "E_dot_emb_tilde"
#> 
#> $M_dot_tilde
#> [1] "M_dot_tilde"
#> 
#> $q_dot_s_tilde
#> [1] "q_dot_s_tilde"
#> 
#> $E_dot_s_tilde
#> [1] "E_dot_s_tilde"
#> 
#> $C_dot_s_tilde
#> [1] "C_dot_s_tilde"
#> 
#> $C_dot_o_tilde
#> [1] "C_dot_o_tilde"
#> 
#> $f_Cs_tilde
#> [1] "f_Cs_tilde"
#> 
#> $e_qs_ps_C_tilde
#> [1] "e_qs_ps_C_tilde"
#> 
#> $e_qs_ps_UC_tilde
#> [1] "e_qs_ps_UC_tilde"
#> 
#> $e_qo_ps_C_tilde
#> [1] "e_qo_ps_C_tilde"
#> 
#> $e_qo_ps_UC_tilde
#> [1] "e_qo_ps_UC_tilde"
#> 
#> $N_dot_tilde
#> [1] "N_dot_tilde"
#>