Difference variables

This is the list of difference variables of the form Delta_var_stage, where var is the variable name (from ReboundTools::key_analysis_vars) and stage is the rebound stage being considered (from ReboundTools::rebound_stages).

Delta_vars

Format

A string list with 48 entries.

Delta_eta_star

Calculated as eta_star - eta_orig.

Delta_p_s_star

Calculated as p_s_star - p_s_orig.

.

Etc.

.

Etc.

.

Etc.

Delta_M_dot_tilde

Calculated as M_dot_tilde - M_dot_bar.

Delta_N_dot_tilde

Calculated as N_dot_tilde - N_dot_bar.

Examples

Delta_vars

#> $Delta_eta_engr_units_star
#> [1] "Delta_eta_engr_units_star"
#> 
#> $Delta_eta_star
#> [1] "Delta_eta_star"
#> 
#> $Delta_p_s_star
#> [1] "Delta_p_s_star"
#> 
#> $Delta_q_dot_s_star
#> [1] "Delta_q_dot_s_star"
#> 
#> $Delta_E_dot_s_star
#> [1] "Delta_E_dot_s_star"
#> 
#> $Delta_E_dot_emb_star
#> [1] "Delta_E_dot_emb_star"
#> 
#> $Delta_C_dot_s_star
#> [1] "Delta_C_dot_s_star"
#> 
#> $Delta_C_dot_cap_star
#> [1] "Delta_C_dot_cap_star"
#> 
#> $Delta_C_dot_md_star
#> [1] "Delta_C_dot_md_star"
#> 
#> $Delta_C_dot_o_star
#> [1] "Delta_C_dot_o_star"
#> 
#> $Delta_N_dot_star
#> [1] "Delta_N_dot_star"
#> 
#> $Delta_M_dot_star
#> [1] "Delta_M_dot_star"
#> 
#> $Delta_eta_engr_units_hat
#> [1] "Delta_eta_engr_units_hat"
#> 
#> $Delta_eta_hat
#> [1] "Delta_eta_hat"
#> 
#> $Delta_p_s_hat
#> [1] "Delta_p_s_hat"
#> 
#> $Delta_q_dot_s_hat
#> [1] "Delta_q_dot_s_hat"
#> 
#> $Delta_E_dot_s_hat
#> [1] "Delta_E_dot_s_hat"
#> 
#> $Delta_E_dot_emb_hat
#> [1] "Delta_E_dot_emb_hat"
#> 
#> $Delta_C_dot_s_hat
#> [1] "Delta_C_dot_s_hat"
#> 
#> $Delta_C_dot_cap_hat
#> [1] "Delta_C_dot_cap_hat"
#> 
#> $Delta_C_dot_md_hat
#> [1] "Delta_C_dot_md_hat"
#> 
#> $Delta_C_dot_o_hat
#> [1] "Delta_C_dot_o_hat"
#> 
#> $Delta_N_dot_hat
#> [1] "Delta_N_dot_hat"
#> 
#> $Delta_M_dot_hat
#> [1] "Delta_M_dot_hat"
#> 
#> $Delta_eta_engr_units_bar
#> [1] "Delta_eta_engr_units_bar"
#> 
#> $Delta_eta_bar
#> [1] "Delta_eta_bar"
#> 
#> $Delta_p_s_bar
#> [1] "Delta_p_s_bar"
#> 
#> $Delta_q_dot_s_bar
#> [1] "Delta_q_dot_s_bar"
#> 
#> $Delta_E_dot_s_bar
#> [1] "Delta_E_dot_s_bar"
#> 
#> $Delta_E_dot_emb_bar
#> [1] "Delta_E_dot_emb_bar"
#> 
#> $Delta_C_dot_s_bar
#> [1] "Delta_C_dot_s_bar"
#> 
#> $Delta_C_dot_cap_bar
#> [1] "Delta_C_dot_cap_bar"
#> 
#> $Delta_C_dot_md_bar
#> [1] "Delta_C_dot_md_bar"
#> 
#> $Delta_C_dot_o_bar
#> [1] "Delta_C_dot_o_bar"
#> 
#> $Delta_N_dot_bar
#> [1] "Delta_N_dot_bar"
#> 
#> $Delta_M_dot_bar
#> [1] "Delta_M_dot_bar"
#> 
#> $Delta_eta_engr_units_tilde
#> [1] "Delta_eta_engr_units_tilde"
#> 
#> $Delta_eta_tilde
#> [1] "Delta_eta_tilde"
#> 
#> $Delta_p_s_tilde
#> [1] "Delta_p_s_tilde"
#> 
#> $Delta_q_dot_s_tilde
#> [1] "Delta_q_dot_s_tilde"
#> 
#> $Delta_E_dot_s_tilde
#> [1] "Delta_E_dot_s_tilde"
#> 
#> $Delta_E_dot_emb_tilde
#> [1] "Delta_E_dot_emb_tilde"
#> 
#> $Delta_C_dot_s_tilde
#> [1] "Delta_C_dot_s_tilde"
#> 
#> $Delta_C_dot_cap_tilde
#> [1] "Delta_C_dot_cap_tilde"
#> 
#> $Delta_C_dot_md_tilde
#> [1] "Delta_C_dot_md_tilde"
#> 
#> $Delta_C_dot_o_tilde
#> [1] "Delta_C_dot_o_tilde"
#> 
#> $Delta_N_dot_tilde
#> [1] "Delta_N_dot_tilde"
#> 
#> $Delta_M_dot_tilde
#> [1] "Delta_M_dot_tilde"
#>