Calculate differences between stages

Calculate differences between stages

Usage

calc_Deltas(
  .tilde_data = NULL,
  key_analysis_vars = ReboundTools::key_analysis_vars,
  rebound_stages = ReboundTools::rebound_stages,
  .all_vars = "all_vars"
)

Arguments

.tilde_data

A data frame containing rebound calculations, original data, star data, hat data, and tilde data, likely calculated by calc_tilde().

key_analysis_vars

See ReboundTools::key_analysis_vars.

rebound_stages

See ReboundTools::rebound_stages.

.all_vars

A column names used internally for all variables. Default is ".all_vars".

Value

.tilde_data with additional columns containing all possible difference terms.

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde() %>% 
  calc_Deltas()
#> # A tibble: 3 × 211
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 204 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>, R_alpha_orig <dbl>,
#> #   R_omega_orig <dbl>, p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, …