Calculate differences between stages

Calculate differences between stages

calc_Deltas(
  .tilde_data = NULL,
  key_analysis_vars = ReboundTools::key_analysis_vars,
  rebound_stages = ReboundTools::rebound_stages
)

Arguments

.tilde_data
A data frame containing rebound calculations, original data, star data, hat data, and tilde data, likely calculated by calc_tilde().

key_analysis_vars
See ReboundTools::key_analysis_vars.

rebound_stages
See ReboundTools::rebound_stages.

Value

.tilde_data with additional columns containing all possible difference terms.

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde() %>% 
  calc_Deltas()

#> # A tibble: 2 x 139
#>   Reference Case  Original     Upgrade            service_unit energy_engr_unit
#>   <chr>     <chr> <chr>        <chr>              <chr>        <chr>           
#> 1 None yet  Car   Ford Fusion  Ford Fusion Hybrid mile         gal             
#> 2 None yet  Lamp  Incandescent LED                lm-hr        W-hr            
#> # … with 133 more variables: MJ/energy_engr_unit <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>,
#> #   e_qs_ps <dbl>, e_qo_ps <dbl>, sigma <dbl>, rho <dbl>, E_dot_emb_orig <dbl>,
#> #   N_dot_orig <dbl>, eta_star <dbl>, eta_ratio <dbl>, S_dot_dev <dbl>,
#> #   G_dot <dbl>, p_s_star <dbl>, q_dot_s_star <dbl>, C_dot_cap_star <dbl>,
#> #   E_dot_emb_star <dbl>, C_dot_s_star <dbl>, M_dot_star <dbl>,
#> #   N_dot_star <dbl>, C_dot_o_star <dbl>, E_dot_s_star <dbl>,
#> #   eta_engr_units_hat <dbl>, eta_hat <dbl>, p_s_hat <dbl>,
#> #   C_dot_cap_hat <dbl>, C_dot_md_hat <dbl>, E_dot_emb_hat <dbl>,
#> #   M_dot_hat <dbl>, q_dot_s_hat <dbl>, E_dot_s_hat <dbl>, C_dot_s_hat <dbl>,
#> #   C_dot_o_hat <dbl>, N_dot_hat <dbl>, M_dot_hat_prime <dbl>,
#> #   eta_engr_units_bar <dbl>, eta_bar <dbl>, p_s_bar <dbl>,
#> #   C_dot_cap_bar <dbl>, C_dot_md_bar <dbl>, E_dot_emb_bar <dbl>,
#> #   M_dot_bar <dbl>, q_dot_s_bar <dbl>, E_dot_s_bar <dbl>, C_dot_s_bar <dbl>,
#> #   C_dot_o_bar <dbl>, N_dot_bar <dbl>, eta_engr_units_tilde <dbl>,
#> #   eta_tilde <dbl>, p_s_tilde <dbl>, C_dot_cap_tilde <dbl>,
#> #   C_dot_md_tilde <dbl>, E_dot_emb_tilde <dbl>, M_dot_tilde <dbl>,
#> #   q_dot_s_tilde <dbl>, E_dot_s_tilde <dbl>, C_dot_s_tilde <dbl>,
#> #   C_dot_o_tilde <dbl>, N_dot_tilde <dbl>, Delta_eta_engr_units_star <dbl>,
#> #   Delta_eta_star <dbl>, Delta_p_s_star <dbl>, Delta_q_dot_s_star <dbl>,
#> #   Delta_E_dot_s_star <dbl>, Delta_E_dot_emb_star <dbl>,
#> #   Delta_C_dot_s_star <dbl>, Delta_C_dot_cap_star <dbl>,
#> #   Delta_C_dot_md_star <dbl>, Delta_C_dot_o_star <dbl>,
#> #   Delta_N_dot_star <dbl>, Delta_M_dot_star <dbl>,
#> #   Delta_eta_engr_units_hat <dbl>, Delta_eta_hat <dbl>, Delta_p_s_hat <dbl>, …