Calculate energy rebound data at the bar stage

This function calculates energy rebound information for the bar stage (immediately after the income effect).

calc_bar(
  .hat_data = NULL,
  tol = 1e-06,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  p_E = ReboundTools::orig_vars$p_E,
  eta_engr_units_hat = ReboundTools::hat_vars$eta_engr_units_hat,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  e_qs_ps_UC_hat = ReboundTools::hat_vars$e_qs_ps_UC_hat,
  e_qo_ps_UC_hat = ReboundTools::hat_vars$e_qo_ps_UC_hat,
  e_qs_ps_C_hat = ReboundTools::hat_vars$e_qs_ps_C_hat,
  e_qo_ps_C_hat = ReboundTools::hat_vars$e_qo_ps_C_hat,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  eta_engr_units_bar = ReboundTools::bar_vars$eta_engr_units_bar,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  C_dot_md_bar = ReboundTools::bar_vars$C_dot_md_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  f_Cs_bar = ReboundTools::bar_vars$f_Cs_bar,
  e_qs_ps_UC_bar = ReboundTools::bar_vars$e_qs_ps_UC_bar,
  e_qo_ps_UC_bar = ReboundTools::bar_vars$e_qo_ps_UC_bar,
  e_qs_ps_C_bar = ReboundTools::bar_vars$e_qs_ps_C_bar,
  e_qo_ps_C_bar = ReboundTools::bar_vars$e_qo_ps_C_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar
)

Arguments

.hat_data

An optional data frame containing rebound calculations, original data, star data, and hat data, likely calculated by calc_hat().

tol

The tolerance with which the budget constraint should be satisfied. Default is 1e-6.

e_qs_M, e_qo_M, p_E

See ReboundTools::eeu_base_params.

eta_engr_units_hat, eta_hat, p_s_hat, C_dot_cap_hat, C_dot_md_hat, E_dot_emb_hat, M_dot_hat, q_dot_s_hat, N_dot_hat, M_dot_hat_prime, C_dot_o_hat, e_qs_ps_UC_hat, e_qo_ps_UC_hat, e_qs_ps_C_hat, e_qo_ps_C_hat, E_dot_s_hat

See ReboundTools::hat_vars.

eta_engr_units_bar, eta_bar, p_s_bar, C_dot_cap_bar, C_dot_md_bar, E_dot_emb_bar, M_dot_bar, q_dot_s_bar, E_dot_s_bar, C_dot_s_bar, C_dot_o_bar, f_Cs_bar, e_qs_ps_UC_bar, e_qo_ps_UC_bar, e_qs_ps_C_bar, e_qo_ps_C_bar, N_dot_bar

See ReboundTools::bar_vars.

Value

A list or data frame of derived rebound values for the bar stage (after the income effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar()
#> # A tibble: 2 × 95
#>   Reference Case  Original   Upgrade servi…¹ energ…² MJ/en…³   I_E     k p_E_e…⁴
#>   <chr>     <chr> <chr>      <chr>   <chr>   <chr>     <dbl> <dbl> <dbl>   <dbl>
#> 1 None yet  Car   Ford Fusi… Ford F… mile    gal     1.27e+2  3.39     1 2.21e+0
#> 2 None yet  Lamp  Incandesc… LED     lm-hr   W-hr    3.6 e-3  3.39     1 1.35e-4
#> # … with 85 more variables: e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>, …