Calculate energy rebound data at the hat stage

This function calculates energy rebound information for the hat stage (immediately after the substitution effect).

Usage

calc_hat(
  .star_data = NULL,
  use_sub_approx = FALSE,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  p_E = ReboundTools::orig_vars$p_E,
  e_qo_ps_C = ReboundTools::orig_vars$e_qo_ps_C,
  e_qs_ps_C = ReboundTools::orig_vars$e_qs_ps_C,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  sigma = ReboundTools::orig_vars$sigma,
  rho = ReboundTools::orig_vars$rho,
  eta_engr_units_star = ReboundTools::star_vars$eta_engr_units_star,
  eta_star = ReboundTools::star_vars$eta_star,
  p_s_star = ReboundTools::star_vars$p_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  M_dot_star = ReboundTools::star_vars$M_dot_star,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  e_qs_ps_UC_star = ReboundTools::star_vars$e_qs_ps_UC_star,
  e_qo_ps_UC_star = ReboundTools::star_vars$e_qo_ps_UC_star,
  e_qs_ps_C_star = ReboundTools::star_vars$e_qs_ps_C_star,
  e_qo_ps_C_star = ReboundTools::star_vars$e_qo_ps_C_star,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star,
  G_dot = ReboundTools::star_vars$G_dot,
  eta_engr_units_hat = ReboundTools::hat_vars$eta_engr_units_hat,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  C_dot_s_hat = ReboundTools::hat_vars$C_dot_s_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  f_Cs_hat = ReboundTools::hat_vars$f_Cs_hat,
  e_qs_ps_UC_hat = ReboundTools::hat_vars$e_qs_ps_UC_hat,
  e_qo_ps_UC_hat = ReboundTools::hat_vars$e_qo_ps_UC_hat,
  e_qs_ps_C_hat = ReboundTools::hat_vars$e_qs_ps_C_hat,
  e_qo_ps_C_hat = ReboundTools::hat_vars$e_qo_ps_C_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime
)

Arguments

.star_data

An optional data frame containing EEU base data, original data, and star data, likely calculated by calc_star().

use_sub_approx

Tells whether to use an approximate method for calculating the substitution effect. Default is FALSE. See details.

e_qs_M, e_qo_M

See ReboundTools::eeu_base_params.

p_E

See ReboundTools::eeu_base_params.

e_qo_ps_C, e_qs_ps_C, C_dot_cap_orig, C_dot_md_orig, f_Cs_orig, q_dot_s_orig, C_dot_o_orig, sigma, rho

See ReboundTools::orig_vars.

eta_engr_units_star, eta_star, p_s_star, C_dot_cap_star, C_dot_md_star, E_dot_emb_star, M_dot_star, q_dot_s_star, eta_ratio, C_dot_o_star, e_qs_ps_UC_star, e_qo_ps_UC_star, e_qs_ps_C_star, e_qo_ps_C_star, N_dot_star, E_dot_s_star, G_dot

See ReboundTools::star_vars.

eta_engr_units_hat, eta_hat, p_s_hat, C_dot_cap_hat, C_dot_md_hat, E_dot_emb_hat, M_dot_hat, q_dot_s_hat, E_dot_s_hat, C_dot_s_hat, C_dot_o_hat, f_Cs_hat, e_qs_ps_UC_hat, e_qo_ps_UC_hat, e_qs_ps_C_hat, e_qo_ps_C_hat, N_dot_hat, M_dot_hat_prime

See ReboundTools::hat_vars.

Value

A list or data frame of derived rebound values for the hat stage (after the substitution effect).

Details

By default, this function uses an exact method for calculating the substitution effect. The exact method uses a constant elasticity of substitution (CES) formulation of utility, which assumes constant elasticity of substitution (sigma) along the indifference curve.

The approximate method is employed with use_sub_approx = TRUE. The approximate method assumes constant price elasticity of energy service consumption. (In fact, the price elasticity of energy service consumption is not constant along an indifference curve.) The approximation leads to a simple and elegant rebound expression but a different (and incorrect) consumption bundle after the substitution effect.

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat()
#> # A tibble: 2 × 78
#>   Reference Case  Original   Upgrade servi…¹ energ…² MJ/en…³   I_E     k p_E_e…⁴
#>   <chr>     <chr> <chr>      <chr>   <chr>   <chr>     <dbl> <dbl> <dbl>   <dbl>
#> 1 None yet  Car   Ford Fusi… Ford F… mile    gal     1.27e+2  3.39     1 2.21e+0
#> 2 None yet  Lamp  Incandesc… LED     lm-hr   W-hr    3.6 e-3  3.39     1 1.35e-4
#> # … with 68 more variables: e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>, …