Calculate energy rebound data at the hat stage

This function calculates energy rebound information for the hat stage (immediately after the substitution effect).

Usage

calc_hat(
  .star_data = NULL,
  t_life_star = ReboundTools::star_vars$t_life_star,
  tau_alpha_star = ReboundTools::star_vars$tau_alpha_star,
  tau_omega_star = ReboundTools::star_vars$tau_omega_star,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qg_M = ReboundTools::eeu_base_params$e_qg_M,
  p_E = ReboundTools::orig_vars$p_E,
  e_qo_ps_C = ReboundTools::orig_vars$e_qo_ps_C,
  e_qs_ps_C = ReboundTools::orig_vars$e_qs_ps_C,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  sigma = ReboundTools::orig_vars$sigma,
  rho = ReboundTools::orig_vars$rho,
  eta_engr_units_star = ReboundTools::star_vars$eta_engr_units_star,
  eta_star = ReboundTools::star_vars$eta_star,
  p_s_star = ReboundTools::star_vars$p_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  C_dot_om_star = ReboundTools::star_vars$C_dot_om_star,
  C_d_star = ReboundTools::star_vars$C_d_star,
  C_dot_d_star = ReboundTools::star_vars$C_dot_d_star,
  C_dot_omd_star = ReboundTools::star_vars$C_dot_omd_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  M_dot_star = ReboundTools::star_vars$M_dot_star,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  C_dot_g_star = ReboundTools::star_vars$C_dot_g_star,
  e_qs_ps_UC_star = ReboundTools::star_vars$e_qs_ps_UC_star,
  e_qg_ps_UC_star = ReboundTools::star_vars$e_qg_ps_UC_star,
  e_qs_ps_C_star = ReboundTools::star_vars$e_qs_ps_C_star,
  e_qg_ps_C_star = ReboundTools::star_vars$e_qg_ps_C_star,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star,
  G_dot = ReboundTools::star_vars$G_dot,
  t_life_hat = ReboundTools::hat_vars$t_life_hat,
  tau_alpha_hat = ReboundTools::hat_vars$tau_alpha_hat,
  tau_omega_hat = ReboundTools::hat_vars$tau_omega_hat,
  eta_engr_units_hat = ReboundTools::hat_vars$eta_engr_units_hat,
  eta_hat = ReboundTools::hat_vars$eta_hat,
  p_s_hat = ReboundTools::hat_vars$p_s_hat,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  tau_alpha_C_dot_cap_hat = ReboundTools::hat_vars$tau_alpha_C_dot_cap_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  M_dot_hat = ReboundTools::hat_vars$M_dot_hat,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  p_E_hat = ReboundTools::hat_vars$p_E_hat,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  C_dot_s_hat = ReboundTools::hat_vars$C_dot_s_hat,
  C_d_hat = ReboundTools::hat_vars$C_d_hat,
  C_dot_d_hat = ReboundTools::hat_vars$C_dot_d_hat,
  tau_omega_C_dot_d_hat = ReboundTools::hat_vars$tau_omega_C_dot_d_hat,
  C_dot_om_hat = ReboundTools::hat_vars$C_dot_om_hat,
  C_dot_omd_hat = ReboundTools::hat_vars$C_dot_omd_hat,
  C_dot_g_hat = ReboundTools::hat_vars$C_dot_g_hat,
  f_Cs_hat = ReboundTools::hat_vars$f_Cs_hat,
  e_qs_ps_UC_hat = ReboundTools::hat_vars$e_qs_ps_UC_hat,
  e_qg_ps_UC_hat = ReboundTools::hat_vars$e_qg_ps_UC_hat,
  e_qs_ps_C_hat = ReboundTools::hat_vars$e_qs_ps_C_hat,
  e_qg_ps_C_hat = ReboundTools::hat_vars$e_qg_ps_C_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime
)

Arguments

.star_data

An optional data frame containing EEU base data, original data, and star data, likely calculated by calc_star().

t_life_star, C_dot_om_star, C_d_star, C_dot_d_star, C_dot_omd_star, tau_alpha_star, tau_omega_star, eta_engr_units_star, eta_star, p_s_star, C_dot_cap_star, E_dot_emb_star, M_dot_star, q_dot_s_star, eta_ratio, C_dot_g_star, e_qs_ps_UC_star, e_qg_ps_UC_star, e_qs_ps_C_star, e_qg_ps_C_star, N_dot_star, E_dot_s_star, G_dot

See ReboundTools::star_vars.

e_qs_M, e_qg_M

See ReboundTools::eeu_base_params.

p_E

See ReboundTools::eeu_base_params.

e_qo_ps_C, e_qs_ps_C, C_dot_cap_orig, f_Cs_orig, q_dot_s_orig, C_dot_g_orig, sigma, rho

See ReboundTools::orig_vars.

t_life_hat, C_d_hat, C_dot_d_hat, tau_omega_C_dot_d_hat, C_dot_om_hat, C_dot_omd_hat, tau_alpha_hat, tau_omega_hat, eta_engr_units_hat, eta_hat, p_s_hat, C_dot_cap_hat, tau_alpha_C_dot_cap_hat, E_dot_emb_hat, M_dot_hat, q_dot_s_hat, E_dot_s_hat, C_dot_s_hat, C_dot_g_hat, f_Cs_hat, e_qs_ps_UC_hat, e_qg_ps_UC_hat, e_qs_ps_C_hat, e_qg_ps_C_hat, N_dot_hat, M_dot_hat_prime, p_E_hat

See ReboundTools::hat_vars.

Value

A list or data frame of derived rebound values for the hat stage (after the substitution effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat()
#> # A tibble: 3 × 102
#>   Reference Case       Original     Upgrade      r service_unit energy_engr_unit
#>   <chr>     <chr>      <chr>        <chr>    <dbl> <chr>        <chr>           
#> 1 None yet  Car        Ford Fusion  Ford Fu…  0.03 mile         gal             
#> 2 None yet  Lamp       Incandescent LED       0.03 lm-hr        W-hr            
#> 3 None yet  Car, r = 0 Ford Fusion  Ford Fu…  0    mile         gal             
#> # ℹ 95 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC_orig <dbl>, e_qs_M <dbl>, e_qg_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, C_cap_star <dbl>, C_dot_om_orig <dbl>,
#> #   C_d_orig <dbl>, C_dot_om_star <dbl>, C_d_star <dbl>, E_emb_orig <dbl>,
#> #   t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   tau_alpha_orig <dbl>, tau_omega_orig <dbl>, p_E <dbl>, p_E_orig <dbl>, …