Calculate energy rebound data at the original stage

This function calculates energy rebound information for the original stage (pre-EEU).

calc_orig(
  .eeu_data = NULL,
  MJ_engr_unit = ReboundTools::eeu_base_params$MJ_engr_unit,
  p_E_engr_units = ReboundTools::eeu_base_params$p_E_engr_units,
  e_qs_ps_UC = ReboundTools::eeu_base_params$e_qs_ps_UC,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  eta_engr_units_orig = ReboundTools::orig_vars$eta_engr_units_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_cap_orig = ReboundTools::orig_vars$C_cap_orig,
  t_own_orig = ReboundTools::orig_vars$t_own_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  E_emb_orig = ReboundTools::orig_vars$E_emb_orig,
  t_life_orig = ReboundTools::orig_vars$t_life_orig,
  p_E = ReboundTools::orig_vars$p_E,
  eta_orig = ReboundTools::orig_vars$eta_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  p_s_orig = ReboundTools::orig_vars$p_s_orig,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  e_qs_ps = ReboundTools::orig_vars$e_qs_ps,
  e_qo_ps = ReboundTools::orig_vars$e_qo_ps,
  sigma = ReboundTools::orig_vars$sigma,
  rho = ReboundTools::orig_vars$rho,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  N_dot_orig = ReboundTools::orig_vars$N_dot_orig
)

Arguments

.eeu_data
An optional data frame containing EEU base data. See ReboundTools::eeu_base_params.

MJ_engr_unit, p_E_engr_units, e_qs_ps_UC, e_qs_M
See ReboundTools::eeu_base_params.

eta_engr_units_orig, q_dot_s_orig, C_cap_orig, t_own_orig, M_dot_orig, C_dot_md_orig, E_emb_orig, t_life_orig, p_E, eta_orig, E_dot_s_orig, C_dot_cap_orig, p_s_orig, C_dot_s_orig, C_dot_o_orig, f_Cs_orig, e_qs_ps, e_qo_ps, sigma, rho, E_dot_emb_orig, N_dot_orig
See ReboundTools::orig_vars.

Value

A list or data frame of derived rebound values.

Examples

load_eeu_data() %>% 
  calc_orig()

#> # A tibble: 2 x 41
#>   Reference Case  Original     Upgrade            service_unit energy_engr_unit
#>   <chr>     <chr> <chr>        <chr>              <chr>        <chr>           
#> 1 None yet  Car   Ford Fusion  Ford Fusion Hybrid mile         gal             
#> 2 None yet  Lamp  Incandescent LED                lm-hr        W-hr            
#> # … with 35 more variables: MJ/energy_engr_unit <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>,
#> #   e_qs_ps <dbl>, e_qo_ps <dbl>, sigma <dbl>, rho <dbl>, E_dot_emb_orig <dbl>,
#> #   N_dot_orig <dbl>