Calculate rebound terms.

This function calculates rebound terms from a data frame that already contains Delta terms. Note that each rebound term is calculated twice as a way of validating the derived expression for rebound.

calc_rebound(
  .Deltas_data = NULL,
  tol = 1e-06,
  I_E = ReboundTools::eeu_base_params$I_E,
  e_qs_M = ReboundTools::eeu_base_params$e_qs_M,
  e_qo_M = ReboundTools::eeu_base_params$e_qo_M,
  k = ReboundTools::eeu_base_params$k,
  e_qs_ps_C = ReboundTools::orig_vars$e_qs_ps_C,
  e_qo_ps_C = ReboundTools::orig_vars$e_qo_ps_C,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  M_dot_hat_prime = ReboundTools::hat_vars$M_dot_hat_prime,
  Delta_E_dot_emb_star = ReboundTools::Delta_vars$Delta_E_dot_emb_star,
  Delta_C_dot_cap_star = ReboundTools::Delta_vars$Delta_C_dot_cap_star,
  Delta_C_dot_md_star = ReboundTools::Delta_vars$Delta_C_dot_md_star,
  Delta_E_dot_s_hat = ReboundTools::Delta_vars$Delta_E_dot_s_hat,
  Delta_C_dot_o_hat = ReboundTools::Delta_vars$Delta_C_dot_o_hat,
  Delta_E_dot_s_bar = ReboundTools::Delta_vars$Delta_E_dot_s_bar,
  Delta_C_dot_o_bar = ReboundTools::Delta_vars$Delta_C_dot_o_bar,
  Re_dempl = ReboundTools::rebound_terms$Re_dempl,
  Re_emb = ReboundTools::rebound_terms$Re_emb,
  Re_cap = ReboundTools::rebound_terms$Re_cap,
  Re_md = ReboundTools::rebound_terms$Re_md,
  Re_empl = ReboundTools::rebound_terms$Re_empl,
  Re_dsub = ReboundTools::rebound_terms$Re_dsub,
  Re_isub = ReboundTools::rebound_terms$Re_isub,
  Re_sub = ReboundTools::rebound_terms$Re_sub,
  Re_dinc = ReboundTools::rebound_terms$Re_dinc,
  Re_iinc = ReboundTools::rebound_terms$Re_iinc,
  Re_inc = ReboundTools::rebound_terms$Re_inc,
  Re_macro = ReboundTools::rebound_terms$Re_macro,
  Re_dir = ReboundTools::rebound_terms$Re_dir,
  Re_indir = ReboundTools::rebound_terms$Re_indir,
  Re_tot = ReboundTools::rebound_terms$Re_tot
)

Arguments

.Deltas_data
A data frame containing Delta values, likely created by ReboundTools::calc_Deltas()

tol
The tolerance for checking internal consistency of rebound calculations. Default is 1e-6.

I_E, e_qs_M, e_qo_M, k
See ReboundTools::eeu_base_params.

e_qs_ps_C, e_qo_ps_C, C_dot_o_orig, E_dot_s_orig
See ReboundTools::orig_vars.

S_dot_dev, eta_ratio
See ReboundTools::star_vars.

N_dot_hat, M_dot_hat_prime
See ReboundTools::hat_vars.

Delta_E_dot_emb_star, Delta_C_dot_cap_star, Delta_C_dot_md_star, Delta_E_dot_s_hat, Delta_C_dot_o_hat, Delta_E_dot_s_bar, Delta_C_dot_o_bar
See ReboundTools::Delta_vars.

Re_dempl, Re_emb, Re_cap, Re_md, Re_empl, Re_dsub, Re_isub, Re_sub, Re_dinc, Re_iinc, Re_inc, Re_macro, Re_dir, Re_indir, Re_tot
See ReboundTools::rebound_terms.

Value

A data frame with rebound terms added as columns.

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde() %>% 
  calc_Deltas() %>% 
  calc_rebound()

#> # A tibble: 2 x 154
#>   Reference Case  Original     Upgrade            service_unit energy_engr_unit
#>   <chr>     <chr> <chr>        <chr>              <chr>        <chr>           
#> 1 None yet  Car   Ford Fusion  Ford Fusion Hybrid mile         gal             
#> 2 None yet  Lamp  Incandescent LED                lm-hr        W-hr            
#> # … with 148 more variables: MJ/energy_engr_unit <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>,
#> #   e_qs_ps_C <dbl>, e_qo_ps_C <dbl>, sigma <dbl>, rho <dbl>,
#> #   E_dot_emb_orig <dbl>, N_dot_orig <dbl>, eta_star <dbl>, eta_ratio <dbl>,
#> #   S_dot_dev <dbl>, G_dot <dbl>, p_s_star <dbl>, q_dot_s_star <dbl>,
#> #   C_dot_cap_star <dbl>, E_dot_emb_star <dbl>, C_dot_s_star <dbl>,
#> #   M_dot_star <dbl>, N_dot_star <dbl>, C_dot_o_star <dbl>, E_dot_s_star <dbl>,
#> #   eta_engr_units_hat <dbl>, eta_hat <dbl>, p_s_hat <dbl>,
#> #   C_dot_cap_hat <dbl>, C_dot_md_hat <dbl>, E_dot_emb_hat <dbl>,
#> #   M_dot_hat <dbl>, q_dot_s_hat <dbl>, E_dot_s_hat <dbl>, C_dot_s_hat <dbl>,
#> #   C_dot_o_hat <dbl>, N_dot_hat <dbl>, M_dot_hat_prime <dbl>,
#> #   eta_engr_units_bar <dbl>, eta_bar <dbl>, p_s_bar <dbl>,
#> #   C_dot_cap_bar <dbl>, C_dot_md_bar <dbl>, E_dot_emb_bar <dbl>,
#> #   M_dot_bar <dbl>, q_dot_s_bar <dbl>, E_dot_s_bar <dbl>, C_dot_s_bar <dbl>,
#> #   C_dot_o_bar <dbl>, N_dot_bar <dbl>, eta_engr_units_tilde <dbl>,
#> #   eta_tilde <dbl>, p_s_tilde <dbl>, C_dot_cap_tilde <dbl>,
#> #   C_dot_md_tilde <dbl>, E_dot_emb_tilde <dbl>, M_dot_tilde <dbl>,
#> #   q_dot_s_tilde <dbl>, E_dot_s_tilde <dbl>, C_dot_s_tilde <dbl>,
#> #   C_dot_o_tilde <dbl>, N_dot_tilde <dbl>, Delta_eta_engr_units_star <dbl>,
#> #   Delta_eta_star <dbl>, Delta_p_s_star <dbl>, Delta_q_dot_s_star <dbl>,
#> #   Delta_E_dot_s_star <dbl>, Delta_E_dot_emb_star <dbl>,
#> #   Delta_C_dot_s_star <dbl>, Delta_C_dot_cap_star <dbl>,
#> #   Delta_C_dot_md_star <dbl>, Delta_C_dot_o_star <dbl>,
#> #   Delta_N_dot_star <dbl>, Delta_M_dot_star <dbl>,
#> #   Delta_eta_engr_units_hat <dbl>, Delta_eta_hat <dbl>, Delta_p_s_hat <dbl>, …