Calculate energy rebound data at the star stage

This function calculates energy rebound information for the star stage (immediately after the emplacement effect, with no behavior change).

calc_star(
  .orig_data = NULL,
  MJ_engr_unit = ReboundTools::eeu_base_params$MJ_engr_unit,
  p_E = ReboundTools::orig_vars$p_E,
  eta_orig = ReboundTools::orig_vars$eta_orig,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  M_dot_orig = ReboundTools::orig_vars$M_dot_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  eta_star_engr_units = ReboundTools::star_vars$eta_star_engr_units,
  E_emb_star = ReboundTools::star_vars$E_emb_star,
  t_life_star = ReboundTools::star_vars$t_life_star,
  C_cap_star = ReboundTools::star_vars$C_cap_star,
  t_own_star = ReboundTools::star_vars$t_own_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  eta_star = ReboundTools::star_vars$eta_star,
  eta_ratio = ReboundTools::star_vars$eta_ratio,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  G_dot = ReboundTools::star_vars$G_dot,
  p_s_star = ReboundTools::star_vars$p_s_star,
  q_dot_s_star = ReboundTools::star_vars$q_dot_s_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  C_dot_s_star = ReboundTools::star_vars$C_dot_s_star,
  M_dot_star = ReboundTools::star_vars$M_dot_star,
  N_dot_star = ReboundTools::star_vars$N_dot_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star
)

Arguments

.orig_data
An optional data frame containing EEU base data and original data, likely calculated by calc_orig().

MJ_engr_unit, p_E
See ReboundTools::eeu_base_params.

eta_orig, E_dot_s_orig, q_dot_s_orig, M_dot_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig
See ReboundTools::orig_vars.

eta_star_engr_units, E_emb_star, t_life_star, C_cap_star, t_own_star, C_dot_md_star, eta_star, eta_ratio, S_dot_dev, G_dot, p_s_star, q_dot_s_star, C_dot_cap_star, E_dot_emb_star, C_dot_s_star, M_dot_star, N_dot_star, C_dot_o_star, E_dot_s_star
See ReboundTools::star_vars.

Value

A list or data frame of derived rebound values for the star stage (after the emplacement effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star()

#> # A tibble: 2 x 53
#>   Reference Case  Original Upgrade service_unit energy_engr_unit
#>   <chr>     <chr> <chr>    <chr>   <chr>        <chr>           
#> 1 None yet  Car   Ford Fu… Ford F… mile         gal             
#> 2 None yet  Lamp  Incande… LED     lm-hr        kW-hr           
#> # … with 47 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>,
#> #   e_qs_ps <dbl>, e_qo_ps <dbl>, sigma <dbl>, E_dot_emb_orig <dbl>,
#> #   N_dot_orig <dbl>, eta_star <dbl>, eta_ratio <dbl>, S_dot_dev <dbl>,
#> #   G_dot <dbl>, p_s_star <dbl>, q_dot_s_star <dbl>, C_dot_cap_star <dbl>,
#> #   E_dot_emb_star <dbl>, C_dot_s_star <dbl>, M_dot_star <dbl>,
#> #   N_dot_star <dbl>, C_dot_o_star <dbl>, E_dot_s_star <dbl>