Calculate energy rebound data at the tilde stage

This function calculates energy rebound information for the tilde stage (immediately after the productivity effect).

calc_tilde(
  .bar_data = NULL,
  eta_engr_units_bar = ReboundTools::bar_vars$eta_engr_units_bar,
  eta_bar = ReboundTools::bar_vars$eta_bar,
  p_s_bar = ReboundTools::bar_vars$p_s_bar,
  C_dot_cap_bar = ReboundTools::bar_vars$C_dot_cap_bar,
  C_dot_md_bar = ReboundTools::bar_vars$C_dot_md_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  M_dot_bar = ReboundTools::bar_vars$M_dot_bar,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  C_dot_s_bar = ReboundTools::bar_vars$C_dot_s_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  N_dot_bar = ReboundTools::bar_vars$N_dot_bar,
  eta_engr_units_tilde = ReboundTools::tilde_vars$eta_engr_units_tilde,
  eta_tilde = ReboundTools::tilde_vars$eta_tilde,
  p_s_tilde = ReboundTools::tilde_vars$p_s_tilde,
  C_dot_cap_tilde = ReboundTools::tilde_vars$C_dot_cap_tilde,
  C_dot_md_tilde = ReboundTools::tilde_vars$C_dot_md_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  M_dot_tilde = ReboundTools::tilde_vars$M_dot_tilde,
  q_dot_s_tilde = ReboundTools::tilde_vars$q_dot_s_tilde,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  C_dot_s_tilde = ReboundTools::tilde_vars$C_dot_s_tilde,
  C_dot_o_tilde = ReboundTools::tilde_vars$C_dot_o_tilde,
  N_dot_tilde = ReboundTools::tilde_vars$N_dot_tilde
)

Arguments

.bar_data
An optional data frame containing rebound calculations, original data, star data, hat data, and bar data, likely calculated by calc_bar().

eta_engr_units_bar, eta_bar, p_s_bar, C_dot_cap_bar, C_dot_md_bar, E_dot_emb_bar, M_dot_bar, q_dot_s_bar, E_dot_s_bar, C_dot_s_bar, C_dot_o_bar, N_dot_bar
See ReboundTools::bar_vars.

eta_engr_units_tilde, eta_tilde, p_s_tilde, C_dot_cap_tilde, C_dot_md_tilde, E_dot_emb_tilde, M_dot_tilde, q_dot_s_tilde, E_dot_s_tilde, C_dot_s_tilde, C_dot_o_tilde, N_dot_tilde
See ReboundTools::tilde_vars.

Value

A list or data frame of derived rebound values for the bar stage (after the income effect).

Examples

load_eeu_data() %>% 
  calc_orig() %>% 
  calc_star() %>% 
  calc_hat() %>% 
  calc_bar() %>% 
  calc_tilde()

#> # A tibble: 2 x 90
#>   Reference Case  Original Upgrade service_unit energy_engr_unit
#>   <chr>     <chr> <chr>    <chr>   <chr>        <chr>           
#> 1 None yet  Car   Ford Fu… Ford F… mile         gal             
#> 2 None yet  Lamp  Incande… LED     lm-hr        kW-hr           
#> # … with 84 more variables: `MJ/energy_engr_unit` <dbl>, I_E <dbl>, k <dbl>,
#> #   p_E_engr_units <dbl>, e_qs_ps_UC <dbl>, e_qs_M <dbl>, e_qo_M <dbl>,
#> #   eta_engr_units_orig <dbl>, eta_engr_units_star <dbl>, q_dot_s_orig <dbl>,
#> #   M_dot_orig <dbl>, C_cap_orig <dbl>, t_own_orig <dbl>, C_cap_star <dbl>,
#> #   t_own_star <dbl>, C_dot_md_orig <dbl>, C_dot_md_star <dbl>,
#> #   E_emb_orig <dbl>, t_life_orig <dbl>, E_emb_star <dbl>, t_life_star <dbl>,
#> #   p_E <dbl>, eta_orig <dbl>, E_dot_s_orig <dbl>, C_dot_cap_orig <dbl>,
#> #   p_s_orig <dbl>, C_dot_s_orig <dbl>, C_dot_o_orig <dbl>, f_Cs_orig <dbl>,
#> #   e_qs_ps <dbl>, e_qo_ps <dbl>, sigma <dbl>, E_dot_emb_orig <dbl>,
#> #   N_dot_orig <dbl>, eta_star <dbl>, eta_ratio <dbl>, S_dot_dev <dbl>,
#> #   G_dot <dbl>, p_s_star <dbl>, q_dot_s_star <dbl>, C_dot_cap_star <dbl>,
#> #   E_dot_emb_star <dbl>, C_dot_s_star <dbl>, M_dot_star <dbl>,
#> #   N_dot_star <dbl>, C_dot_o_star <dbl>, E_dot_s_star <dbl>,
#> #   eta_engr_units_hat <dbl>, eta_hat <dbl>, p_s_hat <dbl>,
#> #   C_dot_cap_hat <dbl>, C_dot_md_hat <dbl>, E_dot_emb_hat <dbl>,
#> #   M_dot_hat <dbl>, q_dot_s_hat <dbl>, E_dot_s_hat <dbl>, C_dot_s_hat <dbl>,
#> #   C_dot_o_hat <dbl>, N_dot_hat <dbl>, M_dot_hat_prime <dbl>,
#> #   eta_engr_units_bar <dbl>, eta_bar <dbl>, p_s_bar <dbl>,
#> #   C_dot_cap_bar <dbl>, C_dot_md_bar <dbl>, E_dot_emb_bar <dbl>,
#> #   M_dot_bar <dbl>, q_dot_s_bar <dbl>, E_dot_s_bar <dbl>, C_dot_s_bar <dbl>,
#> #   C_dot_o_bar <dbl>, N_dot_bar <dbl>, eta_engr_units_tilde <dbl>,
#> #   eta_tilde <dbl>, p_s_tilde <dbl>, C_dot_cap_tilde <dbl>,
#> #   C_dot_md_tilde <dbl>, E_dot_emb_tilde <dbl>, M_dot_tilde <dbl>,
#> #   q_dot_s_tilde <dbl>, E_dot_s_tilde <dbl>, C_dot_s_tilde <dbl>,
#> #   C_dot_o_tilde <dbl>, N_dot_tilde <dbl>