A data frame of expenditure path segments

Makes a data frame of segments for an expenditure path graph. Each stage of the rebound process is represented in the data frame.

Usage

expenditure_paths(
  .rebound_data,
  indexed = FALSE,
  graph_params = ReboundTools::path_graph_params,
  rebound_segments = ReboundTools::rebound_segments,
  graph_type = ReboundTools::graph_types$expenditure,
  tau_alpha_orig = ReboundTools::orig_vars$tau_alpha_orig,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_omd_orig = ReboundTools::orig_vars$C_dot_omd_orig,
  C_dot_g_orig = ReboundTools::orig_vars$C_dot_g_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  tau_alpha_star = ReboundTools::star_vars$tau_alpha_star,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  Delta_C_dot_omd_star = ReboundTools::Delta_vars$Delta_C_dot_omd_star,
  Delta_C_dot_s_hat = ReboundTools::Delta_vars$Delta_C_dot_s_hat,
  Delta_C_dot_g_hat = ReboundTools::Delta_vars$Delta_C_dot_g_hat,
  Delta_C_dot_s_bar = ReboundTools::Delta_vars$Delta_C_dot_s_bar,
  Delta_C_dot_g_bar = ReboundTools::Delta_vars$Delta_C_dot_g_bar,
  graph_df_colnames = ReboundTools::graph_df_colnames
)

Arguments

.rebound_data

A data frame of rebound analysis results, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_params

See ReboundTools::graph_params.

rebound_segments

See ReboundTools::rebound_segments.

graph_type

See ReboundTools::graph_types.

tau_alpha_orig, C_dot_s_orig, C_dot_cap_orig, C_dot_omd_orig, C_dot_g_orig

See ReboundTools::orig_vars.

G_dot

See ReboundTools::star_vars.

tau_alpha_star, C_dot_cap_star, Delta_C_dot_omd_star, Delta_C_dot_s_hat, Delta_C_dot_g_hat, Delta_C_dot_s_bar, Delta_C_dot_g_bar

See ReboundTools::Delta_vars.

graph_df_colnames

See ReboundTools::graph_df_colnames.

Value

A data frame with cost rebound path segments.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  expenditure_paths()
#> # A tibble: 21 × 14
#>    Reference Case       Original   Upgrade graph_type line_name colour linewidth
#>    <chr>     <chr>      <chr>      <chr>   <chr>      <chr>     <chr>      <dbl>
#>  1 None yet  Car        Ford Fusi… Ford F… Expenditu… dempl     #1507…       1  
#>  2 None yet  Lamp       Incandesc… LED     Expenditu… dempl     #1507…       1  
#>  3 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… dempl     #1507…       1  
#>  4 None yet  Car        Ford Fusi… Ford F… Expenditu… cap       #1507…       1.5
#>  5 None yet  Lamp       Incandesc… LED     Expenditu… cap       #1507…       1.5
#>  6 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… cap       #1507…       1.5
#>  7 None yet  Car        Ford Fusi… Ford F… Expenditu… OMd       #1507…       1  
#>  8 None yet  Lamp       Incandesc… LED     Expenditu… OMd       #1507…       1  
#>  9 None yet  Car, r = 0 Ford Fusi… Ford F… Expenditu… OMd       #1507…       1  
#> 10 None yet  Car        Ford Fusi… Ford F… Expenditu… isub      #9914…       1  
#> # ℹ 11 more rows
#> # ℹ 6 more variables: linetype <chr>, x <dbl>, y <dbl>, xend <dbl>, yend <dbl>,
#> #   end_arrow <lgl>