Extract points from a data frame of paths

Make a data frame of points between energy rebound effects for any type of rebound path graph. Each available stage of the rebound process is represented in the data frame.

extract_points(
  .paths,
  graph_params = ReboundTools::path_graph_params,
  rebound_stages = ReboundTools::rebound_stages,
  rebound_segments = ReboundTools::rebound_segments,
  graph_df_colnames = ReboundTools::graph_df_colnames
)

Arguments

.paths
A data frame of rebound paths, likely created by energy_paths(), expenditure_paths(), or consumption_paths().

graph_params
See ReboundTools::graph_params.

rebound_stages
See ReboundTools::rebound_stages.

rebound_segments
See ReboundTools::rebound_segments.

graph_df_colnames
See ReboundTools::graph_df_colnames.

Value

A data frame containing points extracted from paths.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  energy_paths() %>% 
  extract_points()

#> # A tibble: 18 x 14
#>    Reference Case  Original   Upgrade     graph_type colour   size      x      y
#>    <chr>     <chr> <chr>      <chr>       <chr>      <chr>   <dbl>  <dbl>  <dbl>
#>  1 None yet  Car   Ford Fusi… Ford Fusio… Energy     #F7E22…     1 4.59e4 82080.
#>  2 None yet  Lamp  Incandesc… LED         Energy     #F7E22…     1 6.22e1 92905.
#>  3 None yet  Car   Ford Fusi… Ford Fusio… Energy     #15078…     1 7.31e4 77317.
#>  4 None yet  Lamp  Incandesc… LED         Energy     #15078…     1 2.37e2 92840.
#>  5 None yet  Car   Ford Fusi… Ford Fusio… Energy     #15078…     1 4.35e4 77317.
#>  6 None yet  Lamp  Incandesc… LED         Energy     #15078…     1 2.55e1 92840.
#>  7 None yet  Car   Ford Fusi… Ford Fusio… Energy     #15078…     1 4.35e4 77745.
#>  8 None yet  Lamp  Incandesc… LED         Energy     #15078…     1 2.55e1 92839.
#>  9 None yet  Car   Ford Fusi… Ford Fusio… Energy     #99149…     1 4.35e4 77452.
#> 10 None yet  Lamp  Incandesc… LED         Energy     #99149…     1 2.55e1 92839.
#> 11 None yet  Car   Ford Fusi… Ford Fusio… Energy     #99149…     1 4.35e4 77385.
#> 12 None yet  Lamp  Incandesc… LED         Energy     #99149…     1 2.55e1 92824.
#> 13 None yet  Car   Ford Fusi… Ford Fusio… Energy     #E76F5…     1 4.44e4 77385.
#> 14 None yet  Lamp  Incandesc… LED         Energy     #E76F5…     1 6.22e1 92824.
#> 15 None yet  Car   Ford Fusi… Ford Fusio… Energy     #E76F5…     1 4.59e4 77385.
#> 16 None yet  Lamp  Incandesc… LED         Energy     #E76F5…     1 6.22e1 92824.
#> 17 None yet  Car   Ford Fusi… Ford Fusio… Energy     #F7E22…     1 4.59e4 79686.
#> 18 None yet  Lamp  Incandesc… LED         Energy     #F7E22…     1 6.22e1 92865.
#> # … with 5 more variables: point_name <chr>, start_point <lgl>, shape <dbl>,
#> #   stroke <dbl>, fill <chr>

load_eeu_data() %>% 
  rebound_analysis() %>% 
  expenditure_paths() %>% 
  extract_points()

#> # A tibble: 16 x 14
#>    Reference Case  Original   Upgrade     graph_type colour  size       x      y
#>    <chr>     <chr> <chr>      <chr>       <chr>      <chr>  <dbl>   <dbl>  <dbl>
#>  1 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #E76F…     1 8.02e+2 26600.
#>  2 None yet  Lamp  Incandesc… LED         Expenditu… #E76F…     1 2.34e+0 27399.
#>  3 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #1507…     1 1.28e+3 26126.
#>  4 None yet  Lamp  Incandesc… LED         Expenditu… #1507…     1 8.90e+0 27392.
#>  5 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #1507…     1 7.59e+2 26126.
#>  6 None yet  Lamp  Incandesc… LED         Expenditu… #1507…     1 9.61e-1 27392.
#>  7 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #1507…     1 7.59e+2 26027.
#>  8 None yet  Lamp  Incandesc… LED         Expenditu… #1507…     1 9.61e-1 27391.
#>  9 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #9914…     1 7.59e+2 25941.
#> 10 None yet  Lamp  Incandesc… LED         Expenditu… #9914…     1 9.61e-1 27391.
#> 11 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #9914…     1 7.59e+2 25921.
#> 12 None yet  Lamp  Incandesc… LED         Expenditu… #9914…     1 9.61e-1 27387.
#> 13 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #E76F…     1 7.74e+2 25921.
#> 14 None yet  Lamp  Incandesc… LED         Expenditu… #E76F…     1 2.34e+0 27387.
#> 15 None yet  Car   Ford Fusi… Ford Fusio… Expenditu… #E76F…     1 8.02e+2 25921.
#> 16 None yet  Lamp  Incandesc… LED         Expenditu… #E76F…     1 2.34e+0 27387.
#> # … with 5 more variables: point_name <chr>, start_point <lgl>, shape <dbl>,
#> #   stroke <dbl>, fill <chr>

load_eeu_data() %>% 
  rebound_analysis() %>% 
  consumption_paths() %>% 
  extract_points()

#> # A tibble: 10 x 14
#>    Reference Case  Original   Upgrade       graph_type colour   size     x     y
#>    <chr>     <chr> <chr>      <chr>         <chr>      <chr>   <dbl> <dbl> <dbl>
#>  1 None yet  Car   Ford Fusi… Ford Fusion … Consumpti… #E76F5…     1  1.06 1.03 
#>  2 None yet  Lamp  Incandesc… LED           Consumpti… #E76F5…     1  2.44 1.00 
#>  3 None yet  Car   Ford Fusi… Ford Fusion … Consumpti… #99149…     1  1    1    
#>  4 None yet  Lamp  Incandesc… LED           Consumpti… #99149…     1  1    1    
#>  5 None yet  Car   Ford Fusi… Ford Fusion … Consumpti… #99149…     1  1    0.999
#>  6 None yet  Lamp  Incandesc… LED           Consumpti… #99149…     1  1    1.00 
#>  7 None yet  Car   Ford Fusi… Ford Fusion … Consumpti… #E76F5…     1  1.02 0.999
#>  8 None yet  Lamp  Incandesc… LED           Consumpti… #E76F5…     1  2.43 1.00 
#>  9 None yet  Car   Ford Fusi… Ford Fusion … Consumpti… #E76F5…     1  1.06 0.999
#> 10 None yet  Lamp  Incandesc… LED           Consumpti… #E76F5…     1  2.44 1.00 
#> # … with 5 more variables: point_name <chr>, start_point <lgl>, shape <dbl>,
#> #   stroke <dbl>, fill <chr>