Extract points from a data frame of paths

Make a data frame of points between energy rebound effects for any type of rebound path graph. Each available stage of the rebound process is represented in the data frame.

extract_points(
  .paths,
  graph_params = ReboundTools::path_graph_params,
  rebound_stages = ReboundTools::rebound_stages,
  rebound_segments = ReboundTools::rebound_segments,
  graph_df_colnames = ReboundTools::graph_df_colnames
)

Arguments

.paths

A data frame of rebound paths, likely created by energy_paths(), expenditure_paths(), or consumption_paths().

graph_params

See ReboundTools::graph_params.

rebound_stages

See ReboundTools::rebound_stages.

rebound_segments

See ReboundTools::rebound_segments.

graph_df_colnames

See ReboundTools::graph_df_colnames.

Value

A data frame containing points extracted from paths.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  energy_paths() %>% 
  extract_points()
#> # A tibble: 18 × 15
#>    Reference Case  Original Upgrade graph…¹ colour linew…²      x      y point…³
#>    <chr>     <chr> <chr>    <chr>   <chr>   <chr>    <dbl>  <dbl>  <dbl> <chr>  
#>  1 None yet  Car   Ford Fu… Ford F… Energy  #F7E2…     1   4.59e4 82080. .last_…
#>  2 None yet  Lamp  Incande… LED     Energy  #F7E2…     1   6.22e1 92905. .last_…
#>  3 None yet  Car   Ford Fu… Ford F… Energy  #1507…     1   7.31e4 77317. orig   
#>  4 None yet  Lamp  Incande… LED     Energy  #1507…     1   2.37e2 92840. orig   
#>  5 None yet  Car   Ford Fu… Ford F… Energy  #1507…     1.5 4.35e4 77317. NA     
#>  6 None yet  Lamp  Incande… LED     Energy  #1507…     1.5 2.55e1 92840. NA     
#>  7 None yet  Car   Ford Fu… Ford F… Energy  #1507…     1   4.35e4 77745. NA     
#>  8 None yet  Lamp  Incande… LED     Energy  #1507…     1   2.55e1 92839. NA     
#>  9 None yet  Car   Ford Fu… Ford F… Energy  #9914…     1   4.35e4 77452. star   
#> 10 None yet  Lamp  Incande… LED     Energy  #9914…     1   2.55e1 92839. star   
#> 11 None yet  Car   Ford Fu… Ford F… Energy  #9914…     1   4.35e4 77385. NA     
#> 12 None yet  Lamp  Incande… LED     Energy  #9914…     1   2.55e1 92824. NA     
#> 13 None yet  Car   Ford Fu… Ford F… Energy  #E76F…     1   4.44e4 77385. hat    
#> 14 None yet  Lamp  Incande… LED     Energy  #E76F…     1   6.22e1 92824. hat    
#> 15 None yet  Car   Ford Fu… Ford F… Energy  #E76F…     1   4.59e4 77385. NA     
#> 16 None yet  Lamp  Incande… LED     Energy  #E76F…     1   6.22e1 92824. NA     
#> 17 None yet  Car   Ford Fu… Ford F… Energy  #F7E2…     1   4.59e4 79686. bar    
#> 18 None yet  Lamp  Incande… LED     Energy  #F7E2…     1   6.22e1 92865. bar    
#> # … with 5 more variables: start_point <lgl>, shape <dbl>, size <dbl>,
#> #   stroke <dbl>, fill <chr>, and abbreviated variable names ¹​graph_type,
#> #   ²​linewidth, ³​point_name
load_eeu_data() %>% 
  rebound_analysis() %>% 
  expenditure_paths() %>% 
  extract_points()
#> # A tibble: 16 × 15
#>    Reference Case  Origi…¹ Upgrade graph…² colour linew…³       x      y point…⁴
#>    <chr>     <chr> <chr>   <chr>   <chr>   <chr>    <dbl>   <dbl>  <dbl> <chr>  
#>  1 None yet  Car   Ford F… Ford F… Expend… #E76F…     1   8.02e+2 26600. .last_…
#>  2 None yet  Lamp  Incand… LED     Expend… #E76F…     1   2.34e+0 27399. .last_…
#>  3 None yet  Car   Ford F… Ford F… Expend… #1507…     1   1.28e+3 26126. orig   
#>  4 None yet  Lamp  Incand… LED     Expend… #1507…     1   8.90e+0 27392. orig   
#>  5 None yet  Car   Ford F… Ford F… Expend… #1507…     1.5 7.59e+2 26126. NA     
#>  6 None yet  Lamp  Incand… LED     Expend… #1507…     1.5 9.61e-1 27392. NA     
#>  7 None yet  Car   Ford F… Ford F… Expend… #1507…     1   7.59e+2 26027. NA     
#>  8 None yet  Lamp  Incand… LED     Expend… #1507…     1   9.61e-1 27391. NA     
#>  9 None yet  Car   Ford F… Ford F… Expend… #9914…     1   7.59e+2 25941. star   
#> 10 None yet  Lamp  Incand… LED     Expend… #9914…     1   9.61e-1 27391. star   
#> 11 None yet  Car   Ford F… Ford F… Expend… #9914…     1   7.59e+2 25921. NA     
#> 12 None yet  Lamp  Incand… LED     Expend… #9914…     1   9.61e-1 27387. NA     
#> 13 None yet  Car   Ford F… Ford F… Expend… #E76F…     1   7.74e+2 25921. hat    
#> 14 None yet  Lamp  Incand… LED     Expend… #E76F…     1   2.34e+0 27387. hat    
#> 15 None yet  Car   Ford F… Ford F… Expend… #E76F…     1   8.02e+2 25921. NA     
#> 16 None yet  Lamp  Incand… LED     Expend… #E76F…     1   2.34e+0 27387. NA     
#> # … with 5 more variables: start_point <lgl>, shape <dbl>, size <dbl>,
#> #   stroke <dbl>, fill <chr>, and abbreviated variable names ¹​Original,
#> #   ²​graph_type, ³​linewidth, ⁴​point_name
load_eeu_data() %>% 
  rebound_analysis() %>% 
  consumption_paths() %>% 
  extract_points()
#> # A tibble: 10 × 15
#>    Reference Case  Original   Upgrade graph…¹ colour linew…²     x     y point…³
#>    <chr>     <chr> <chr>      <chr>   <chr>   <chr>    <dbl> <dbl> <dbl> <chr>  
#>  1 None yet  Car   Ford Fusi… Ford F… Consum… #E76F…       1  1.06 1.03  .last_…
#>  2 None yet  Lamp  Incandesc… LED     Consum… #E76F…       1  2.44 1.00  .last_…
#>  3 None yet  Car   Ford Fusi… Ford F… Consum… #9914…       1  1    1     star   
#>  4 None yet  Lamp  Incandesc… LED     Consum… #9914…       1  1    1     star   
#>  5 None yet  Car   Ford Fusi… Ford F… Consum… #9914…       1  1    0.999 NA     
#>  6 None yet  Lamp  Incandesc… LED     Consum… #9914…       1  1    1.00  NA     
#>  7 None yet  Car   Ford Fusi… Ford F… Consum… #E76F…       1  1.02 0.999 hat    
#>  8 None yet  Lamp  Incandesc… LED     Consum… #E76F…       1  2.43 1.00  hat    
#>  9 None yet  Car   Ford Fusi… Ford F… Consum… #E76F…       1  1.06 0.999 NA     
#> 10 None yet  Lamp  Incandesc… LED     Consum… #E76F…       1  2.44 1.00  NA     
#> # … with 5 more variables: start_point <lgl>, shape <dbl>, size <dbl>,
#> #   stroke <dbl>, fill <chr>, and abbreviated variable names ¹​graph_type,
#> #   ²​linewidth, ³​point_name