Extract points from a data frame of paths

Make a data frame of points between energy rebound effects for any type of rebound path graph. Each available stage of the rebound process is represented in the data frame.

Usage

extract_points(
  .paths,
  graph_params = ReboundTools::path_graph_params,
  rebound_stages = ReboundTools::rebound_stages,
  rebound_segments = ReboundTools::rebound_segments,
  graph_df_colnames = ReboundTools::graph_df_colnames
)

Arguments

.paths

A data frame of rebound paths, likely created by energy_paths(), expenditure_paths(), or consumption_paths().

graph_params

See ReboundTools::graph_params.

rebound_stages

See ReboundTools::rebound_stages.

rebound_segments

See ReboundTools::rebound_segments.

graph_df_colnames

See ReboundTools::graph_df_colnames.

Value

A data frame containing points extracted from paths.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  energy_paths() %>% 
  extract_points()
#> # A tibble: 27 × 15
#>    Reference Case     Original Upgrade graph_type colour linewidth      x      y
#>    <chr>     <chr>    <chr>    <chr>   <chr>      <chr>      <dbl>  <dbl>  <dbl>
#>  1 None yet  Car      Ford Fu… Ford F… Energy     #F7E2…       1   4.58e4 87262.
#>  2 None yet  Lamp     Incande… LED     Energy     #F7E2…       1   6.22e1 92905.
#>  3 None yet  Car, r … Ford Fu… Ford F… Energy     #F7E2…       1   4.58e4 88584.
#>  4 None yet  Car      Ford Fu… Ford F… Energy     #1507…       1   7.31e4 82759.
#>  5 None yet  Lamp     Incande… LED     Energy     #1507…       1   2.37e2 92840.
#>  6 None yet  Car, r … Ford Fu… Ford F… Energy     #1507…       1   7.31e4 84148.
#>  7 None yet  Car      Ford Fu… Ford F… Energy     #1507…       1.5 4.35e4 82759.
#>  8 None yet  Lamp     Incande… LED     Energy     #1507…       1.5 2.55e1 92840.
#>  9 None yet  Car, r … Ford Fu… Ford F… Energy     #1507…       1.5 4.35e4 84148.
#> 10 None yet  Car      Ford Fu… Ford F… Energy     #1507…       1   4.35e4 83188.
#> # ℹ 17 more rows
#> # ℹ 6 more variables: point_name <chr>, start_point <lgl>, shape <dbl>,
#> #   size <dbl>, stroke <dbl>, fill <chr>
load_eeu_data() %>% 
  rebound_analysis() %>% 
  expenditure_paths() %>% 
  extract_points()
#> # A tibble: 24 × 15
#>    Reference Case    Original Upgrade graph_type colour linewidth       x      y
#>    <chr>     <chr>   <chr>    <chr>   <chr>      <chr>      <dbl>   <dbl>  <dbl>
#>  1 None yet  Car     Ford Fu… Ford F… Expenditu… #E76F…       1   8.00e+2 26601.
#>  2 None yet  Lamp    Incande… LED     Expenditu… #E76F…       1   2.34e+0 27399.
#>  3 None yet  Car, r… Ford Fu… Ford F… Expenditu… #E76F…       1   8.00e+2 26602.
#>  4 None yet  Car     Ford Fu… Ford F… Expenditu… #1507…       1   1.28e+3 26126.
#>  5 None yet  Lamp    Incande… LED     Expenditu… #1507…       1   8.90e+0 27392.
#>  6 None yet  Car, r… Ford Fu… Ford F… Expenditu… #1507…       1   1.28e+3 26126.
#>  7 None yet  Car     Ford Fu… Ford F… Expenditu… #1507…       1.5 7.59e+2 26126.
#>  8 None yet  Lamp    Incande… LED     Expenditu… #1507…       1.5 9.61e-1 27392.
#>  9 None yet  Car, r… Ford Fu… Ford F… Expenditu… #1507…       1.5 7.59e+2 26126.
#> 10 None yet  Car     Ford Fu… Ford F… Expenditu… #1507…       1   7.59e+2 26067.
#> # ℹ 14 more rows
#> # ℹ 6 more variables: point_name <chr>, start_point <lgl>, shape <dbl>,
#> #   size <dbl>, stroke <dbl>, fill <chr>
load_eeu_data() %>% 
  rebound_analysis() %>% 
  consumption_paths() %>% 
  extract_points()
#> # A tibble: 15 × 15
#>    Reference Case       Original Upgrade graph_type colour linewidth     x     y
#>    <chr>     <chr>      <chr>    <chr>   <chr>      <chr>      <dbl> <dbl> <dbl>
#>  1 None yet  Car        Ford Fu… Ford F… Consumpti… #E76F…         1  1.05 1.03 
#>  2 None yet  Lamp       Incande… LED     Consumpti… #E76F…         1  2.44 1.00 
#>  3 None yet  Car, r = 0 Ford Fu… Ford F… Consumpti… #E76F…         1  1.05 1.03 
#>  4 None yet  Car        Ford Fu… Ford F… Consumpti… #9914…         1  1    1    
#>  5 None yet  Lamp       Incande… LED     Consumpti… #9914…         1  1    1    
#>  6 None yet  Car, r = 0 Ford Fu… Ford F… Consumpti… #9914…         1  1    1    
#>  7 None yet  Car        Ford Fu… Ford F… Consumpti… #9914…         1  1    0.999
#>  8 None yet  Lamp       Incande… LED     Consumpti… #9914…         1  1    1.00 
#>  9 None yet  Car, r = 0 Ford Fu… Ford F… Consumpti… #9914…         1  1    0.999
#> 10 None yet  Car        Ford Fu… Ford F… Consumpti… #E76F…         1  1.02 0.999
#> 11 None yet  Lamp       Incande… LED     Consumpti… #E76F…         1  2.43 1.00 
#> 12 None yet  Car, r = 0 Ford Fu… Ford F… Consumpti… #E76F…         1  1.02 0.999
#> 13 None yet  Car        Ford Fu… Ford F… Consumpti… #E76F…         1  1.05 0.999
#> 14 None yet  Lamp       Incande… LED     Consumpti… #E76F…         1  2.44 1.00 
#> 15 None yet  Car, r = 0 Ford Fu… Ford F… Consumpti… #E76F…         1  1.05 0.999
#> # ℹ 6 more variables: point_name <chr>, start_point <lgl>, shape <dbl>,
#> #   size <dbl>, stroke <dbl>, fill <chr>