Names of graph data frame columns

The list of names of graph data frame columns. These are default names for columns produced internally.

Usage

graph_df_colnames

Format

A string list with 26 entries.

colour_col

The name of the column containing line colours.

size_col

The name of the column containing line sizes (widths).

linetype_col

The name of the column containing line types.

graph_type_col

The name of the column containing graph types.

line_name_col

The name of the column containing names (string identifiers) for lines, segments, and grids.

slope_col

The name of the column containing line slopes.

intercept_col

The name of the column containing line intercepts.

x_col

The name of the column containing starting x values.

y_col

The name of the column containing starting y values.

xend_col

The name of the column containing ending x values.

yend_col

The name of the column containing ending y values.

qs1_qs0_col

The name of the column containing a q_s/q_s_0 point on this indifference curve.

Co1_Co0_col

The name of the column containing a C_s/C_s_0 point on this indifference curve.

f_Cs_orig_col

The name of the column containing the original value of f_Cs for this indifference curve.

sigma_col

The name of the column containing ending sigma values for this indifference curve.

start_point_col

The name of the boolean column telling whether this row contains a segment that should have a starting point.

end_arrow_col

The name of the boolean column telling whether this row contains a segment that should have an ending arrow.

Re_names

The name of the column of rebound names.

Re_values

The name of the column of rebound values.

y_names_col

The name of the column of y-axis variable names.

y_vals_col

The name of the column of y-axis variable names.

Examples

graph_df_colnames
#> $colour_col
#> [1] "colour"
#> 
#> $size_col
#> [1] "size"
#> 
#> $shape_col
#> [1] "shape"
#> 
#> $fill_col
#> [1] "fill"
#> 
#> $stroke_col
#> [1] "stroke"
#> 
#> $linewidth_col
#> [1] "linewidth"
#> 
#> $linetype_col
#> [1] "linetype"
#> 
#> $graph_type_col
#> [1] "graph_type"
#> 
#> $line_name_col
#> [1] "line_name"
#> 
#> $point_name_col
#> [1] "point_name"
#> 
#> $slope_col
#> [1] "slope"
#> 
#> $intercept_col
#> [1] "intercept"
#> 
#> $x_col
#> [1] "x"
#> 
#> $y_col
#> [1] "y"
#> 
#> $xend_col
#> [1] "xend"
#> 
#> $yend_col
#> [1] "yend"
#> 
#> $qs1_qs0_col
#> [1] "qs1_qs0"
#> 
#> $Co1_Co0_col
#> [1] "Co1_Co0"
#> 
#> $f_Cs_orig_col
#> [1] "f_Cs_orig"
#> 
#> $sigma_col
#> [1] "sigma"
#> 
#> $start_point_col
#> [1] "start_point"
#> 
#> $end_arrow_col
#> [1] "end_arrow"
#> 
#> $Re_names
#> [1] "Re_names"
#> 
#> $Re_values
#> [1] "Re_values"
#> 
#> $y_names_col
#> [1] "y_names"
#> 
#> $y_vals_col
#> [1] "y_vals"
#>