Indifference curve lines

This function calculates indifference curves to be drawn on the preferences graph.

indifference_lines(
  .rebound_data,
  graph_type = ReboundTools::graph_types$preferences,
  graph_params = ReboundTools::default_graph_params,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  sigma = ReboundTools::orig_vars$sigma,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar
)

Arguments

.rebound_data
A data frame of rebound information, likely created by rebound_analysis().

graph_type
See ReboundTools::graph_types. Default is ReboundTools::graph_types$preferences.

graph_params
Parameters that control the appearance of the graph. Default is ReboundTools::default_graph_params.

q_dot_s_orig, C_dot_o_orig, f_Cs_orig, sigma
See ReboundTools::orig_vars.

q_dot_s_hat
See ReboundTools::hat_vars.

q_dot_s_bar, C_dot_o_bar
See ReboundTools::bar_vars.

Value

A data frame of indifference curves for a preferences graph.

Details

The preferences graph is always indexed, so there is no indexed argument.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  indifference_lines()

#> # A tibble: 617 x 15
#>    Reference Case  Original Upgrade graph_type line_name colour  size linetype
#>    <chr>     <chr> <chr>    <chr>   <chr>      <chr>     <chr>  <dbl> <chr>   
#>  1 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  2 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  3 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  4 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  5 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  6 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  7 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  8 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#>  9 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#> 10 None yet  Car   Ford Fu… Ford F… Preferenc… orig      gray     0.1 solid   
#> # … with 607 more rows, and 6 more variables: qs1_qs0 <dbl>, Co1_Co0 <dbl>,
#> #   f_Cs_orig <dbl>, sigma <dbl>, x <dbl>, y <dbl>