Indifference curve lines

This function calculates indifference curves to be drawn on the consumption path graph.

Usage

indifference_lines(
  .rebound_data,
  graph_type = ReboundTools::graph_types$consumption,
  graph_params = ReboundTools::path_graph_params,
  q_dot_s_orig = ReboundTools::orig_vars$q_dot_s_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  f_Cs_orig = ReboundTools::orig_vars$f_Cs_orig,
  sigma = ReboundTools::orig_vars$sigma,
  q_dot_s_hat = ReboundTools::hat_vars$q_dot_s_hat,
  q_dot_s_bar = ReboundTools::bar_vars$q_dot_s_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar
)

Arguments

.rebound_data

A data frame of rebound information, likely created by rebound_analysis().

graph_type

See ReboundTools::graph_types. Default is ReboundTools::graph_types$consumption.

graph_params

Parameters that control the appearance of the graph. Default is ReboundTools::path_graph_params.

q_dot_s_orig, C_dot_o_orig, f_Cs_orig, sigma

See ReboundTools::orig_vars.

q_dot_s_hat

See ReboundTools::hat_vars.

q_dot_s_bar, C_dot_o_bar

See ReboundTools::bar_vars.

Value

A data frame of indifference curves for a consumption path graph.

Details

The consumption path graph is always indexed, so there is no indexed argument.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  indifference_lines()
#> # A tibble: 617 × 15
#>    Refere…¹ Case  Origi…² Upgrade graph…³ line_…⁴ colour linew…⁵ linet…⁶ qs1_qs0
#>    <chr>    <chr> <chr>   <chr>   <chr>   <chr>   <chr>    <dbl> <chr>     <dbl>
#>  1 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  2 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  3 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  4 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  5 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  6 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  7 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  8 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#>  9 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#> 10 None yet Car   Ford F… Ford F… Consum… orig    black      0.5 solid         1
#> # … with 607 more rows, 5 more variables: Co1_Co0 <dbl>, f_Cs_orig <dbl>,
#> #   sigma <dbl>, x <dbl>, y <dbl>, and abbreviated variable names ¹​Reference,
#> #   ²​Original, ³​graph_type, ⁴​line_name, ⁵​linewidth, ⁶​linetype