Iso-energy lines for energy rebound path graphs

This function creates a data frame of constant-energy lines at various rebound stages. The iso lines are described by slope and intercept.

iso_energy_lines(
  .rebound_data,
  indexed = FALSE,
  graph_type = ReboundTools::graph_types$energy,
  graph_params = ReboundTools::path_graph_params,
  I_E = ReboundTools::eeu_base_params$I_E,
  k = ReboundTools::eeu_base_params$k,
  E_dot_s_orig = ReboundTools::orig_vars$E_dot_s_orig,
  E_dot_emb_orig = ReboundTools::orig_vars$E_dot_emb_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  S_dot_dev = ReboundTools::star_vars$S_dot_dev,
  E_dot_s_star = ReboundTools::star_vars$E_dot_s_star,
  E_dot_emb_star = ReboundTools::star_vars$E_dot_emb_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  E_dot_s_hat = ReboundTools::hat_vars$E_dot_s_hat,
  E_dot_emb_hat = ReboundTools::hat_vars$E_dot_emb_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat,
  N_dot_hat = ReboundTools::hat_vars$N_dot_hat,
  E_dot_s_bar = ReboundTools::bar_vars$E_dot_s_bar,
  E_dot_emb_bar = ReboundTools::bar_vars$E_dot_emb_bar,
  C_dot_md_bar = ReboundTools::bar_vars$C_dot_md_bar,
  C_dot_o_bar = ReboundTools::bar_vars$C_dot_o_bar,
  E_dot_s_tilde = ReboundTools::tilde_vars$E_dot_s_tilde,
  E_dot_emb_tilde = ReboundTools::tilde_vars$E_dot_emb_tilde,
  C_dot_md_tilde = ReboundTools::tilde_vars$C_dot_md_tilde,
  C_dot_o_tilde = ReboundTools::tilde_vars$C_dot_o_tilde
)

Arguments

.rebound_data

A data frame of rebound information, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_type

See ReboundTools::graph_types.

graph_params

See ReboundTools::path_graph_params.

I_E, k

See ReboundTools::eeu_base_params.

E_dot_s_orig, E_dot_emb_orig, C_dot_md_orig, C_dot_o_orig

See ReboundTools::orig_vars.

S_dot_dev, E_dot_s_star, E_dot_emb_star, C_dot_md_star, C_dot_o_star

See ReboundTools::star_vars.

E_dot_s_hat, E_dot_emb_hat, C_dot_md_hat, C_dot_o_hat, N_dot_hat

See ReboundTools::hat_vars.

E_dot_s_bar, E_dot_emb_bar, C_dot_md_bar, C_dot_o_bar

See ReboundTools::bar_vars.

E_dot_s_tilde, E_dot_emb_tilde, C_dot_md_tilde, C_dot_o_tilde

See ReboundTools::tilde_vars.

Value

A data frame containing iso-energy lines to be drawn on an energy rebound path graph.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  iso_energy_lines()
#> # A tibble: 12 × 11
#>    Reference Case  Original Upgrade graph…¹ line_…² colour linew…³ linet…⁴ slope
#>    <chr>     <chr> <chr>    <chr>   <chr>   <chr>   <chr>    <dbl> <chr>   <dbl>
#>  1 None yet  Car   Ford Fu… Ford F… Energy  Re = 1… black      0.3 solid      -1
#>  2 None yet  Lamp  Incande… LED     Energy  Re = 1… black      0.3 solid      -1
#>  3 None yet  Car   Ford Fu… Ford F… Energy  Re = 0% black      0.3 solid      -1
#>  4 None yet  Lamp  Incande… LED     Energy  Re = 0% black      0.3 solid      -1
#>  5 None yet  Car   Ford Fu… Ford F… Energy  Re_empl black      0.1 solid      -1
#>  6 None yet  Lamp  Incande… LED     Energy  Re_empl black      0.1 solid      -1
#>  7 None yet  Car   Ford Fu… Ford F… Energy  Re_sub  black      0.1 solid      -1
#>  8 None yet  Lamp  Incande… LED     Energy  Re_sub  black      0.1 solid      -1
#>  9 None yet  Car   Ford Fu… Ford F… Energy  Re_inc  black      0.1 solid      -1
#> 10 None yet  Lamp  Incande… LED     Energy  Re_inc  black      0.1 solid      -1
#> 11 None yet  Car   Ford Fu… Ford F… Energy  Re_tot  black      0.1 solid      -1
#> 12 None yet  Lamp  Incande… LED     Energy  Re_tot  black      0.1 solid      -1
#> # … with 1 more variable: intercept <dbl>, and abbreviated variable names
#> #   ¹​graph_type, ²​line_name, ³​linewidth, ⁴​linetype