Create a data frame of iso-expenditure lines

This function creates a data frame of constant expenditure lines at various rebound stages. The iso lines are described by slope and intercept.

iso_expenditure_lines(
  .rebound_data,
  indexed = FALSE,
  graph_type = ReboundTools::graph_types$expenditure,
  graph_params = ReboundTools::path_graph_params,
  C_dot_s_orig = ReboundTools::orig_vars$C_dot_s_orig,
  C_dot_cap_orig = ReboundTools::orig_vars$C_dot_cap_orig,
  C_dot_md_orig = ReboundTools::orig_vars$C_dot_md_orig,
  C_dot_o_orig = ReboundTools::orig_vars$C_dot_o_orig,
  G_dot = ReboundTools::star_vars$G_dot,
  C_dot_s_hat = ReboundTools::hat_vars$C_dot_s_hat,
  C_dot_cap_star = ReboundTools::star_vars$C_dot_cap_star,
  C_dot_md_star = ReboundTools::star_vars$C_dot_md_star,
  C_dot_o_star = ReboundTools::star_vars$C_dot_o_star,
  C_dot_cap_hat = ReboundTools::hat_vars$C_dot_cap_hat,
  C_dot_md_hat = ReboundTools::hat_vars$C_dot_md_hat,
  C_dot_o_hat = ReboundTools::hat_vars$C_dot_o_hat
)

Arguments

.rebound_data

A data frame of rebound information, likely created by rebound_analysis().

indexed

A boolean telling whether the rebound path should be indexed to 1 at its start.

graph_type

See ReboundTools::graph_types.

graph_params

See ReboundTools::graph_params.

C_dot_s_orig, C_dot_cap_orig, C_dot_md_orig, C_dot_o_orig

See ReboundTools::orig_vars.

G_dot

See ReboundTools::star_vars.

C_dot_s_hat, C_dot_cap_hat, C_dot_md_hat, C_dot_o_hat

See ReboundTools::hat_vars.

C_dot_cap_star, C_dot_md_star, C_dot_o_star

See ReboundTools::star_vars.

Value

A data frame of iso-expenditure lines for an expenditure path graph.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  iso_expenditure_lines()
#> # A tibble: 8 × 11
#>   Reference Case  Original  Upgrade graph…¹ line_…² colour linew…³ linet…⁴ slope
#>   <chr>     <chr> <chr>     <chr>   <chr>   <chr>   <chr>    <dbl> <chr>   <dbl>
#> 1 None yet  Car   Ford Fus… Ford F… Expend… orig    black      0.3 solid      -1
#> 2 None yet  Lamp  Incandes… LED     Expend… orig    black      0.3 solid      -1
#> 3 None yet  Car   Ford Fus… Ford F… Expend… G_dot   black      0.3 solid      -1
#> 4 None yet  Lamp  Incandes… LED     Expend… G_dot   black      0.3 solid      -1
#> 5 None yet  Car   Ford Fus… Ford F… Expend… star    black      0.3 solid      -1
#> 6 None yet  Lamp  Incandes… LED     Expend… star    black      0.3 solid      -1
#> 7 None yet  Car   Ford Fus… Ford F… Expend… hat     black      0.3 solid      -1
#> 8 None yet  Lamp  Incandes… LED     Expend… hat     black      0.3 solid      -1
#> # … with 1 more variable: intercept <dbl>, and abbreviated variable names
#> #   ¹​graph_type, ²​line_name, ³​linewidth, ⁴​linetype