Key analysis variables

This is the list of key rebound analysis variables. These variables are differenced later in the analysis.

Usage

key_analysis_vars

Format

A string list with 12 entries.

eta_engr_units

Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service, in engineering units.

eta

Energy service efficiency, calculated by energy service divided by final energy consumed to provide that service.

p_s

Energy service price [service/MJ], calculated by p_E/eta.

q_dot_s

The rate of energy service consumption [service/yr], calculated by eta*E_dot_s.

E_dot_s

The rate of final energy consumption by the energy conversion device [MJ/yr].

E_dot_emb

The rate of embodied energy demand by the energy conversion device [MJ/yr], calculated by E_emb/t_life.

C_dot_s

The expenditure rate of energy consumption by the device [$/yr], calculated by p_s*q_dot_s.

C_dot_cap

The capital expenditure rate of the device [$/yr], calculated by C_cap/t_own.

C_dot_md

The maintenance and disposal expenditure rate of the device [$/yr].

C_dot_o

The other goods consumption rate [$/yr], calculated, initially, as a residual of the budget constraint.

N_dot

Net income [$/yr].

M_dot

Real income [$/yr].

Examples

key_analysis_vars
#>   eta_engr_units              eta              p_s          q_dot_s 
#> "eta_engr_units"            "eta"            "p_s"        "q_dot_s" 
#>          E_dot_s        E_dot_emb          C_dot_s        C_dot_cap 
#>        "E_dot_s"      "E_dot_emb"        "C_dot_s"      "C_dot_cap" 
#>         C_dot_md          C_dot_o            N_dot            M_dot 
#>       "C_dot_md"        "C_dot_o"          "N_dot"          "M_dot"