EEU orig data

This is the list of the derived variables at the original (pre-EEU) stage of a rebound analysis.

Usage

orig_vars

Format

A string list with 31 entries.

tau_alpha_orig

The original discount multiplier for beginning-of-life expenses, such as capital cost.

tau_omega_orig

The original discount multiplier for end-of-life expenses, such as disposal cost.

p_E

The price of energy [$/MJ], calculated by p_E_engr_units / MJ_engr_unit.

p_E_orig

The price of energy [$/MJ].

q_dot_s_orig

The original (pre-EEU) consumption rate of the energy service. Example units are [miles/yr] [lumen-hours/yr].

C_cap_orig

The net capital expenditure of the original device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].

M_dot_orig

The disposable income rate, exclusive of taxes and savings [$/yr].

eta_engr_units_orig

The original (pre-EEU) energy service efficiency. This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of eta_engr_units_orig is assumed to be the same as the denominator unit of MJ_engr_unit.

eta_orig

Energy service efficiency of the original (pre-EEU) device on a per-MJ basis [service/MJ], calculated by eta_engr_units_orig / MJ_engr_unit.

E_dot_s_orig

The final energy consumption rate of the original (pre-EEU) device [MJ/yr], calculated by q_dot_s_orig / eta_orig.

C_dot_cap_orig

The capital expenditure rate of the device without discounting [$/yr], calculated by C_cap/t_own.

tau_alpha_C_dot_cap_orig

The product of tau_alpha and the undiscounted capital cost rate, itself annualized and discounted in [$/yr].

C_dot_om_orig

The operations and maintenance expenditure rate of the device [$/yr].

C_d_orig

The disposal cost for the device [$/yr].

C_dot_d_orig

The disposal cost rate of the device without discounting [$/yr].

tau_omega_C_dot_d_orig

The product of tau_omega and the undiscounted disposal cost rate, itself annualized and discounted in [$/yr].

C_dot_omd_orig

The operations, maintenance, and disposal expenditure rate of the device [$/yr].

p_s_orig

The original (pre-EEU) energy service price [$/service], calculated by p_E / eta_orig.

C_dot_s_orig

The original (pre-EEU) rate of energy expenditures for the device [$/yr], calculated by p_E * E_dot_s_orig.

C_dot_g_orig

The original (pre-EEU) rate of expenditure on other goods [$/yr], calculated by M_dot_orig - C_dot_s_orig - C_dot_cap_orig - C_dot_md_orig.

f_Cs_orig

The original (pre-EEU) fraction of the energy and other budget spent on the energy service [–], calculated by C_dot_s_orig / (C_dot_s_orig + C_dot_g_orig).

e_qs_ps_C_orig

The original compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [–], calculated by e_qs_ps_UC_orig + f_Cs_orig*e_qs_M.

e_qg_ps_C_orig

The original compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [–], calculated by f_Cs_orig*(f_Cs_orig + e_qs_ps_UC_orig) / (f_Cs_orig - 1).

e_qg_ps_UC_orig

The original uncompensated ("UC") Marshallian energy service price ("ps") elasticity ("e") of other goods ("qo") consumption (cross-price elasticity) [–].

sigma

The elasticity of substitution between energy service consumption and other goods consumption [–].

rho

The exponent in the CES utility model, defined as rho = 1/sigma - 1 [–].

E_emb_orig

The embodied energy of the original (pre-EEU) device [MJ].

t_life_orig

The expected lifetime of the original (pre-EEU) device [yr].

E_dot_emb_orig

The original (pre-EEU) rate of embodied energy demand[MJ/yr], calculated by E_dot_emb / t_orig.

N_dot_orig

The original (pre-EEU) freed cash rate [$/yr], exactly 0.

Examples

orig_vars
#> $tau_alpha_orig
#> [1] "tau_alpha_orig"
#> 
#> $tau_omega_orig
#> [1] "tau_omega_orig"
#> 
#> $p_E
#> [1] "p_E"
#> 
#> $p_E_orig
#> [1] "p_E_orig"
#> 
#> $q_dot_s_orig
#> [1] "q_dot_s_orig"
#> 
#> $M_dot_orig
#> [1] "M_dot_orig"
#> 
#> $C_cap_orig
#> [1] "C_cap_orig"
#> 
#> $eta_engr_units_orig
#> [1] "eta_engr_units_orig"
#> 
#> $eta_orig
#> [1] "eta_orig"
#> 
#> $E_dot_s_orig
#> [1] "E_dot_s_orig"
#> 
#> $C_dot_cap_orig
#> [1] "C_dot_cap_orig"
#> 
#> $tau_alpha_C_dot_cap_orig
#> [1] "tau_alpha_C_dot_cap_orig"
#> 
#> $p_s_orig
#> [1] "p_s_orig"
#> 
#> $C_dot_s_orig
#> [1] "C_dot_s_orig"
#> 
#> $C_dot_om_orig
#> [1] "C_dot_om_orig"
#> 
#> $C_d_orig
#> [1] "C_d_orig"
#> 
#> $C_dot_d_orig
#> [1] "C_dot_d_orig"
#> 
#> $tau_omega_C_dot_d_orig
#> [1] "tau_omega_C_dot_d_orig"
#> 
#> $C_dot_omd_orig
#> [1] "C_dot_omd_orig"
#> 
#> $C_dot_g_orig
#> [1] "C_dot_g_orig"
#> 
#> $f_Cs_orig
#> [1] "f_Cs_orig"
#> 
#> $e_qs_ps_C_orig
#> [1] "e_qs_ps_C_orig"
#> 
#> $e_qs_ps_UC_orig
#> [1] "e_qs_ps_UC_orig"
#> 
#> $e_qg_ps_C_orig
#> [1] "e_qg_ps_C_orig"
#> 
#> $e_qg_ps_UC_orig
#> [1] "e_qg_ps_UC_orig"
#> 
#> $sigma
#> [1] "sigma"
#> 
#> $rho
#> [1] "rho"
#> 
#> $E_emb_orig
#> [1] "E_emb_orig"
#> 
#> $t_life_orig
#> [1] "t_life_orig"
#> 
#> $E_dot_emb_orig
#> [1] "E_dot_emb_orig"
#> 
#> $N_dot_orig
#> [1] "N_dot_orig"
#>