EEU orig data

This is the list of the derived variables at the original (pre-EEU) stage of a rebound analysis.

orig_vars

Format

A string list with 22 entries.

p_E

The price of energy [$/MJ], calculated by p_E_engr_units / MJ_engr_unit.

q_dot_s_orig

The original (pre-EEU) consumption rate of the energy service. Example units are [miles/year] [lumen-hours/year].

C_cap_orig

The net capital expenditure of the original device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].

M_dot_orig

The disposable income rate, exclusive of taxes and savings [$/year].

t_own_orig

The expected ownership duration of the original device [year].

C_dot_md_orig

The original (pre-EEU) maintenance and disposal expenditure rate [$/year].

eta_engr_units_orig

The original (pre-EEU) energy service efficiency. This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of eta_engr_units_orig is assumed to be the same as the denominator unit of MJ_engr_unit.

eta_orig

Energy service efficiency of the original (pre-EEU) device on a per-MJ basis [service/MJ], calculated by eta_engr_units_orig / MJ_engr_unit.

E_dot_s_orig

The final energy consumption rate of the original (pre-EEU) device [MJ/year], calculated by q_dot_s_orig / eta_orig.

C_dot_cap_orig

The original (pre-EEU) capital expenditure rate [$/year], calculated by C_cap_orig / t_orig.

p_s_orig

The original (pre-EEU) energy service price [$/service], calculated by p_E / eta_orig.

C_dot_s_orig

The original (pre-EEU) rate of energy expenditures for the device [$/year], calculated by p_E * E_dot_s_orig.

C_dot_o_orig

The original (pre-EEU) rate of expenditure on other goods [$/year], calculated by M_dot_orig - C_dot_s_orig - C_dot_cap_orig - C_dot_md_orig.

f_Cs_orig

The original (pre-EEU) fraction of the energy and other budget spent on the energy service [--], calculated by C_dot_s_orig / (C_dot_s_orig + C_dot_o_orig).

e_qs_ps_C

The compensated energy service price ("ps") elasticity ("e") of energy service ("qs") consumption [--], calculated by e_qs_ps_UC + f_Cs_orig*e_qs_M.

e_qo_ps_C

The compensated energy service price ("ps") elasticity ("e") of other goods ("qo") consumption [--], calculated by f_Cs_orig*(f_Cs_orig + e_qs_ps_UC) / (f_Cs_orig - 1).

sigma

The elasticity of substitution between energy service consumption and other goods consumption [--].

rho

The exponent in the CES utility model, defined as rho = 1/sigma - 1 [--].

E_emb_orig

The embodied energy of the original (pre-EEU) device [MJ].

t_life_orig

The expected lifetime of the original (pre-EEU) device [year].

E_dot_emb_orig

The original (pre-EEU) rate of embodied energy demand[MJ/year], calculated by E_dot_emb / t_orig.

N_dot_orig

The original (pre-EEU) freed cash rate [$/year], exactly 0.

Examples

orig_vars

#> $p_E
#> [1] "p_E"
#> 
#> $q_dot_s_orig
#> [1] "q_dot_s_orig"
#> 
#> $M_dot_orig
#> [1] "M_dot_orig"
#> 
#> $C_cap_orig
#> [1] "C_cap_orig"
#> 
#> $t_own_orig
#> [1] "t_own_orig"
#> 
#> $C_dot_md_orig
#> [1] "C_dot_md_orig"
#> 
#> $eta_engr_units_orig
#> [1] "eta_engr_units_orig"
#> 
#> $eta_orig
#> [1] "eta_orig"
#> 
#> $E_dot_s_orig
#> [1] "E_dot_s_orig"
#> 
#> $C_dot_cap_orig
#> [1] "C_dot_cap_orig"
#> 
#> $p_s_orig
#> [1] "p_s_orig"
#> 
#> $C_dot_s_orig
#> [1] "C_dot_s_orig"
#> 
#> $C_dot_o_orig
#> [1] "C_dot_o_orig"
#> 
#> $f_Cs_orig
#> [1] "f_Cs_orig"
#> 
#> $e_qs_ps_C
#> [1] "e_qs_ps_C"
#> 
#> $e_qo_ps_C
#> [1] "e_qo_ps_C"
#> 
#> $sigma
#> [1] "sigma"
#> 
#> $rho
#> [1] "rho"
#> 
#> $E_emb_orig
#> [1] "E_emb_orig"
#> 
#> $t_life_orig
#> [1] "t_life_orig"
#> 
#> $E_dot_emb_orig
#> [1] "E_dot_emb_orig"
#> 
#> $N_dot_orig
#> [1] "N_dot_orig"
#>