Run parametric analyses

After defining a base case for rebound analyses, parametric analyses and sensitivity studies are often helpful. This function provides a convenient way to do parametric studies.

Usage

parametric_analysis(
  original_cases,
  parameterization,
  include_orig_point = TRUE,
  case_colname = ReboundTools::eeu_base_params$case,
  point_type_colname = ReboundTools::parametric_analysis_point_types$point_type_colname,
  orig = ReboundTools::parametric_analysis_point_types$orig,
  sweep = ReboundTools::parametric_analysis_point_types$sweep
)

Arguments

original_cases

A data frame with baseline data from which parametric studies are desired. This data frame is expected to contain 1 row, a single case.

parameterization

A named list of lists. See details.

include_orig_point

A boolean that tells whether the original point for each case (from original_cases) should be included in the output. Default is TRUE.

case_colname

See ReboundTools::eeu_base_params.

point_type_colname, orig, sweep

See ReboundTools::parametric_analysis_point_types.

Value

A data frame containing results of parametric studies.

Details

original_cases must have the same structure as the example data frames in this package. See the file at sample_eeu_data_path(). Failures will occur if original_cases contains columns calculated by rebound_analysis().

parameterization is a list and must have the following structure. Top-level elements are named for cases. Case names must match the case column in original_cases. Not every case in original_cases must be present. Each top-level element in parameterization must itself also be a list. Each 2nd-level list must contain named vectors, with each name corresponding to a parameter in original_cases that will be swept in the parametric study. If more than one entry is included in a 2nd-level list, all combinations of values will be used, via expand.grid(). Note that the original value of each parameter (in original_cases) will not be inserted into the parameterization lists, so be sure to include that value, if desired.

Note that the case names in parameterization (i.e., the names of the top-level entries) are filtering toward original_cases. I.e., the output of this function will contain only those cases in parameterization.

Examples

car_case <- load_eeu_data() %>% 
  dplyr::filter(.data[[ReboundTools::eeu_base_params$case]] == "Car")
# This example creates a parametric study for the "Car" case only,
# sweeping through values of both `k` and `p_E_engr_units`.
params <- list(Car = list(k = seq(0, 2, by = 1), 
                          p_E_engr_units = seq(1.5, 2.5, by = 0.25)))
res <- parametric_analysis(car_case, params)
dplyr::glimpse(res)
#> Rows: 17
#> Columns: 177
#> $ Reference                  <chr> "None yet", "None yet", "None yet", "None y…
#> $ Case                       <chr> "Car", "Car", "Car", "Car", "Car", "Car", "…
#> $ Original                   <chr> "Ford Fusion", "Ford Fusion", "Ford Fusion"…
#> $ Upgrade                    <chr> "Ford Fusion Hybrid", "Ford Fusion Hybrid",…
#> $ service_unit               <chr> "mile", "mile", "mile", "mile", "mile", "mi…
#> $ energy_engr_unit           <chr> "gal", "gal", "gal", "gal", "gal", "gal", "…
#> $ `MJ/energy_engr_unit`      <dbl> 126.6216, 126.6216, 126.6216, 126.6216, 126…
#> $ I_E                        <dbl> 3.389339, 3.389339, 3.389339, 3.389339, 3.3…
#> $ k                          <dbl> 1, 1, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0, 1, 2, 0…
#> $ p_E_engr_units             <dbl> 2.21, 2.21, 1.50, 1.50, 1.50, 1.75, 1.75, 1…
#> $ e_qs_ps_UC_orig            <dbl> -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -…
#> $ e_qs_M                     <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ e_qo_M                     <dbl> 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1…
#> $ eta_engr_units_orig        <dbl> 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25,…
#> $ eta_engr_units_star        <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42, 42,…
#> $ q_dot_s_orig               <dbl> 14425, 14425, 14425, 14425, 14425, 14425, 1…
#> $ M_dot_orig                 <dbl> 27401.28, 27401.28, 27401.28, 27401.28, 274…
#> $ C_cap_orig                 <dbl> 28216.1, 28216.1, 28216.1, 28216.1, 28216.1…
#> $ t_own_orig                 <dbl> 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7…
#> $ C_cap_star                 <dbl> 27523.4, 27523.4, 27523.4, 27523.4, 27523.4…
#> $ t_own_star                 <dbl> 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7, 7…
#> $ C_dot_md_orig              <dbl> 2861.134, 2861.134, 2861.134, 2861.134, 286…
#> $ C_dot_md_star              <dbl> 2774.668, 2774.668, 2774.668, 2774.668, 277…
#> $ E_emb_orig                 <dbl> 34000, 34000, 34000, 34000, 34000, 34000, 3…
#> $ t_life_orig                <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14,…
#> $ E_emb_star                 <dbl> 40000, 40000, 40000, 40000, 40000, 40000, 4…
#> $ t_life_star                <dbl> 14, 14, 14, 14, 14, 14, 14, 14, 14, 14, 14,…
#> $ point_type                 <chr> "orig", "sweep", "sweep", "sweep", "sweep",…
#> $ p_E                        <dbl> 0.01745357, 0.01745357, 0.01184632, 0.01184…
#> $ eta_orig                   <dbl> 0.1974386, 0.1974386, 0.1974386, 0.1974386,…
#> $ E_dot_s_orig               <dbl> 73060.68, 73060.68, 73060.68, 73060.68, 730…
#> $ C_dot_cap_orig             <dbl> 4030.871, 4030.871, 4030.871, 4030.871, 403…
#> $ p_s_orig                   <dbl> 0.0884, 0.0884, 0.0600, 0.0600, 0.0600, 0.0…
#> $ C_dot_s_orig               <dbl> 1275.17, 1275.17, 865.50, 865.50, 865.50, 1…
#> $ C_dot_o_orig               <dbl> 19234.10, 19234.10, 19643.77, 19643.77, 196…
#> $ f_Cs_orig                  <dbl> 0.06217529, 0.06217529, 0.04220043, 0.04220…
#> $ e_qo_ps_UC_orig            <dbl> -0.05966762, -0.05966762, -0.03965379, -0.0…
#> $ e_qs_ps_C_orig             <dbl> -0.03782471, -0.03782471, -0.05779957, -0.0…
#> $ e_qo_ps_C_orig             <dbl> 0.002507678, 0.002507678, 0.002546636, 0.00…
#> $ sigma                      <dbl> 0.04033238, 0.04033238, 0.06034621, 0.06034…
#> $ rho                        <dbl> -23.79397, -23.79397, -15.57105, -15.57105,…
#> $ E_dot_emb_orig             <dbl> 2428.571, 2428.571, 2428.571, 2428.571, 242…
#> $ N_dot_orig                 <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ eta_star                   <dbl> 0.3316969, 0.3316969, 0.3316969, 0.3316969,…
#> $ eta_ratio                  <dbl> 1.68, 1.68, 1.68, 1.68, 1.68, 1.68, 1.68, 1…
#> $ S_dot_dev                  <dbl> 29572.18, 29572.18, 29572.18, 29572.18, 295…
#> $ G_dot                      <dbl> 516.1402, 516.1402, 350.3214, 350.3214, 350…
#> $ p_s_star                   <dbl> 0.05261905, 0.05261905, 0.03571429, 0.03571…
#> $ q_dot_s_star               <dbl> 14425, 14425, 14425, 14425, 14425, 14425, 1…
#> $ C_dot_cap_star             <dbl> 3931.914, 3931.914, 3931.914, 3931.914, 393…
#> $ E_dot_emb_star             <dbl> 2857.143, 2857.143, 2857.143, 2857.143, 285…
#> $ C_dot_s_star               <dbl> 759.0298, 759.0298, 515.1786, 515.1786, 515…
#> $ M_dot_star                 <dbl> 27401.28, 27401.28, 27401.28, 27401.28, 274…
#> $ N_dot_star                 <dbl> 701.5635, 701.5635, 535.7447, 535.7447, 535…
#> $ C_dot_o_star               <dbl> 19234.10, 19234.10, 19643.77, 19643.77, 196…
#> $ E_dot_s_star               <dbl> 43488.5, 43488.5, 43488.5, 43488.5, 43488.5…
#> $ f_Cs_star                  <dbl> 0.03796453, 0.03796453, 0.02555582, 0.02555…
#> $ e_qs_ps_UC_star            <dbl> -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -0.1, -…
#> $ e_qo_ps_UC_star            <dbl> -0.05966762, -0.05966762, -0.03965379, -0.0…
#> $ e_qs_ps_C_star             <dbl> -0.03782471, -0.03782471, -0.05779957, -0.0…
#> $ e_qo_ps_C_star             <dbl> 0.002507678, 0.002507678, 0.002546636, 0.00…
#> $ eta_engr_units_hat         <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42, 42,…
#> $ eta_hat                    <dbl> 0.3316969, 0.3316969, 0.3316969, 0.3316969,…
#> $ p_s_hat                    <dbl> 0.05261905, 0.05261905, 0.03571429, 0.03571…
#> $ C_dot_cap_hat              <dbl> 3931.914, 3931.914, 3931.914, 3931.914, 393…
#> $ C_dot_md_hat               <dbl> 2774.668, 2774.668, 2774.668, 2774.668, 277…
#> $ E_dot_emb_hat              <dbl> 2857.143, 2857.143, 2857.143, 2857.143, 285…
#> $ M_dot_hat                  <dbl> 27401.28, 27401.28, 27401.28, 27401.28, 274…
#> $ q_dot_s_hat                <dbl> 14714.74, 14714.74, 14868.07, 14868.07, 148…
#> $ E_dot_s_hat                <dbl> 44362.00, 44362.00, 44824.26, 44824.26, 448…
#> $ C_dot_s_hat                <dbl> 774.2755, 774.2755, 531.0024, 531.0024, 531…
#> $ C_dot_o_hat                <dbl> 19214.16, 19214.16, 19623.07, 19623.07, 196…
#> $ f_Cs_hat                   <dbl> 0.03873618, 0.03873618, 0.02634715, 0.02634…
#> $ e_qs_ps_C_hat              <dbl> -0.03877006, -0.03877006, -0.05875626, -0.0…
#> $ e_qo_ps_C_hat              <dbl> 0.001562322, 0.001562322, 0.001589951, 0.00…
#> $ e_qs_ps_UC_hat             <dbl> -0.07750624, -0.07750624, -0.08510341, -0.0…
#> $ e_qo_ps_UC_hat             <dbl> -0.03717385, -0.03717385, -0.02475720, -0.0…
#> $ N_dot_hat                  <dbl> 706.2625, 706.2625, 540.6200, 540.6200, 540…
#> $ M_dot_hat_prime            <dbl> 19988.43, 19988.43, 20154.08, 20154.08, 201…
#> $ eta_engr_units_bar         <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42, 42,…
#> $ eta_bar                    <dbl> 0.3316969, 0.3316969, 0.3316969, 0.3316969,…
#> $ p_s_bar                    <dbl> 0.05261905, 0.05261905, 0.03571429, 0.03571…
#> $ C_dot_cap_bar              <dbl> 3931.914, 3931.914, 3931.914, 3931.914, 393…
#> $ C_dot_md_bar               <dbl> 2774.668, 2774.668, 2774.668, 2774.668, 277…
#> $ E_dot_emb_bar              <dbl> 2857.143, 2857.143, 2857.143, 2857.143, 285…
#> $ M_dot_bar                  <dbl> 27401.28, 27401.28, 27401.28, 27401.28, 274…
#> $ q_dot_s_bar                <dbl> 15234.66, 15234.66, 15266.89, 15266.89, 152…
#> $ E_dot_s_bar                <dbl> 45929.47, 45929.47, 46026.64, 46026.64, 460…
#> $ C_dot_s_bar                <dbl> 801.6334, 801.6334, 545.2462, 545.2462, 545…
#> $ C_dot_o_bar                <dbl> 19893.06, 19893.06, 20149.45, 20149.45, 201…
#> $ f_Cs_bar                   <dbl> 0.03873618, 0.03873618, 0.02634715, 0.02634…
#> $ e_qs_ps_UC_bar             <dbl> -0.07750624, -0.07750624, -0.08510341, -0.0…
#> $ e_qo_ps_UC_bar             <dbl> -0.03717385, -0.03717385, -0.02475720, -0.0…
#> $ e_qs_ps_C_bar              <dbl> -0.03877006, -0.03877006, -0.05875626, -0.0…
#> $ e_qo_ps_C_bar              <dbl> 0.001562322, 0.001562322, 0.001589951, 0.00…
#> $ N_dot_bar                  <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ eta_engr_units_tilde       <dbl> 42, 42, 42, 42, 42, 42, 42, 42, 42, 42, 42,…
#> $ eta_tilde                  <dbl> 0.3316969, 0.3316969, 0.3316969, 0.3316969,…
#> $ p_s_tilde                  <dbl> 0.05261905, 0.05261905, 0.03571429, 0.03571…
#> $ C_dot_cap_tilde            <dbl> 3931.914, 3931.914, 3931.914, 3931.914, 393…
#> $ C_dot_md_tilde             <dbl> 2774.668, 2774.668, 2774.668, 2774.668, 277…
#> $ E_dot_emb_tilde            <dbl> 2857.143, 2857.143, 2857.143, 2857.143, 285…
#> $ M_dot_tilde                <dbl> 27401.28, 27401.28, 27401.28, 27401.28, 274…
#> $ q_dot_s_tilde              <dbl> 15234.66, 15234.66, 15266.89, 15266.89, 152…
#> $ E_dot_s_tilde              <dbl> 45929.47, 45929.47, 46026.64, 46026.64, 460…
#> $ C_dot_s_tilde              <dbl> 801.6334, 801.6334, 545.2462, 545.2462, 545…
#> $ C_dot_o_tilde              <dbl> 19893.06, 19893.06, 20149.45, 20149.45, 201…
#> $ f_Cs_tilde                 <dbl> 0.03873618, 0.03873618, 0.02634715, 0.02634…
#> $ e_qs_ps_UC_tilde           <dbl> -0.07750624, -0.07750624, -0.08510341, -0.0…
#> $ e_qo_ps_UC_tilde           <dbl> -0.03717385, -0.03717385, -0.02475720, -0.0…
#> $ e_qs_ps_C_tilde            <dbl> -0.03877006, -0.03877006, -0.05875626, -0.0…
#> $ e_qo_ps_C_tilde            <dbl> 0.001562322, 0.001562322, 0.001589951, 0.00…
#> $ N_dot_tilde                <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_engr_units_star  <dbl> 17, 17, 17, 17, 17, 17, 17, 17, 17, 17, 17,…
#> $ Delta_eta_star             <dbl> 0.1342583, 0.1342583, 0.1342583, 0.1342583,…
#> $ Delta_p_s_star             <dbl> -0.03578095, -0.03578095, -0.02428571, -0.0…
#> $ Delta_q_dot_s_star         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_E_dot_s_star         <dbl> -29572.18, -29572.18, -29572.18, -29572.18,…
#> $ Delta_E_dot_emb_star       <dbl> 428.5714, 428.5714, 428.5714, 428.5714, 428…
#> $ Delta_C_dot_s_star         <dbl> -516.1402, -516.1402, -350.3214, -350.3214,…
#> $ Delta_C_dot_cap_star       <dbl> -98.95714, -98.95714, -98.95714, -98.95714,…
#> $ Delta_C_dot_md_star        <dbl> -86.46614, -86.46614, -86.46614, -86.46614,…
#> $ Delta_C_dot_o_star         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_N_dot_star           <dbl> 701.5635, 701.5635, 535.7447, 535.7447, 535…
#> $ Delta_M_dot_star           <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_engr_units_hat   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_hat              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_p_s_hat              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_q_dot_s_hat          <dbl> 289.7372, 289.7372, 443.0672, 443.0672, 443…
#> $ Delta_E_dot_s_hat          <dbl> 873.4998, 873.4998, 1335.7592, 1335.7592, 1…
#> $ Delta_E_dot_emb_hat        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_s_hat          <dbl> 15.24569, 15.24569, 15.82383, 15.82383, 15.…
#> $ Delta_C_dot_cap_hat        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_md_hat         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_o_hat          <dbl> -19.94471, -19.94471, -20.69917, -20.69917,…
#> $ Delta_N_dot_hat            <dbl> 4.699019, 4.699019, 4.875341, 4.875341, 4.8…
#> $ Delta_M_dot_hat            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_engr_units_bar   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_bar              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_p_s_bar              <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_q_dot_s_bar          <dbl> 519.9241, 519.9241, 398.8263, 398.8263, 398…
#> $ Delta_E_dot_s_bar          <dbl> 1567.468, 1567.468, 1202.382, 1202.382, 120…
#> $ Delta_E_dot_emb_bar        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_s_bar          <dbl> 27.35791, 27.35791, 14.24380, 14.24380, 14.…
#> $ Delta_C_dot_cap_bar        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_md_bar         <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_o_bar          <dbl> 678.9046, 678.9046, 526.3763, 526.3763, 526…
#> $ Delta_N_dot_bar            <dbl> -706.2625, -706.2625, -540.6200, -540.6200,…
#> $ Delta_M_dot_bar            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_engr_units_tilde <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_eta_tilde            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_p_s_tilde            <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_q_dot_s_tilde        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_E_dot_s_tilde        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_E_dot_emb_tilde      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_s_tilde        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_cap_tilde      <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_md_tilde       <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_C_dot_o_tilde        <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_N_dot_tilde          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Delta_M_dot_tilde          <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Re_dempl                   <dbl> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ Re_emb                     <dbl> 0.01449239, 0.01449239, 0.01449239, 0.01449…
#> $ Re_cap                     <dbl> -0.01134172, -0.01134172, -0.01134172, -0.0…
#> $ Re_md                      <dbl> -0.009910093, -0.009910093, -0.009910093, -…
#> $ Re_empl                    <dbl> 0.004582292, 0.004582292, 0.004582292, 0.00…
#> $ Re_isub                    <dbl> -0.002285912, -0.002285912, -0.002372382, -…
#> $ Re_dsub                    <dbl> 0.02953789, 0.02953789, 0.04516945, 0.04516…
#> $ Re_sub                     <dbl> 0.02725198, 0.02725198, 0.04279707, 0.04279…
#> $ Re_dinc                    <dbl> 0.05300480, 0.05300480, 0.04065922, 0.04065…
#> $ Re_iinc                    <dbl> 0.07781090, 0.07781090, 0.06032925, 0.06032…
#> $ Re_inc                     <dbl> 0.1308157, 0.1308157, 0.1009885, 0.1009885,…
#> $ Re_micro                   <dbl> 0.1626500, 0.1626500, 0.1483678, 0.1483678,…
#> $ Re_macro                   <dbl> 0.08040789, 0.08040789, 0.00000000, 0.06140…
#> $ Re_dir                     <dbl> 0.08254269, 0.08254269, 0.08582867, 0.08582…
#> $ Re_indir                   <dbl> 0.16051517, 0.16051517, 0.06253917, 0.12394…
#> $ Re_tot                     <dbl> 0.2430579, 0.2430579, 0.1483678, 0.2097708,…