Path graph parameters

The list of graph parameters for drawing energy path graphs, expenditure path graphs, and preference path graphs in the ReboundTools package.

path_graph_params

Format

A string list with 84 entries.

which_points

A data frame telling which points to include in the graph.

last_point

Tells whether to show the last point in a path. Overrides which_points.

point_shape

The shape for points between rebound effects. Default is 21, a filled circle..

point_size

The size for points between rebound effects. Default is 1.

point_stroke

The size of the line surrounding points between rebound effects. Default is 1.

which_arrows

A data frame telling which ending arrows to include in the graph.

last_arrow

Tells whether to show the last arrow. Overrides which_arrows.

arrow_style

An arrow object created by grid::arrow.

dempl_colour

The colour for direct emplacment lines.

emb_colour

The colour for embodied energy lines.

cap_colour

The colour for capital expenditure lines.

md_colour

The colour for maintenance and disposal lines.

empl_colour

The colour for emplacement lines.

dsub_colour

The colour for direct substitution lines.

isub_colour

The colour for indirect substitution lines.

sub_colour

The colour for substitution lines.

dinc_colour

The colour for direct income lines.

iinc_colour

The colour for indirect income lines.

inc_colour

The colour for income lines.

macro_colour

The colour for macro lines.

dir_colour

The colour for direct lines.

indir_colour

The colour for macro lines.

tot_colour

The colour for total rebound lines.

dempl_size

The size for direct emplacment lines.

emb_size

The size for embodied energy lines.

cap_size

The size for capital expenditure lines.

md_size

The size for maintenance and disposal lines.

empl_size

The size for emplacement lines.

dsub_size

The size for direct substitution lines.

isub_size

The size for indirect substitution lines.

sub_size

The size for substitution lines.

dinc_size

The size for direct income lines.

iinc_size

The size for indirect income lines.

inc_size

The size for income lines.

macro_size

The size for macro lines.

dir_size

The size for direct lines.

indir_size

The size for indirect lines.

tot_size

The size for total rebound lines.

dempl_linetyps

The linetype for direct emplacment lines.

emb_linetype

The linetype for embodied energy lines.

cap_linetype

The linetype for capital expenditure lines.

md_linetype

The linetype for maintenance and disposal lines.

empl_linetype

The linetype for emplacement lines.

dsub_linetype

The linetype for direct substitution lines.

isub_linetype

The linetype for indirect substitution lines.

sub_linetype

The linetype for substitution lines.

dinc_linetype

The linetype for direct income lines.

iinc_linetype

The linetype for indirect income lines.

inc_linetype

The linetype for income lines.

macro_linetype

The linetype for macro lines.

dir_linetype

The linetype for direct lines.

indir_linetype

The linetype for indirect lines.

tot_linetype

The linetype for total rebound lines.

lineend

The line end style.

linejoin

The line join style.

reverse_path_drawing_order

Tells whether to reverse the drawing order for paths. The default (FALSE) draws emplacement on the bottom, followed by substitution, income, and macro paths. TRUE puts macro paths on the bottom, followed by income, substitution, and emplacement paths. Setting TRUE produces attractive layering when many paths have arrows, because arrows overlay their following points.

points_atop_paths

Tells whether to draw points above paths (TRUE) or beneath paths (FALSE). Default is TRUE.

energy_grid_colour

The colour for energy grid lines.

zero_perc_rebound_grid_colour

The colour for energy the 0% rebound lines.

hundred_perc_rebound_grid_colour

The colour for the 100% rebound lines.

energy_rebound_lines_colour

The colour for energy rebound lines.

expenditure_grid_colour

The colour for expenditure grid lines.

expenditure_ray_colour

The colour for expenditure rays.

prefs_grid_colour

The colour for preferences grid lines.

prefs_indiff_curve_colour

The colour for indifference curve lines.

energy_grid_size

The size for energy grid lines.

zero_perc_rebound_grid_size

The size for 0% rebound grid lines.

hundred_perc_rebound_grid_size

The size for the 100% rebound grid lines.

energy_rebound_lines_size

The size for energy rebound lines.

expenditure_grid_size

The size for expenditure grid lines.

expenditure_ray_size

The size for expenditure rays.

prefs_grid_size

The size for preferences grid lines.

prefs_indiff_curve_size

The size for indifference curve lines.

energy_grid_linetype

The linetype for energy grid lines.

zero_perc_rebound_grid_linetype

The linetype for 0% rebound grid lines.

hundred_perc_rebound_grid_linetype

The linetype for the 100% rebound grid lines.

energy_rebound_lines_linetype

The linetype for energy rebound lines.

expenditure_grid_linetype

The linetype for expenditure grid lines.

expenditure_ray_linetype

The linetype for expenditure rays.

prefs_grid_linetype

The linetype for preferences grid lines.

prefs_indiff_curve_linetype

The linetype for indifference curve lines.

n_indiff_curve_points

The number of points on the indifference curves.

qs_qs0_lower

The lower bound for the x value in the indifference curves.

qs_qs0_upper

The upper bound for the x value in the indifference curves.

include_start_point

A boolean that tells whether to include the starting point.

start_point_size

The size of the start point.

start_point_shape

The shape of the start point.

include_end_arrow

A boolean that tells whether to include an ending arrow.

arrow_angle

The angle for the arrow head, in degrees.

arrow_length

The length of the arrow. See grid::unit.

arrow_type

The arrow type, "closed" (the default) or "open". See grid::arrow.

Details

This list is passed to several graphing functions. Callers may pass a modified version of this list to change graph appearance.

Examples

path_graph_params

#> $which_points
#> # A tibble: 5 x 2
#>   point_name start_point
#>   <chr>      <lgl>      
#> 1 orig       TRUE       
#> 2 star       TRUE       
#> 3 hat        TRUE       
#> 4 bar        TRUE       
#> 5 tilde      FALSE      
#> 
#> $last_point
#> [1] FALSE
#> 
#> $point_shape
#> [1] 19
#> 
#> $point_size
#> [1] 1
#> 
#> $point_stroke
#> [1] 1
#> 
#> $which_arrows
#> # A tibble: 9 x 2
#>   line_name end_arrow
#>   <chr>     <lgl>    
#> 1 dempl     FALSE    
#> 2 emb       FALSE    
#> 3 cap       FALSE    
#> 4 md        FALSE    
#> 5 isub      FALSE    
#> 6 dsub      FALSE    
#> 7 dinc      FALSE    
#> 8 iinc      FALSE    
#> 9 macro     FALSE    
#> 
#> $last_arrow
#> [1] TRUE
#> 
#> $arrow_style
#> $angle
#> [1] 20
#> 
#> $length
#> [1] 0.1inches
#> 
#> $ends
#> [1] 2
#> 
#> $type
#> [1] 2
#> 
#> attr(,"class")
#> [1] "arrow"
#> 
#> $dempl_colour
#> [1] "#150789FF"
#> 
#> $emb_colour
#> [1] "#150789FF"
#> 
#> $cap_colour
#> [1] "#150789FF"
#> 
#> $md_colour
#> [1] "#150789FF"
#> 
#> $empl_colour
#> [1] "#150789FF"
#> 
#> $isub_colour
#> [1] "#99149FFF"
#> 
#> $dsub_colour
#> [1] "#99149FFF"
#> 
#> $sub_colour
#> [1] "#99149FFF"
#> 
#> $dinc_colour
#> [1] "#E76F5AFF"
#> 
#> $iinc_colour
#> [1] "#E76F5AFF"
#> 
#> $inc_colour
#> [1] "#E76F5AFF"
#> 
#> $macro_colour
#> [1] "#F7E225FF"
#> 
#> $dir_colour
#> [1] "black"
#> 
#> $indir_colour
#> [1] "black"
#> 
#> $tot_colour
#> [1] "black"
#> 
#> $dempl_size
#> [1] 1
#> 
#> $emb_size
#> [1] 1.5
#> 
#> $cap_size
#> [1] 1.5
#> 
#> $md_size
#> [1] 1
#> 
#> $empl_size
#> [1] 1
#> 
#> $isub_size
#> [1] 1
#> 
#> $dsub_size
#> [1] 1
#> 
#> $sub_size
#> [1] 1
#> 
#> $dinc_size
#> [1] 1
#> 
#> $iinc_size
#> [1] 1
#> 
#> $inc_size
#> [1] 1
#> 
#> $macro_size
#> [1] 1
#> 
#> $dir_size
#> [1] 1
#> 
#> $indir_size
#> [1] 1
#> 
#> $tot_size
#> [1] 2
#> 
#> $dempl_linetype
#> [1] "solid"
#> 
#> $emb_linetype
#> [1] "11"
#> 
#> $cap_linetype
#> [1] "11"
#> 
#> $md_linetype
#> [1] "solid"
#> 
#> $empl_linetype
#> [1] "solid"
#> 
#> $dsub_linetype
#> [1] "solid"
#> 
#> $isub_linetype
#> [1] "solid"
#> 
#> $sub_linetype
#> [1] "solid"
#> 
#> $dinc_linetype
#> [1] "solid"
#> 
#> $iinc_linetype
#> [1] "solid"
#> 
#> $sinc_linetype
#> [1] "solid"
#> 
#> $macro_linetype
#> [1] "solid"
#> 
#> $dir_linetype
#> [1] "solid"
#> 
#> $indir_linetype
#> [1] "solid"
#> 
#> $tot_linetype
#> [1] "solid"
#> 
#> $lineend
#> [1] "round"
#> 
#> $linejoin
#> [1] "round"
#> 
#> $reverse_path_drawing_order
#> [1] FALSE
#> 
#> $points_atop_paths
#> [1] TRUE
#> 
#> $energy_grid_colour
#> [1] "black"
#> 
#> $zero_perc_rebound_grid_colour
#> [1] "black"
#> 
#> $hundred_perc_rebound_grid_colour
#> [1] "black"
#> 
#> $energy_rebound_lines_colour
#> [1] "black"
#> 
#> $expenditure_grid_colour
#> [1] "black"
#> 
#> $prefs_grid_colour
#> [1] "black"
#> 
#> $prefs_ray_colour
#> [1] "black"
#> 
#> $prefs_indiff_grid_colour
#> [1] "black"
#> 
#> $energy_grid_size
#> [1] 0.1
#> 
#> $zero_perc_rebound_grid_size
#> [1] 0.3
#> 
#> $hundred_perc_rebound_grid_size
#> [1] 0.3
#> 
#> $energy_rebound_lines_size
#> [1] 0.1
#> 
#> $expenditure_grid_size
#> [1] 0.3
#> 
#> $prefs_grid_size
#> [1] 0.1
#> 
#> $prefs_ray_size
#> [1] 0.1
#> 
#> $prefs_indiff_grid_size
#> [1] 0.5
#> 
#> $energy_grid_linetype
#> [1] "solid"
#> 
#> $zero_perc_rebound_grid_linetype
#> [1] "solid"
#> 
#> $hundred_perc_rebound_grid_linetype
#> [1] "solid"
#> 
#> $energy_rebound_lines_linetype
#> [1] "solid"
#> 
#> $expenditure_grid_linetype
#> [1] "solid"
#> 
#> $prefs_grid_linetype
#> [1] "solid"
#> 
#> $prefs_ray_linetype
#> [1] "solid"
#> 
#> $prefs_indiff_grid_linetype
#> [1] "solid"
#> 
#> $n_indiff_curve_points
#> [1] 200
#> 
#> $qs_qs0_lower
#> [1] 0.1
#> 
#> $qs_qs0_upper
#> [1] 10
#>