Build a rebound results table

Build a rebound results table

rebound_results_table(
  .analysis_data,
  escape_latex = TRUE,
  include_subtotals = TRUE,
  include_total = TRUE,
  as_percent = TRUE,
  rebound_terms = ReboundTools::rebound_terms,
  latex_rebound_terms = ReboundTools::latex_rebound_terms,
  case = ReboundTools::eeu_base_params$case,
  subtotals = c(ReboundTools::rebound_terms$Re_empl, ReboundTools::rebound_terms$Re_sub,
    ReboundTools::rebound_terms$Re_inc, ReboundTools::rebound_terms$Re_dir,
    ReboundTools::rebound_terms$Re_indir, ReboundTools::rebound_terms$Re_micro),
  total = ReboundTools::rebound_terms$Re_tot,
  term_name = "Rebound term",
  latex_term_name = "LaTeX rebound term",
  Re_val_colname = "Value [-]",
  perc_Re_val_colname = "Value [%]",
  latex_Re_val_colname = "Value [--]",
  latex_perc_Re_val_colname = "Value [\\%]",
  ...
)

Arguments

.analysis_data

Rebound analysis results. Probably the output of a call to rebound_analysis().

escape_latex

When TRUE (the default), return LaTeX-compatible versions of strings.

include_subtotals

Tells whether to include rebound subtotals by stage and direct/indirect. Default is TRUE.

include_total

Tells whether to include rebound total. Default is TRUE.

as_percent

When TRUE, rebound results are reported as percentages. When FALSE, rebound results are reported as fractions. Default is TRUE.

rebound_terms

See ReboundTools::rebound_terms.

latex_rebound_terms

See ReboundTools::latex_rebound_terms.

case

See ReboundTools::eeu_base_params.

subtotals

The rebound terms that represent subtotals. Default is c(ReboundTools::rebound_terms$Re_empl, ReboundTools::rebound_terms$Re_sub, ReboundTools::rebound_terms$Re_inc, ReboundTools::rebound_terms$Re_dir, ReboundTools::rebound_terms$Re_indir, ReboundTools::rebound_terms$Re_micro).

total

The rebound term that represents total rebound. Default is ReboundTools::rebound_terms$Re_tot.

term_name

The title of the rebound term column. Default is "Rebound term".

latex_term_name

The LaTeX term name column. Default is "LaTeX rebound term".

Re_val_colname

The title of the rebound value column. Default is "Value [--]".

perc_Re_val_colname

The title of the rebound value column when percentages are requested. Default is "Value [%]".

latex_Re_val_colname

The LaTeX rebound value column name. Default is "Value [--]".

latex_perc_Re_val_colname

The LaTeX rebound value column name when percentages are requested. Default is "Value [\%]".

...

Arguments passed to xtable::xtable(), possibly label, caption, digits, etc.

Value

An xtable object suitable for printing.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  rebound_results_table()
#> % latex table generated in R 4.2.1 by xtable 1.8-4 package
#> % Mon Jan  2 13:33:42 2023
#> \begin{table}[ht]
#> \centering
#> \begin{tabular}{rllr}
#>   \hline
#>  & Case & Rebound term & Value [$\backslash$\%] \\ 
#>   \hline
#> 1 & Car & \$Re\_\{dempl\}\$ & 0.00 \\ 
#>   2 & Car & \$Re\_\{emb\}\$ & 1.45 \\ 
#>   3 & Car & \$Re\_\{cap\}\$ & -1.13 \\ 
#>   4 & Car & \$Re\_\{md\}\$ & -0.99 \\ 
#>   5 & Car & \$Re\_\{empl\}\$ & 0.46 \\ 
#>   6 & Car & \$Re\_\{dsub\}\$ & 2.95 \\ 
#>   7 & Car & \$Re\_\{isub\}\$ & -0.23 \\ 
#>   8 & Car & \$Re\_\{sub\}\$ & 2.73 \\ 
#>   9 & Car & \$Re\_\{dinc\}\$ & 5.30 \\ 
#>   10 & Car & \$Re\_\{iinc\}\$ & 7.78 \\ 
#>   11 & Car & \$Re\_\{inc\}\$ & 13.08 \\ 
#>   12 & Car & \$Re\_\{micro\}\$ & 16.26 \\ 
#>   13 & Car & \$Re\_\{macro\}\$ & 8.04 \\ 
#>   14 & Car & \$Re\_\{dir\}\$ & 8.25 \\ 
#>   15 & Car & \$Re\_\{indir\}\$ & 16.05 \\ 
#>   16 & Car & \$Re\_\{tot\}\$ & 24.31 \\ 
#>   17 & Lamp & \$Re\_\{dempl\}\$ & 0.00 \\ 
#>   18 & Lamp & \$Re\_\{emb\}\$ & -0.27 \\ 
#>   19 & Lamp & \$Re\_\{cap\}\$ & -1.48 \\ 
#>   20 & Lamp & \$Re\_\{md\}\$ & 0.00 \\ 
#>   21 & Lamp & \$Re\_\{empl\}\$ & -0.27 \\ 
#>   22 & Lamp & \$Re\_\{dsub\}\$ & 17.36 \\ 
#>   23 & Lamp & \$Re\_\{isub\}\$ & -7.02 \\ 
#>   24 & Lamp & \$Re\_\{sub\}\$ & 10.35 \\ 
#>   25 & Lamp & \$Re\_\{dinc\}\$ & 0.01 \\ 
#>   26 & Lamp & \$Re\_\{iinc\}\$ & 19.04 \\ 
#>   27 & Lamp & \$Re\_\{inc\}\$ & 19.06 \\ 
#>   28 & Lamp & \$Re\_\{micro\}\$ & 29.13 \\ 
#>   29 & Lamp & \$Re\_\{macro\}\$ & 14.24 \\ 
#>   30 & Lamp & \$Re\_\{dir\}\$ & 17.38 \\ 
#>   31 & Lamp & \$Re\_\{indir\}\$ & 25.99 \\ 
#>   32 & Lamp & \$Re\_\{tot\}\$ & 43.37 \\ 
#>    \hline
#> \end{tabular}
#> \end{table}