Build a stages table

A stages table has variables in rows and stages in columns. These tables enable tracking of variables across the different stages of rebound.

stages_table(
  .analysis_data,
  add_units = TRUE,
  escape_latex = TRUE,
  vars = ReboundTools::key_analysis_vars,
  latex_vars = ReboundTools::latex_key_analysis_vars,
  stages = ReboundTools::rebound_stages,
  latex_stages = ReboundTools::latex_rebound_stages,
  case = ReboundTools::eeu_base_params$case,
  service_unit = ReboundTools::eeu_base_params$service_unit,
  energy_engr_unit = ReboundTools::eeu_base_params$energy_engr_unit,
  ...,
  .var = ".var",
  .stage = ".stage",
  .var_stage = ".var_stage",
  .value = ".value",
  .name = ".name",
  .unit_col = ".unit_col"
)

Arguments

.analysis_data
A data frame, usually the result of calling rebound_analysis(). Default is rebound_analysis(load_eeu_data(file)).

add_units
When TRUE (the default), adds a unit specification to variable names in the table.

escape_latex
When TRUE (the default), return LaTeX-compatible versions of strings.

vars
A list of variables for rows of the table. Default is ReboundTools::key_analysis_vars. Variable order is preserved in the table.

latex_vars
See ReboundTools::latex_key_analysis_vars. Set NULL to prevent conversion to LaTeX variable names.

stages
A list of stages for columns of the table. Default is ReboundTools::rebound_stages. Stage order is preserved in the table.

latex_stages
See ReboundTools::latex_rebound_stages. Set NULL to prevent conversion to LaTeX stage names.

case
See ReboundTools::eeu_base_params.

service_unit, energy_engr_unit
See ReboundTools::eeu_base_params.

...
Arguments passed to xtable::xtable(), possibly label, caption, digits, etc.

.var, .stage, .var_stage, .value, .name, .unit_col
Column names used internally.

Value

An xtable object suitable for printing.

Examples

load_eeu_data() %>% 
  rebound_analysis() %>% 
  stages_table()

#> % latex table generated in R 4.0.3 by xtable 1.8-4 package
#> % Sat Jan 16 22:09:46 2021
#> \begin{table}[ht]
#> \centering
#> \begin{tabular}{rllrrrrr}
#>   \hline
#>  & Case &   & \$\{ \}\verb|^|\{$\backslash$scriptscriptstyle $\backslash$circ\}\$ (orig) & \$\{ \}\verb|^|*\$ (star) & \$$\backslash$hat\{ \}\$ (hat) & \$$\backslash$bar\{ \}\$ (bar) & \$$\backslash$tilde\{ \}\$ (tilde) \\ 
#>   \hline
#> 1 & Car & \$$\backslash$eta\$ [mile/MJ] & 0.20 & 0.33 & 0.33 & 0.33 & 0.33 \\ 
#>   2 & Car & \$p\_s\$ [$\backslash$\$/mile] & 0.09 & 0.05 & 0.05 & 0.05 & 0.05 \\ 
#>   3 & Car & \$$\backslash$dot\{q\}\_s\$ [mile/year] & 14425.00 & 14425.00 & 14710.86 & 15234.66 & 15234.66 \\ 
#>   4 & Car & \$$\backslash$dot\{E\}\_s\$ [MJ/year] & 73060.68 & 43488.50 & 44350.31 & 45929.47 & 45929.47 \\ 
#>   5 & Car & \$$\backslash$dot\{E\}\_\{emb\}\$ [MJ/year] & 2428.57 & 2857.14 & 2857.14 & 2857.14 & 2857.14 \\ 
#>   6 & Car & \$$\backslash$dot\{C\}\_s\$ [$\backslash$\$/year] & 1275.17 & 759.03 & 774.07 & 801.63 & 801.63 \\ 
#>   7 & Car & \$$\backslash$dot\{C\}\_\{cap\}\$ [$\backslash$\$/year] & 4030.87 & 3931.91 & 3931.91 & 3931.91 & 3931.91 \\ 
#>   8 & Car & \$$\backslash$dot\{C\}\_\{md\}\$ [$\backslash$\$/year] & 2861.13 & 2774.67 & 2774.67 & 2774.67 & 2774.67 \\ 
#>   9 & Car & \$$\backslash$dot\{C\}\_o\$ [$\backslash$\$/year] & 19234.10 & 19234.10 & 19209.10 & 19893.06 & 19893.06 \\ 
#>   10 & Car & \$$\backslash$dot\{N\}\$ [$\backslash$\$/year] & 0.00 & 701.56 & 711.53 & 0.00 & 0.00 \\ 
#>   11 & Car & \$$\backslash$dot\{M\}\$ [$\backslash$\$/year] & 27401.28 & 27401.28 & 27401.28 & 27401.28 & 27401.28 \\ 
#>   12 & Lamp & \$$\backslash$eta\$ [lm-hr/MJ] & 2453.70 & 22722.22 & 22722.22 & 22722.22 & 22722.22 \\ 
#>   13 & Lamp & \$p\_s\$ [$\backslash$\$/lm-hr] & 0.00 & 0.00 & 0.00 & 0.00 & 0.00 \\ 
#>   14 & Lamp & \$$\backslash$dot\{q\}\_s\$ [lm-hr/year] & 580350.00 & 580350.00 & 1412626.68 & 1413421.26 & 1413421.26 \\ 
#>   15 & Lamp & \$$\backslash$dot\{E\}\_s\$ [MJ/year] & 236.52 & 25.54 & 62.17 & 62.20 & 62.20 \\ 
#>   16 & Lamp & \$$\backslash$dot\{E\}\_\{emb\}\$ [MJ/year] & 1.22 & 0.65 & 0.65 & 0.65 & 0.65 \\ 
#>   17 & Lamp & \$$\backslash$dot\{C\}\_s\$ [$\backslash$\$/year] & 8.90 & 0.96 & 2.34 & 2.34 & 2.34 \\ 
#>   18 & Lamp & \$$\backslash$dot\{C\}\_\{cap\}\$ [$\backslash$\$/year] & 1.04 & 0.12 & 0.12 & 0.12 & 0.12 \\ 
#>   19 & Lamp & \$$\backslash$dot\{C\}\_\{md\}\$ [$\backslash$\$/year] & 0.00 & 0.00 & 0.00 & 0.00 & 0.00 \\ 
#>   20 & Lamp & \$$\backslash$dot\{C\}\_o\$ [$\backslash$\$/year] & 27391.33 & 27391.33 & 27383.41 & 27398.82 & 27398.82 \\ 
#>   21 & Lamp & \$$\backslash$dot\{N\}\$ [$\backslash$\$/year] & 0.00 & 8.86 & 15.40 & 0.00 & 0.00 \\ 
#>   22 & Lamp & \$$\backslash$dot\{M\}\$ [$\backslash$\$/year] & 27401.28 & 27401.28 & 27401.28 & 27401.28 & 27401.28 \\ 
#>    \hline
#> \end{tabular}
#> \end{table}