EEU star data

This is the list of the derived variables at the star (post-emplacement) stage of a rebound analysis.

star_vars

Format

A string list with 19 entries.

C_cap_star

The net capital cost of the upgraded device: the sum of purchase price and financing costs less rebates and resale value at end of ownership [$].

t_own_star

The expected ownership duration of the upgraded device [year].

C_dot_md_star

The upgraded (post-EEU) maintenance and disposal cost rate [$/year].

E_emb_star

The embodied energy of the upgraded (post-EEU) device [MJ].

t_life_star

The expected lifetime of the upgraded (post-EEU) device [year].

eta_star_engr_units

The upgraded (post-EEU) energy service efficiency. This number should have engineering units in the denominator, e.g., [miles/gallon] [lumens/kW]. Note that the denominator unit of eta_engr_units_orig is assumed to be the same as the denominator unit of MJ_engr_unit.

eta_star

Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], calculated by eta_tilde_engr_units / MJ_engr_unit.

eta_ratio

The energy service efficiency ratio [--], calculated by eta_star/eta_orig.

S_dot_dev

The expected device-level energy savings rate [MJ/year], calculated by (eta_ratio - 1) * (1/eta_ratio) * E_dot_s_orig.

G_dot

The expected device-level energy gross cost savings rate [MJ/year], calculated by p_E * S_dot_dev.

p_s_star

The upgraded (post-EEU) energy service price [$/service], calculated by p_E / eta_star = p_E / eta_tilde.

q_dot_s_star

The upgraded (post-EEU) energy service consumption rate [service/year], same as q_dot_s_orig.

C_dot_cap_star

The upgraded (post-EEU) capital cost rate [$/year], calculated by C_cap_star / t_star.

E_dot_emb_star

The upgraded (post-EEU) embodied energy rate [MJ/year], calculated by E_emb_star / t_star.

C_dot_s_star

The upgraded (post-EEU) energy cost rate [$/year], calculated by p_s_star * q_dot_s_star.

M_dot_star

The disposable income rate, exclusive of taxes and savings [$/year], exactly M_dot_orig.

N_dot_star

The freed cash rate [$/year], calculated by G_dot - (C_dot_cap_star - C_dot_cap_orig) - (C_dot_md_star - C_dot_md_orig).

C_dot_o_star

The upgraded (post-EEU) other goods expenditure rate [$/year], exactly C_dot_o_orig.

E_dot_s_star

The upgraded (post-EEU) energy consumption rate [MJ/year], calculated by q_dot_s_star / eta_star.

Examples

star_vars

#> $C_cap_star
#> [1] "C_cap_star"
#> 
#> $t_own_star
#> [1] "t_own_star"
#> 
#> $C_dot_md_star
#> [1] "C_dot_md_star"
#> 
#> $E_emb_star
#> [1] "E_emb_star"
#> 
#> $t_life_star
#> [1] "t_life_star"
#> 
#> $eta_star_engr_units
#> [1] "eta_engr_units_star"
#> 
#> $eta_star
#> [1] "eta_star"
#> 
#> $eta_ratio
#> [1] "eta_ratio"
#> 
#> $S_dot_dev
#> [1] "S_dot_dev"
#> 
#> $G_dot
#> [1] "G_dot"
#> 
#> $p_s_star
#> [1] "p_s_star"
#> 
#> $q_dot_s_star
#> [1] "q_dot_s_star"
#> 
#> $C_dot_cap_star
#> [1] "C_dot_cap_star"
#> 
#> $E_dot_emb_star
#> [1] "E_dot_emb_star"
#> 
#> $C_dot_s_star
#> [1] "C_dot_s_star"
#> 
#> $M_dot_star
#> [1] "M_dot_star"
#> 
#> $N_dot_star
#> [1] "N_dot_star"
#> 
#> $C_dot_o_star
#> [1] "C_dot_o_star"
#> 
#> $E_dot_s_star
#> [1] "E_dot_s_star"
#>