EEU tilde data

This is the list of the derived variables at the tilde stage (after productivity effect) of a rebound analysis.

tilde_vars

Format

A string list with 11 entries.

eta_tilde

Energy service efficiency of the upgraded (post-EEU) device on a per-MJ basks [service/MJ], exactly eta_bar.

p_s_tilde

The energy service price after the productivity effect [$/service], exactly p_s_bar.

C_dot_cap_tilde

The capital expenditure rate after the productivity effect [$/year], exactly C_dot_cap_bar.

C_dot_md_tilde

The maintenance and disposal expenditure rate after the productivity effect [$/year], exactly C_dot_md_bar.

E_dot_emb_tilde

The embodied energy rate after the productivity effect [MJ/year], exactly E_dot_emb_bar.

M_dot_tilde

Real income after the productivity effect [MJ/year], exactly M_dot_bar.

q_dot_s_tilde

The rate of energy service consumption after the productivity effect\ [service/year], exactly q_dot_s_bar.

C_dot_o_tilde

The rate of other goods expenditures after the productivity effect [$/year], exactly C_dot_o_bar.

N_dot_tilde

The freed cash rate after the productivity effect [$/year], exactly 0.

[service/year]: R:service/year%5C

Examples

tilde_vars

#> $eta_tilde
#> [1] "eta_tilde"
#> 
#> $p_s_tilde
#> [1] "p_s_tilde"
#> 
#> $C_dot_cap_tilde
#> [1] "C_dot_cap_tilde"
#> 
#> $C_dot_md_tilde
#> [1] "C_dot_md_tilde"
#> 
#> $E_dot_emb_tilde
#> [1] "E_dot_emb_tilde"
#> 
#> $M_dot_tilde
#> [1] "M_dot_tilde"
#> 
#> $q_dot_s_tilde
#> [1] "q_dot_s_tilde"
#> 
#> $E_dot_s_tilde
#> [1] "E_dot_s_tilde"
#> 
#> $C_dot_s_tilde
#> [1] "C_dot_s_tilde"
#> 
#> $C_dot_o_tilde
#> [1] "C_dot_o_tilde"
#> 
#> $N_dot_tilde
#> [1] "N_dot_tilde"
#>